android: Don't allow AutofillPopup on outside touch
By default, PopupWindows dismiss when the user touches outside the window. The only
way to avert this behavior when using a ListPopupWindow is to rely on a hidden API.

This manifested as the linked bug because each tap dismissed the popup and detached it from the anchor view. C++ didn't realize the popup was dismissed, and tried to update it, but there was no anchor view.

BUG=400601

Review URL: https://codereview.chromium.org/442933002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@291463 0039d316-1c4b-4281-b951-d872f2087c98