aw: Avoid uncontrolled video context destruction
Since r289474, the sync factory nulls out the video context
which means the destruction is uncontrolled. Context
destruction is synchronous, so can cause deadlocks if there
are not webviews attached to view tree to run GL.

This is a rewrite of r289474 that just returns null
VideoContextProvider if no webview is ready to use it. But
still hold on to the reference to the context to avoid
destruction.

BUG=403882

Review URL: https://codereview.chromium.org/498623002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@291527 0039d316-1c4b-4281-b951-d872f2087c98