Begin whitelisting specific extensions for global key registration.
Adds whitelist to commands API to relax restrictions on:
- global accelerator keys
- maximum number of commands
- override previously taken accelerators

This cl also involved fixing the use of the search key while using the global commands listener. The 'command' modifier was being dropped while making its way to the renderer and back.

BUG=404768
TEST=With a sample extension binding
"Search+Shift+L" and
"global": true

press and verify callback when
focus is inside of web content
focus is on the status tray

Review URL: https://codereview.chromium.org/399783002