User Manager: fix text blurriness and "remove" button style
1. The text is blurry because we use a transform: scale3d(), which on non-retina
desktop Chrome is blurry. I assume there is a reason why CrOS chose a scale3d()
so I only made this change for the desktop.
2. This leads to a webkit bug (crbug.com/406529) where transforms and opacity
transitions do not play well together, so I've removed the opacity transition on
the action-area-box.

Also:
3. Fixed the focus style on the "remove user" button to not be the ugly blue
default but match the gmail style.

BUG=399683, 403784
TEST=Start Chrome with -enable-new-avatar-menu. Click on "Switch person".
The profile names in the user manager should not be blurry, and hovering
over the pods should not be janky.

Review URL: https://codereview.chromium.org/476193003