Fix a couple of autofill popup issues
1. The popup needs to hide when the renderer resizes (akin to hiding on scroll or orientation change).
2. There was a race condition where the renderer could say Show() then Hide(), but the Hide() would fall on deaf ears because the popup was still waiting for webdata.

The linked bug occurred when the render view was auto-resized /after/ an orientation change. After the orientation change we'd get a text field changed event which would trigger the popup to reshow. The fix is to make the renderer resizing close the popup, but (2) and (3) above were necessary for this to actually work.

BUG=401752

Review URL: https://codereview.chromium.org/482643004