Move EME code out of WebMediaPlayerImpl.
This change moves all the encrypted media code out of WebMediaPlayerImpl
and into a EncryptedMediaSupportImpl class. This removes the encrypted
media related content/ depencencies from WebMediaPlayerImpl so that it
will be easier to move WebMediaPlayerImpl to a location where it can
be reused by Mojo.

Review URL: https://codereview.chromium.org/501473003