Restrict ETC1 power-of-two rounding to old IMG drivers.
The ETC1 spec says that any multiple of 4 is allowed for ETC1 size, but
certain old IMG devices (notably the Galaxy Nexus) crash unless the
textures are specifically power-of-two. To avoid wasting RAM on all
devices due to this device-specific problem, this patch plumbs a driver
workaround flag to the thumbnail store (the only user of ETC1 at
present).

We considered adding a workaround to the service side like all of the
others, but because texture size impacts memory budgeting and
coordinates, this one seems like it needs to be in the client code.

NOTRY=true
BUG=150500

Review URL: https://codereview.chromium.org/470233003