Add armv7s to the iOS ARCH list
App Store rules require shipping armv7s with arm64, so this is required
for the "both" configuration.

Also adds it to the 32-bit configuration so that local testing will
better match actual builds, but we can revisit if that turns out to be
onerous in practice.

BUG=407295

Review URL: https://codereview.chromium.org/489713003