Reduce input dependencies in GN.
Converts input dependencies for binary targets to be order-only, rather than implicit. See added comments in the patch for reasoning behind this.

Long ago this was order-only, which is not correct for actions (which might directly depend on the output of previous actions). So I changed them to be implicit, which then caused unnecessary rebuilds. This change makes actions and binary targets behave differently. Actions remain implicit, binary targets are order-only.

Add unit test for input deps on binary targets.

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/505353002