Add support for tracing in browser tests.
This CL adds support for collecting tracing data during the execution of browser tests. It is currently not possible to do so, yet it is very useful to be able to collect this data and analyze it without having to step through with debugger.

BUG=

Review URL: https://codereview.chromium.org/506113002