[NaCl SDK] Fix the SDK builders to use Cr-Commit-Position.
We are relying on lastchange.FetchVersionInfo in build_sdk.py. The revision
changed from being an SVN revision to a git sha, but we want to be able to
compare the values numerically.

This change modifies build_version.py to get "Cr-Commit-Position", which for
our purposes is functionally equivalent to the SVN revision.

BUG=none
R=bradnelson@google.com, bradnelson@chromium.org

Review URL: https://codereview.chromium.org/507883003