Fix overzealous CHECK in custom EC_KEYs.
If any EC_KEY with EX_DATA is duplicated, whether or not it is one of ours,
ExDataDup may get called. This is unlikely to ever happen (I don't think
we ever dup any EC_KEY on the client, and dup_ex_data has a fast-path if
the EX_DATA has never been touched anyway), but the CHECK is a little
overzealous.

BUG=none

Review URL: https://codereview.chromium.org/506813002