Implement <extensionoptions> deferred attach logic
Due to how the browser plugin is created, it is possible for the
<extensionoptions> element to attempt to attach to the embedder before
the browser plugin has bindings. This occurs when an ExtensionOptions object
gets an extension id and creates a guest, but has not yet been added
to the DOM. Use the deferred attach logic from <webview> so that if this
case occurs, wait until the browser plugin gets its bindings before
trying to attach the guest view.

BUG=386838

Review URL: https://codereview.chromium.org/457653003