Address flaky test by waiting for focus change.
This change attempts to fix a flaky test by changing the KeyPress function to
use SendKeyPressAndWait which will wait for an event. The event that this test
will wait for is the focus change event that should get called when an element
in the web form changes focus.

BUG=405711

Review URL: https://codereview.chromium.org/503863002