mojo: Fix double character inserts on Windows by removing a Linux hack.
If we're going to make mojo a write-once-run-anywhere platform, we'll
have to smooth over platform differences in event handling. Currently,
we have code that tries to emulate this behaviour at different levels
of the chrome stack on Linux. This emulation code lead to the double
character insertion bug on Windows: things looked fine on Linux
because we had another hack to cover up the lack of character events,
which just caused double character insertion on Windows.

In this patch, we declare the Windows style of events (a keydown,
a translated character event, a keyup) as canonical and what
should be targeted going forward.

BUG=403514

Review URL: https://codereview.chromium.org/476283006