Adding Baseframework of the ResourceManager (all hooks and observers) are being put in place with a very basic management scheme
The current ash implementation of the LowMemoryObserver polls
"/dev/chromeos-low-mem" to find if we are running out of memory.
After discussion with semenzato@ we decided that that is not precise
enough and we should rather poll the system memory information. Going
forward this has multiple benefits:
 1. it does not require special kernel handling.
 2. it also gives us information about GPU memory on chromeOS which
    will come handy once we are adding more sophisticated management.

Beside that, this CL adds the ResourceManager which hooks itself into
several components to detect changes in Activity's and their ordering
as well as memory pressure.

Note: The current management implementation itself is really primitive
and needs to be improved.

TBR for the DEPS file

BUG=381212, 403782
TEST=MemoryPressureTest.*, ResourceManagerTest.*
TBR=oshima@chromium.org

Review URL: https://codereview.chromium.org/513523002