MacViews: Guard Objective-C bits in content::WebContentsViewDelegate with __OBJC__
Currently WebContentsViewDelegate has a mix of ObjectiveC and C++ code
guarded by #ifdef OS_MACOSX. This means, on Mac, it can only be included
in .mm files.

For views on Mac, we want to reuse the toolkit-views WCVD (and tests)
when hosting a WebContents in a views::Widget. So we can't convert
everything to .mm.

This CL changes the guards to use __OBJC__ and ensures the vtable is a
consistent size with the compile flags used on MacViews.

To reduce churn, the WCVD methods are given default implementations.
MacViews needs the other methods, currently guarded by USE_AURA, so
things get really complex without this.

BUG=399191

Review URL: https://codereview.chromium.org/507363002