New SimpleCache Entry test for duplicate dooms.
Adds a regression test for bug 317138. This test is interesting
because it generates exactly the same stacks that we're seeing in
crash reports on that bug. This is a good regression test, so I
suggest we land it, as well as look at it for more confidence that
our fix to bug 317138 is the right fix.

R=ttuttle@chromium.org,jkarlin@chromium.org
BUG=317138

Review URL: https://codereview.chromium.org/509973002