Make DoDeferredClose a non-nestable task
This makes it behave the same as the Close function.

Destroying RenderWidgetCompositor during RenderWidget::DoDeferredClose
can cause problems for nested message loops.  The bug is that
InputHandlerProxy creates a SwapPromiseMonitor which adds itself to the
LayerTreeHost.  In that stack, a nested message loop is run which runs
the RenderWidget::DoDeferredClose function which previously destroyed
the LayerTreeHost.  This explodes, because LayerTreeHost expects all
SwapPromiseMonitors to be removed before it is destroyed.

Making DoDeferredClose non-nestable avoids this problem, as the nested
message loop can't run the task, the SwapPromiseMonitor safely gets
removed, and then the LayerTreeHost can be cleanly deleted.

This patch depends on https://codereview.chromium.org/512533002

R=jamesr@chromium.org,danakj@chromium.org
BUG=403500

Review URL: https://codereview.chromium.org/513433002