Introduce proximity_auth component.
This component handles the protocol between Chrome and an associated phone for
the Easy Unlock and Easy Sign-in features.

The system is currently implemented as a Chrome packaged app, and this is the
first step in migrating this functionality to native code. We will gradually
reimplement the system and switch from the app to this component once the
refactoring is complete.

BUG=385719

Review URL: https://codereview.chromium.org/505373004