Revert of Keep a copy of page id in RenderViewHost. (patchset #2 id:20001 of https://codereview.chromium.org/517813002/)
Reason for revert:
Crash data is coming in from the 39.0.2140.0 canary.  I'll revert this so that it doesn't continue reporting crashes over the long weekend.

(Sadly, the crashes so far appear to be in the browser process and not the renderer process, meaning we don't have a solid explanation yet.)

Original issue's description:
> Keep a copy of page id in RenderViewHost.
>
> This is an instrumented version of the patch that will be reverted in a few days. This is meant to catch crashes in edge cases and log enough for us to repro them.
>
> BUG=407376
> TEST=this may crash for a few people
>
> Committed: https://chromium.googlesource.com/chromium/src/+/ab752e59a10a2a4c322ee7f6dbbb16a4fdb467e0

TBR=avi@chromium.org
NOTREECHECKS=true
NOTRY=true
BUG=407376

Review URL: https://codereview.chromium.org/526493004