Remove RenderThreadImpl dependencies from WebMediaPlayerImpl.
This change replaces RenderThreadImpl calls in WebMediaPlayerImpl with
calls on the WebMediaPlayerParams passed into the constructor. This
removes another content/ depencency so that this code can be moved to
media/blink.

The EncryptedMediaSupport creation was moved to RenderThreadImpl and
passed via WebMediaPlayerParams to avoid linking problems that would
occur if WebMediaPlayerImpl was in media/blink.

BUG=408338

Review URL: https://codereview.chromium.org/517003002