Http cache: Allow caching of byte ranges without strong validators.
In order to fetch a resource using multiple byte range requests, the cache
has to make sure that the resource has not changed, and strong validators
are needed for that.

However, if the cache only has to return data without talking to the server
again, strong validators are not required.

It used to be the case that the cache will not return content from sparse
entries without revalidating tnem with the server first. But that changed a
long time ago to allow regular reuse policies, even for sparse/range requests.

This CL allows caching of partial responses that don't have strong validators,
as they are useful as described above.

BUG=407923

Review URL: https://codereview.chromium.org/509783002