Retry Set RequestContextObject for PNaCl pexe fetches, to fix On-Demand update.
There was a race condition in the added test check in the previous
attempt. That made the test Super Flaky. This is more noticeable in the
Release builds than the Debug build that I was working with.
(see https://codereview.chromium.org/471233003/)

With a Release build I was able to repro the flakiness (40% failure
over 200 runs). After the fix, I tested 500 runs and the success rate
is now 100%. I can't say it won't flake for other reasons, but
this fixes the main flake that was introduced. Historically,
the test does ocassionally need to be retried due to timeouts,
according to the bot log history. We may want to split the
CORS and the non-CORS test case to see if that will help w/ timeouts.

I didn't want to do that for this patch, since it would get
much more complicated and this is intended for merging to M38.

Retry with the race condition fixed. Also re-enable the test since
it got disabled by another CL for being flaky.

BUG=401755
BUG=315328

Review URL: https://codereview.chromium.org/524683002