Fixed flakyness of PolicyTest.FileURLBlacklist.
The problem was introduced in https://codereview.chromium.org/110643005.
The URLBlacklistManager used to build the blacklist in the FILE thread
but it's now using the blocking pool; however, FlushBlacklistPolicy
wasn't updated to flush the blocking pool instead of the FILE thread.

This test (and the URL blacklist tests too) have been flaking now
and then on every bot but this race is extremely hard to trigger.
DrMemory on Windows was a consistent repro though.

BUG=339240

Review URL: https://codereview.chromium.org/520703002