Layout Tests should support '*.test' URLs.
We need, for example, to test that ServiceWorkers can't be registered on
unauthenticated origins[1]. This patch allows '*.test' as a valid URL
for tests, and teaches content shell how to resolve them if (and only
if) the '--dump-render-tree' flag is set.

'.test' is a reserved gTLD[3], and is also used by the W3C's Web Platform
Tests[3]; it seems like the best choice of suffix for layout tests.

[1]: http://w3c.github.io/webappsec/specs/mixedcontent/#authenticated-origin
[2]: http://newgtlds.icann.org/applicants/agb
[3]: https://github.com/w3c/web-platform-tests#running-the-tests

BUG=410190

Review URL: https://codereview.chromium.org/529403002