Add NSString category for providing iOS6-style string drawing APIs.
The iOS6-style APIs have been deprecated, but the replacements are
more verbose and brittle so this category wraps the replacements in
methods similar to the iOS6-style APIs.

This category lives in a new target and directory: src/ui/ios, so
this CL sets up all the gyp/DEPS stuff and only adds a single
method. There will be more methods added in subsequent CLs.

BUG=364419

Review URL: https://codereview.chromium.org/497503004