Adds WARP support to Chromium on Windows 8+
Metro mode requires a GPU process but there are a few
tablets and other devices that don't have a GPU. Currently
we don't expose the menu option for Metro mode on these
devices. This CL exposes the Metro mode option for all
Windows 8+ machines, and if a GPU isn't normally available
it turns on WARP support for Metro mode only.

BUG=314954

Review URL: https://codereview.chromium.org/435383002