USB interrupt and bulk transfer tests using an echo gadget.
Add tests for interrupt and bulk USB transfers using a gadget device
that echos packets back to the host. This change depends on the removal
of the //content dependency in usb_service so that it no longer depends
explicitly on BrowserThread::FILE for I/O operations.

BUG=

Review URL: https://codereview.chromium.org/516003002