[Telemetry] Add 'important' to JSON output.
We have a plan in progress to upgrade the 'important' field to a combination
of two fields 'primary' and 'improvement_direction'. This will be a two-sided
patch that will require reworking the dashboard's handling of chart JSON. Since
the dashboard has been modified to accept chart JSON with 'important' values,
we want to update Telemetry to continue serializing values with this field for
now.

R=nednguyen

Review URL: https://codereview.chromium.org/541253002