mojo: Move initialization of the embedder in java.
The embedder was initialized in the constructor of CoreImpl. This is not
correct, because when it will be used in Chromium, the embedder is
initialized by the native code.

This remove the initialization from CoreImpl. Instead, for tests,
initialize it when loading the native library.

R=ppi@chromium.org

Review URL: https://codereview.chromium.org/544603002