Fix the buttons in the mojo demo window manager debug panel.
NWVM was using the ScopedCaptureClient to attach a capture controller. The problem with this is that the window manager already made one, and there is only one capture controller when you use SCC and this was screwing with event dispatch as coordinate systems between the different root windows were different (not sure this is a condition we've ever really designed for in aura). Anyway. Made NWVM use DefaultCaptureClient directly.

R=sky@chromium.org
BUG=

Review URL: https://codereview.chromium.org/541013002