Enable forced extension updates on NaCl arch mismatch.
This causes ExtensionUpdater to scan the contents of
installed extensions prior to initiating update queries. If
any extensions are found with a _platform_specific
directory that does NOT contain a subdirectory which
corresponds to the current NaCl arch, that extension is
placed into forced-update mode.

An extension in forced-update mode will be treated as
version 0.0.0.0 when querying the update server.

BUG=409948

Review URL: https://codereview.chromium.org/540673002