Fixes possible crash in tab dragging
This crash would happen if you select a couple of tabs, then click but
don't drag and hit escape or control-w. The crash would occur because
we would end up inserting the tabs a second time even though they are
already in the tabstrip.

BUG=360822
TEST=see bug
R=ben@chromium.org

Review URL: https://codereview.chromium.org/548703002