Fix focus on popup GTK dialogs
Previously we didn't have a call to gtk_window_present(), which would
mean that the widgets would become visible, but the necessary actions to
focus weren't always taken. This resulted in the active window not being
correct after opening the file or print dialogs.

BUG=399514

Review URL: https://codereview.chromium.org/543093002