cc: Remove and Create the correct tiles when resizing live tiles rect
When removing tiles, we need to remove everything outside the live
tiles rect even if it has borders inside the live tiles rect. The
DifferenceIterator includes borders when deciding if a tile is in the
included rect, so tiles that are only in the included rect with a border
were not being removed from the tiling. This lead to us having tiles
outside the rect, and to creating a tile that already existed.

When adding tiles, the same problem existed. The DifferenceIterator
includes borders when deciding which tiles are to be excluded, so a
tile that was inside the old live tiles rect with borders only would
not have existed, but it would be excluded by the DifferenceIterator.
This prevented us from making tiles that were inside the live tiles
rect.

Depends on: https://codereview.chromium.org/513903002/

TBR=enne@chromium.org, vmpstr@chromium.org
BUG=405427

Review URL: https://codereview.chromium.org/505913003

Cr-Commit-Position: refs/heads/master@{#292417}
(cherry picked from commit e3b7da9e87ee035c2c4529ccbfe541a7ffebd077)

Review URL: https://codereview.chromium.org/560443002