Supervised user interstitial: Don't destroy WebContents that might still be in use.
Before, the WebContents was closed directly from OnDontProceed. This breaks if the caller expects the WebContents to survive (happened e.g. when a new interstitial was being initialized).
This CL instead posts a task to the message loop.
Also, the tab won't be closed anymore if it's the last one in the window, as that would lead to closing the window.

BUG=411841

Review URL: https://codereview.chromium.org/550973002