Require build directories exist for some GN commands.
Some GN commands just query the state of the build. They generally take the
build directory as a first argument. But this is easy to forget and GN will
auto-create these directories if they don't exist. So I end up with directories
named "base/*" because I was trying to list all targets in base but forgot the
build argument.

This patch requires that the build directory contain a build.ninja file for
the "querying" commands to prevent this case.

Review URL: https://codereview.chromium.org/558763002