[Android] Preserve selection on hide/detach
Previously, the selection was cleared and the handles dismissed if
Chrome was backgrounded (via home or task switcher buttons) or the
screen was locked. This breaks any kind of multitasking with an
active selection. Instead, in such cases preserve the selection,
hiding the action bar and restoring as necessary when the view is
reattached or is reshown.

Also expose a programmatic way for preserving the selection when the
view loses focus, which is typically undesirable (tapping omnibox or
switching tabs), but not always.

BUG=406925

Review URL: https://codereview.chromium.org/516523003