Warn on executing command on unsupported elements.
Elements which don't have corresponding entries can be focused. (for example, when buttons are focused by <Tab>.)
When they are focused, canExecute of commands are called and warnings are emitted in CommandUtil.getCommandEntry().
This CL warn more specific (really suspicious) case. That is, a command executed on an element which doesn't have corresponding entry.

BUG=411790
TEST=manually tested

Review URL: https://codereview.chromium.org/550943003