Eliminate use of 'delete self' pattern in AttachmentUploaderImpl.
Prior to this change UploadState objects would synchronously notify
AttachmentUploaderImpl that they were done and could be deleted.  This
meant they were still on the call stack when their destructor was
invoked.  While it worked, it was subtle and might lead to bugs in the
future.

BUG=

Review URL: https://codereview.chromium.org/556083002