Schedule sliding transition after height change
On the BrowserOptions page, sections that are supposed to hide with an upward sliding transition instead hide instantly.

This is apparently because the style recalc that adds the CSS transition also sets the new section height (i.e., the first style recalc usually doesn't happen before the setTimeout is triggered in animatedSectionHeightChange_). To work around this, force a recalc before setting the new height.

BUG=394053
R=dbeam@chromium.org

Review URL: https://codereview.chromium.org/393943003