Remove the unmaintained performance monitor.
It's not launched nor are there plans to do so. The last commit
was Dec, 2012 and there was a thread about removing it on chromium-dev in May.

My immediate use-case is to quelch the periodic idle wakeups
due to the performance monitor, but at this point, simply removing it seems
preferable to digging in to fix the timer.

BUG=411488,130212

Review URL: https://codereview.chromium.org/547063003