Non-SFI mode: Quick workaround of unexpected CHECK failure.
Recently, crrev.com/418423002 is landed, but it has a bug in Non-SFI mode.
It introduces CHECK for the file token in ManifestService::OpenResource().
However, in Non-SFI mode, there is no NaClIPCAdapter, so the token is passed
from the renderer directly. (Actually the IPC channel is connected directly
to the renderer).
As a result, if the renderer fills the file token properly, it crashes.
As far as I investigated, it happens, at least, when the fast-path is triggered
(i.e. OpenNaClExecutable works in DownloadFile in ppb_nacl_private_impl.cc).
Anyway, we can ignore file tokens in Non-SFI mode, because it is for
SFI NaCl's validation cache.

BUG=394130
TEST=Ran trybots. Patched locally and run our Non-SFI NaCl app.
CQ_EXTRA_TRYBOTS=tryserver.chromium.linux:linux_rel_precise32

Review URL: https://codereview.chromium.org/560983003