Simpler logic to setup/restore Chrome command line
Simplified the logic to save, set, and restore the command line used to
launch a browser. For efficiency, functions are implemented as single
shell commands to run on the device.

This reduces about 20% of the time taken to start a new browser, and
overall 5% of the total running time of telemety unit tests.

BUG=379378

Review URL: https://codereview.chromium.org/547623002