Move the logic for creating data: URLs out of DesktopNotificationService.
This code is only used by the Notification constructor, and doesn't have
to be a public method on a class that's otherwise unrelated to it.

Also get rid of caching the UI manager in the service. All callsites of
the constructor passed in NULL, and BrowserProcess::notification_ui_manager()
actually does the caching for us.

TBR=jochen
BUG=

Review URL: https://codereview.chromium.org/561813002