TPMTokenLoader: Report both the disabled and enabled state of the TPM.
Also change from the observer pattern to the simpler callback pattern, as there is only a single notification type (TPM is enabled/disabled).
This simplifies ownership of the observer/receiver especially if multiple threads (typically UI and IO) are involved: A temporary callback with a WeakPtr can be used instead of an observer object that has to be owned by someone.

BUG=413219

Review URL: https://codereview.chromium.org/567553002