Force calling PostDisplayConfigurationChange() even if earyling out in UpdateDisplays
When changing from software mirroring mode to sinlge display mode, it
is possible there is no need to update |displays_| and we early out
UpdateDisplays(). But we still want to run the PostDisplayConfigurationChange()
cause there are some clients need to act on this, e.g.
TouchTransformerController needs to adjust the TouchTransformer when
switching from dual displays to single display.

BUG=chrome-os-partner:31868
TEST=tested on Big, after existing software mirroring mode, the touch location
     transformation is still correct.

Review URL: https://codereview.chromium.org/559213002