[ServiceWorker] Check is_uninstalled in ServiceWorkerRegisterJob::ContinueWithRegistration()
In the following situations the registration may be already uninstalled.
So existing_registration->GetNewestVersion() could return NULL and crash in ContinueWithRegistration().

- register the ServiceWorker
- open the page in the ServiceWorker scope
- unregister the ServiceWorker
- register the ServiceWorker to the same scope and close the page quickly.

ServiceWorkerRegisterJob::Start() is called to register the ServiceWorker.
 next_step is set to ContinueWithRegistration.
 GetUninstallingRegistration() returns the uninstalling registration.
 RunSoon() calls PostTask(next_step)

ServiceWorkerRegistration::Clear() is called because there is no controllee.

ContinueWithRegistration() is executed.
But at this moment the registration is already uninstalled.
And crashed while calling existing_registration->GetNewestVersion()->script_url().

https://codereview.chromium.org/506043002/ is reverted because of this crash.
So I created the revert of revert cl.
https://codereview.chromium.org/547743003

BUG=398355,408048
TEST=http/tests/serviceworker/unregister-then-register-new-script.html http/tests/serviceworker/register-same-scope-different-script-url.html in https://codereview.chromium.org/547743003

Review URL: https://codereview.chromium.org/556003006