Fix potential race-condition _beginthread
Sometimes ppapi tests fails due to race condition.
It's safer to use _beginthreadex than _beginthread. If the thread
that's generated by _beginthread exits quickly, the handle that's
returned to the caller of _beginthread might be invalid or point to
another thread. However, the handle that's returned by _beginthreadex
has to be closed by the caller of _beginthreadex, so it is guaranteed
to be a valid handle if _beginthreadex did not return an error.

R=bbudge@chromium.org
BUG=

Review URL: https://codereview.chromium.org/547543002