Adding option to move suggestion popup item text to omnibox.
Currently there is no way to move suggestion item text to omnibox
to refine suggestion in a single step.
With this change an arrow will be appearing along with suggestion text
and url. On clicking the arrow corresponding suggestion text is moved
to omnibox and suggestions are refined accordingly.

BUG=

Review URL: https://codereview.chromium.org/557263002