Cast: First stab at implementing adaptive latency
After a bunch of testing, it seems like there is a direct linear
relationship between RTT and the optimal playout delay. Packet loss
is also a factor, but since we expect packet loss to be low in
normal networks, perhaps we can make the formula as easy as RTT * 4.

This is disabled by default, the extension needs to set min_latency_ms
in order to activate it.

BUG=405339

Review URL: https://codereview.chromium.org/562653004