Enabling webview in app_shell.
This CL registers the webview tag in extensions, allowing it to be used inside app_shell. The webview browser sample should be working with this patch.

For now, not all the webview APIs are supported, the APIs that are not supported are: contextMenus, webrequest, declarativeWebRequest and permissions (videoCapture, pointerLock, geolocation).

BUG=352293

Review URL: https://codereview.chromium.org/566863004