Reland: Convert device::UsbConfigDescriptor and friends to structs.
These classes do not need to be classes and expecially don't need to be
abstract classes as this leads to a complicated implementation and
complicated tests. All USB devices no matter the platform will have the
same descriptor data.

This change follows the model of device::HidDeviceInfo.

Relanding this patch (reverted in crrev.com/294640) after restoring the
call to libusb_free_config_descriptor that was removed during
refactoring and caused failures in the Linux ASan LSan tests.

BUG=

Review URL: https://codereview.chromium.org/568943002