STABLE_TEST_SUITES: s/unit_tests/unit_tests_apk/
We used to add "_apk" to all targets from this list in the internal waterfall.
With a recent change by mikecase@ we no longer do that.

It works for most targets, but not for "unit_tests", which does not exist in GN.

This change would need to be reverted if we decide to switch back to GYP, but
it is unlikely to happen.

BUG=607455

Review-Url: https://codereview.chromium.org/1924893004
Cr-Commit-Position: refs/heads/master@{#390342}
(cherry picked from commit 9954a78c165554ffed0a978c5b84738f5fa83531)

Review URL: https://codereview.chromium.org/1928003002 .

Cr-Commit-Position: refs/branch-heads/2719@{#2}
Cr-Branched-From: db7757b63f29696b9b465949f9704a0d9d149f2b-refs/heads/master@{#390251}