Ensure that we always pass a non-null WTF string to mojo calls.
This fixes a crashing bug, where an empty WebString is passed to
WriteImage for the title parameter. When mojo serializes this it is
converted to a WTF::String that mojo detects is null and sends null
across the wire. On the receive side it is expecting a non null
string so deserialization of the message fails.

This patch sends an empty string instead of a null string, which
matches the pre-mojo state of the code that used to call Utf16() on
the WebString.

(cherry picked from commit b1dae1bfc9de4eff221b49bf631e2b2208f2830a)

Bug: 794512
Change-Id: I3eb8d6c23ca6a9ff435802521366f03f7cb60d43
Reviewed-on: https://chromium-review.googlesource.com/826826
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Stuart Langley <slangley@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#524155}
Reviewed-on: https://chromium-review.googlesource.com/827706
Reviewed-by: Krishna Govind <govind@chromium.org>
Cr-Commit-Position: refs/branch-heads/3294@{#3}
Cr-Branched-From: 1e4786a323e8eca4d5a66474bf23aa7fa80fe7ae-refs/heads/master@{#523976}