Lower Goma jobs from 500 to 50 for Cast Linux, which is not as beefy.
LUCI CI compiles currently take 1hr-1:45 whereas LUCI try compiles take
5-15min. That discrepancy is bigger than expected: while Buildbot CI
compiles are also longer than Buildbot try compiles, the difference is
20min-1hr vs. 2-5min.

The hypothesis is that the cause of this discrepancy is LUCI CI running
-j500 vs everything else running -j50, while everything is on 8 core
machines. It's possible that that's not the root cause, but ruling out
the number of jobs as a difference is an easy thing to do here, and this
CL aims to gather data for better basis of comparison.

Bug: 790118,809744
Change-Id: I8c1fe7c1e37166d251534bbea5e7ca839e1d4f06
Reviewed-on: https://chromium-review.googlesource.com/929897
Commit-Queue: Jao-ke Chin-Lee <jchinlee@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>