Fix Windows first run crash related to locale.
The crash would happen if you install Chrome of a specific language
and run it for the first time on a Windows OS of a different
language.

It was caused by code setting the language pref in-between the two
places where it was queried - such that the first time it was
queried resulted in a different value than the second.

This triggered the CHECK that was introduced by this CL:
https://chromium-review.googlesource.com/c/chromium/src/+/1349770

Which correctly identified the problem. This CL fixes the issue
by moving the importing of the first-run locale setting to
earlier.

TBR=asvitkine@chromium.org

(cherry picked from commit 5814800fa70a7a7a7975c9943cdf0c310873321c)

Bug: 908791, 908114
Change-Id: Ie2e0178f298f78d38e0bc7f0b58c2205bb617561
Reviewed-on: https://chromium-review.googlesource.com/c/1352482
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#611480}
Reviewed-on: https://chromium-review.googlesource.com/c/1352673
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/branch-heads/3624@{#3}
Cr-Branched-From: b54fb9cc3bbc522b006c38f4a3e88cbda2be1c59-refs/heads/master@{#611430}