[M71] Roll ICU to 6939c541
The roll has 3 CLs.

The first reintroduces a fix for the timezone detection in remote
desktop connection on Windows (https://crbug.com/854387). It's
included in the first M71 stable release, but
was reverted in https://crrev.com/c/1370582 for the 2nd M71 release
as an emergency fix for https://crbug.com/913298 .

The second and the third (combined) are a just 1-liner to fix
https://crbug.com/913298  in a proper manner without regressiing
https://crbug.com/854387 .

https://chromium.googlesource.com/chromium/deps/icu.git/+log/751b643..6939c54

Bug: 854387, 913298
Test: See the bug.
Change-Id: I6637898926341e3f0ca0268e5f381a813c786342
Reviewed-on: https://chromium-review.googlesource.com/c/1393338
Reviewed-by: Michael Moss <mmoss@chromium.org>
Cr-Commit-Position: refs/branch-heads/3578@{#930}
Cr-Branched-From: 4226ddf99103e493d7afb23a4c7902ee496108b6-refs/heads/master@{#599034}