Fix bug with WorkQueueSets::OnPopQueue called on a non-min queue
WorkQueue::RemoveAllCanceledTasksFromFront and
WorkQueue::DeletePendingTasks where calling WorkQueueSets::OnPopQueue
if something was deleted. This is not safe unless the queue is the
the minimum one in the set.  We have a DCHECK to try and enforce that
and I'm not sure why the fuzzer didn't hit it.

This is a speculative fix for a bug found by the fuzzer.

Bug: 936393
Change-Id: I9be9e0ef18f47b8dfbc095eff849d8307b5eeffd
Reviewed-on: https://chromium-review.googlesource.com/c/1492114
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Auto-Submit: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#636411}(cherry picked from commit f0a6246827895ec7068a2f72d3fa5da44f8776a2)
Reviewed-on: https://chromium-review.googlesource.com/c/1495577
Reviewed-by: Krishna Govind <govind@chromium.org>
Cr-Commit-Position: refs/branch-heads/3720@{#8}
Cr-Branched-From: 08999f291d1d7ac1f2ab5b77f9ead33a26a0f663-refs/heads/master@{#636200}