Make ConnectJobs take CommonConnectJobParams as a const pointer.
Since this parameter is a constant and the same for all ConnectJobs in
a socket pool, there's no reason to copy it to all of them. The
CommonConnectJobParams object is now owned by the SocketFactory, though
I'm planning on moving it up to HttpNetworkSession, to make it easier to
make ConnectJobs without a socket pool.

This CL also moves SocketTag() from CommonConnectJobParams to being a
ConnectJob parameter, since it can vary between ConnectJobs in the same
SocketPool.

Bug: 927100
Change-Id: I1ad39b4e0b8b64a9e6cf82b519cf003f41020bd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1508579
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/master@{#638884}