Limit how many FormData::fields the passwords parser accepts
As discovered by a fuzzer, feeding an excessive amount of fields
inside FormData to the parser producing PasswordForms can cost a lot
of memory. At the same time, password forms with thousands of fields
are unlikely to be useful.

Therefore this CL makes the FormDataParser::Parse method exit early if
more than FormDataParser::kMaxParseableFields (10K) fields are
contained in the FormData. The value was chosen arbitrarily, so that
it is obviously large enough for all legit forms, and not large enough
to produce significant memory pressure.

There already is a similar bound in
autofill::form_util::kMaxParseableFields, but the newly added is not
linked to it for several reasons:
(1) Parsing password forms and other forms could potentially need
    different limits.
(2) The limit in autofill::form_util::kMaxParseableFields may be too
    low (see https://crbug.com/935694#c4).
(3) That constant is defined inside the content layer, on which the
    component core must not depend.

Finally, for fuzzing based on protobuf fuzzer, a limit derived from
FormDataParser::kMaxParseableFields was added, to avoid unnecessary
wasting of computational resources. This introduces the assumption
into the fuzzer, that the parser indeed interprets
FormDataParser::kMaxParseableFields as the limit, but the fuzzer is
still able to supply slightly bigger amount of fields to check
handling of that.

Fuzzers which use non-proto-based generator of FormData do not need a
similar change, because that generator already caps the number of
fields. In fact, it might be a good idea to extend that cap beyond
FormDataParser::kMaxParseableFields, but that should be in a separate
CL to clarify the cause of potential impact on fuzzing with that
generator.

Bug: 935694
Change-Id: I575ee1b63be418f5cc7ef9e9516b95d5ef799c78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1509392
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Commit-Queue: Vaclav Brozek <vabr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#638973}