Fix flaky MachineLevelUserCloudPolicyPolicyFetchTest
The MachineLevelUserCloudPolicyPolicyFetchTest is timeout on Win7. It
might be caused by the race condition that when we check the
|last_policy_timestamp| of CloudPolicyClient, the policy has been fetched
but not validated. Hence the OnPolicyFetched event is never triggered
after observer created.

The fix will be listen to the CloudPolicyStore for any valid DM token test
case. Because OnStoreLoaded will always be called when the timestamp is set.
And only observe to CloudPolicyClient iff the DM token is not valid in which
case there is no CloudPolicyStore event.

Bug: 939261
Change-Id: I5e7698643474ca639bb516f89363ba32537f0782
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1509642
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639001}