Don't use ink drop container bounds in InkDropMask creation
CreateInkDropMask() is called immediately upon bounds change. However,
LabelButton's ink drop container bounds are updated on the next layout.
As a result, if the BrowserAppMenuButton is resized (e.g. due to
maximizing the window) the ink drop container bounds will be wrong in
the call to CreateInkDropMask(). Hence, we should use the
BrowserAppMenuBounds directly.

This fixes a DCHECK crash that happens when the window is maximized
while an in-product help promo is showing.

Bug: 939536
Change-Id: I5b8619c803dbc865be88bf46a725b63d0648a230
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1510773
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639233}