Fold //ash/wm/focus_rules.cc into //ash/wm/ash_focus_rules.cc
The files were split in 2016 when the mustash project introduced
ash/common and WmWindow. Now that we use aura windows everywhere in
ash we don't need the split any more.

This reduces the number of files with "focus_rules" in the name,
and reduces the number of bare functions in the ash namespace.

I took AshFocusRules out of namespace ash::wm because the majority
of code in //ash/wm is in namespace ash, not ash::wm.

Bug: none
Test: ash_unittests
Change-Id: I6e51fda4e018ee041e9c59eb4a67e615c8a6fb8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1512662
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639267}