OMT service worker fetch: Fix CSP related web_tests
If OMT SW is off, CSP is supplied following ways:
- Non-installed script: WebEmbeddedWorkerImpl fetches the script on the main
  thread using ClassicWorkerScriptLoader. CSP comes from the loader and is
  passed to WorkerGlobalScope's constructor.
- Installed script: ServiceWorkerGlobalScope retrieves CSP from
  InstalledScriptsManager on the worker thread and set it via
  ApplyContentSecurityPolicyFromHeaders()

When OMT SW is on, non-installed script case doesn't work because we use
different ClassicWorkerScriptLoader in WorkerGlobalScope::ImportClassicScript().
This CL adds an enum that indicates where CSP comes from. WorkerGlobalScope
uses CSP from the response of the main script when the enum is set to use
the response's CSP.

Bug: 924043,935264
Change-Id: I20ea7ee4a4934cc634a2b1497e24439d07ce26b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1493694
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639308}