BlinkMemoryMgt: Annotate Blink's css with the macros of Allocator
Although the previous CLs have annotated most of classes in the css
directory, there are still some classes which weren't anonated. So this CL
annotates some classes of //third_party/blink/renderer/core/css with
the macros of Allocator.

After some analyzes, it was concluded that for the case of GeneratedImageCache
and CSSSyntaxComponent, class, it makes sense to use DISALLOW_NEW because
it has been used as a member variable with a type of reference.

Lastly, the rest of the classes this CL annotates make sense to use
STACK_ALLOCATED.

Bug: 919389
Change-Id: I60c08eaca40d5c17fc01e3c3187730c64bcd5b61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1514332
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Gyuyoung Kim <gyuyoung@igalia.com>
Cr-Commit-Position: refs/heads/master@{#639397}