[AF Wallet] Implement deleting old orphan metadata
This CL implements the last part of the wallet metadata sync bridge as
designed in the DD. While it solves an extremely corner-case issue, it
is still important for correctness of the overall design.

This CL also fixes a bug in the ParseWalletMetadataStorageKey() helper
function.

Bug: 853688
Change-Id: I4938b45aab68b2a48f90912fccb77782397a8a3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1505499
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639451}