[Sync::USS] Fix bookmarks handling of server-illegal titles
For historical reasons, "", "." and ".." weren't allowed on the server.
Clients should always append a blank space before committing a bookmark
with such a title. The opposite should also be done when processing
remote updates.

Before this CL: the USS version didn't incorporate the same
functionality, which resulted in many user reports that blank spaces
are added to nameless bookmarks when syncing across multiple devices,
where some of them are on USS and some are on the pre-USS
infrastructure.

After this CL: USS version of bookmarks is implementing the same
functionality as the legacy clients.

Bug: 936314
Change-Id: Id28b4af03dcce27cbee0f79410f765b48f9aab2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1514832
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639517}