Deprecate unused SyncEnums::Action entries
These are never sent by the server, so let's get rid of them.
Two tests required updates, because they simulated a dashboard reset
in a non-realistic way.

Note that there's a separate ClientAction enum which has many similar
entries, but these are generated by the client. (Some of them are
becoming unused after this CL and will be cleaned up in a followup.)

Bug: 940502
Change-Id: I60acf188ccff03cb1ae277c88ef6da971796631b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1514755
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639531}