[FCP++] Deprecate Last Text Paint
Last Text Paint started off as an exploration for a better paint metric.
We built it and correlated it with human rating data, and discovered that
this metric didn't reflect user experience of page speed very well. So We
decided to deprecate Last Text Paint.

This CL is to remove the Last Text Paint from the Chromium code base.

Bug: 940039
Change-Id: Ib0238cec078530afed35a84f25331a94d301b35c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1512935
Reviewed-by: Steve Kobes <skobes@chromium.org>
Reviewed-by: Bryan McQuade <bmcquade@chromium.org>
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639608}