Remove WebGPU as a "test" feature in Blink.
This flag previously made WebGPU enabled in Blink when running tests.
However, it does not set --enable-unsafe-webgpu in Chromium as well.
This would lead to a problem where tests in Blink have WebGPU exposed
but they are not enabled in Chromium.

Furthermore, we don't want WebGPU to always be enabled in the tests yet.
All WebGPU tests that are not run with --enable-unsafe-webgpu should fail
(and not crash). They should pass with --enable-unsafe-webgpu.

Bug: 877147
Change-Id: I9c76918a39c90dd976339eb39c846bc2c36f61a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1512177
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Austin Eng <enga@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639648}