Removes the use of --test-filter command line argument in gpu_integration_test_unittest.py
This CL will change a some unit tests so that it calls a function that runs all the
tests in unittest_data.integration_tests.RunTestsWithExpectationsFiles. This will remove the
use of the --test-filter argument from those tests. Removing the use of the --test-filter
argument will fix the issue causing crrev.com/c/1503875 to be reverted.

Bug: chromium:938586
Change-Id: I8433f94b6818294fed9ee50d6b321ed775bc7b09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1516252
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Cr-Commit-Position: refs/heads/master@{#639706}