Migrate sync migration tests away from python server
These are the remaining tests that make use of the test server
in components/sync/tools/testserver/ which can now be deleted in
follow-up patches.

The patch also reenables multiple tests that have been failing
for years, and others that were reported flaky (let's hope that
flakes no longer reproduce).

In order to achieve this, LoopbackServer had to be extended with
support for migration (and corresponding version tracking). This
has been achieved in a backward-compatible way, such that local
sync (enterprise use-cases) is not affected.

Bug: 406545,918124,403778,565559
Change-Id: I73bda43bc95a000c7bf390a6f93b39035c51cdf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1514553
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639859}