sync_ui_util: Don't check for NOT_MY_BIRTHDAY error
Before this CL, sync_ui_util checked for a NOT_MY_BIRTHDAY protocol
error (in addition to !IsSyncRequested()) to determine whether we're in
the "Sync was reset from the dashboard" state. However, that's
a) unnecessary; !IsSyncRequested() is a sufficient condition (since
   resetting Sync from the dashboard is the only way to get into a
   primary-account-but-sync-not-requested state), and
b) incorrect, since the protocol error gets reset when
   Sync-the-transport starts up.

So this CL removes the NOT_MY_BIRTHDAY check. This is also consistent
with the check in GetSyncedStateStatusLabel (which provides the
corresponding message).

As a related cleanup, this CL introduces a GetClientAction() function
which extracts the action from QueryDetailedSyncStatus(). That makes
it clear that this is the only part of the detailed status that's
relevant here.

Bug: 940437, 911153
Change-Id: I00e925073bcfa7b1e13bb6df0b3017a0073fca7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1517652
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639919}