[ElementTiming] Allow timing for video poster image
This CL allows ElementTiming to observe video poster images. Since
VideoPainter delegates the poster image painting to ImagePainter, this
is just a matter of allowing the node used to be a video element. A test
showing that the video poster image is observed is added.

Bug: 879270
Change-Id: Ic5f354180343cb5e3763d715a87290a540fd8e64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1516782
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#639978}