Create AppThemeColorProvider and have ToolbarTablet's MenuButton use it
The AppThemeColorProvider is a theme provider for UI that only
responds to standard or incognito theming (no website branding). This
includes the bottom toolbar and tablets (and eventually the tab switcher
top toolbar).

ToolbarTablet's MenuButton used to listen to the TabThemeColorProvider.
It mostly worked fine since tablet tabs either have a standard or
incognito theme. The issue only arose in the accessibility
tab switcher were there is no tab to theme the menu button. When
the incognito TabLayout toggle is toggled the menu button would not
react to the color change since it is listening to the previous tab's
color.

Bug: 938220
Change-Id: I03941240b4b75f7b74595c67e3f32bd88b0a03a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1512271
Commit-Queue: Pedro Amaral <amaralp@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640058}