Keep SharedContextState alive as long as SharedImageFactory
We want a SharedImageFactory for the GLES2Decoders (passthrough and
validating) to exercise SharedImage endpoints. To do so, We need to keep
the SharedContextState alive as long as the factory. If we do not, the
virtual context associated with SharedContextState will go out of scope,
and will SetCurrent(nullptr) and SetCurrentGL(nullptr) in ~GLContext.

Bug: 940184,939144
Change-Id: I03e37f3e98fd735755d68b058a5f6a0c663d5ae6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1518450
Commit-Queue: Jonathan Backer <backer@chromium.org>
Commit-Queue: Peng Huang <penghuang@chromium.org>
Auto-Submit: Jonathan Backer <backer@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640072}