Update PageInfo to not use a SecurityInfo
This is part of a refactoring to remove the security_state::SecurityInfo
struct. See the commit message of
https://chromium-review.googlesource.com/c/chromium/src/+/1500414 for
more on why. This CL updates Page Info to use a SecurityLevel and
VisibleSecurityState rather than a SecurityInfo. The one piece of
information included in a SecurityInfo but not VisibleSecurityState is
whether there is a SHA1 cert in the chain, so a helper method is added
to the security_state component to compute this.

Change-Id: I7e99edd488d08e72821f19b2710c8914e7f6354a
Bug: 939588
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1500783
Commit-Queue: Emily Stark <estark@chromium.org>
Reviewed-by: Christopher Thompson <cthomp@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640080}