PropertyTreeBuilder: Check subtree for rounded corners
This patch adds code to check the subtree of an effect node for rounded
corners. If any descendant has a rounded corner that may cause blending
artifacts then it is reported to the parent effect node which triggers
the generation of a render surface on that parent effect node.

Also updates the unit tests.

Bug: 903486
Change-Id: I94d80f3a73f1e312003d23bf93dce5b08a89a8b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1512788
Commit-Queue: Malay Keshav <malaykeshav@chromium.org>
Reviewed-by: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640085}