aw: Implement presentation feedback for webview
Webview's implementation makes the assumption that the frame token is a
sequence and that it is ok to drop frames in the middle of the sequence.
See LayerTreeHostImpl::DidPresentCompositorFrame.

Combine the draw constraints post task with this one, which means that
post task will now be delayed until after DrawAndSwap on render thread
most of the time.

Perf note. This is expected to increase the number of tasks per frame
on the UI thread for webview by 1.

Bug: 938956
Change-Id: I2bb895c23049441560118b0190e1e2d2a8534bd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1512340
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Reviewed-by: Tobias Sargeant <tobiasjs@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640234}