RenderDocument: Add base::feature and tests.
Currently, a RenderFrame represents neither a document nor a frame, but
a frame in a process. A new RenderFrame is created on each cross-process
navigation. On navigation commit, the previous and next RenderFrameHost
are swapped.

We would like a RenderFrame to have a document semantic instead. It
requires switching RenderFrame on same-process navigations as well.

This CL:
 * Adds the base::Feature flags to enable/disable RenderDocument.
 * Adds a set of tests.

This doesn't implement RenderDocument at all. This will come later.

(WIP) document:
https://docs.google.com/document/d/1C2VKkFRSc0kdmqjKan1G4NlNlxWZqE4Wam41FNMgnmA/edit?usp=sharing

Bug: 936696
Change-Id: Ic865129c51e873ad20c4718f918cf5abf42d4086
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1511077
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640248}