Use a per-frame task runner at IdleStatus's mojo binding
This CL switches task runners at several places from per-thread to
per-frame. The purpose of this change is to make it possible to freeze
tasks for bfcache.

The actual test that executes JavaScript on the task runner is:
external/wpt/idle-detection/interceptor.https.html

Bug: 913912
Change-Id: I12f64f9e071b841be8a0c1d9003edc81be03f911
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1514460
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640254}