Defer sending a request to close the page for remote main frames
The same arguments apply as for a local main frame, in that JS could
run a nested message loop and the browser may close the RenderWidget
that the JS is running from before the stack unwinds, causing a crash.
So we post a task to send the close request to the browser from
RenderViewImpl in the remote main frame case, matching the local main
frame/popup case.

R=dcheng@chromium.org

Bug: 912193
Change-Id: I5adc9694fc8d9898a98993adafa48b98f04a97de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1480270
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640363}