[Fuchsia] Update LoadUrl() and CreateContext() APIs to use extensible FIDL tables.
In a previous CL, chromium.web.LoadUrlParams and
chromium.web.CreateContextParams were replaced with an extensible
version with the suffix "2". Now that the out-of-tree callers for the
old APIs have been updated, this replaces the old APIs to match the
newer APIs and replaces the Chromium-tree callers.

Bug: 931831
Change-Id: Id1cf567ad1989ffa309cd01470fb66e0752521ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1516887
Commit-Queue: Fabrice de Gans-Riberi <fdegans@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640418}