Do not create PrefetchGCMAppHander in reduced mode
If in reduced mode, we cannot create PrefetchGCMAppHandler. Instead we pass the
prefetch service a closure to create the gcm app handler on demand.

This cl also changes the signature of the getter for PrefetchGCMHandler in
PrefetchService to require a profile to be passed in. This ensures that:
1) the getter can only be called in full browser mode (there is no profile in
reduced mode).
2) if PrefetchGCMHandler hadn't been created yet, then it can be
created on demand using the aforementioned profile.

Bug: 934337
Change-Id: I7e813f425e0ba8687f9b519b4300fa7eec014929
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1495737
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Reviewed-by: Jian Li <jianli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640470}