Setup API for fast rounded corner path
This patch adds an API for a fast rounded corner on any given layer. If
this flag is set on the layer, the rounded corner for that layer will
not have a render surface due to it having any rounded corner. It will
also not contribute to the decision making on its ancestor along the
parent chain for having a render surface(for intersecting rounded
corners).

The patch pipes the value and sets it on the draw property of the
respective layer impl.

Added unittests

Bug: 903486
Change-Id: I5936e1dedb8793c89c1625fcb234509ee9ce7cea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1519467
Commit-Queue: Malay Keshav <malaykeshav@chromium.org>
Reviewed-by: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640553}