[OnionSoup] Reland:Build typemapping for blink::FormDataElement
The original CL has been reverted because it failed on newly add
unittest: "EncodedFormDataMojomTraitsTest.Roundtrips_FormDataElement".
When |original1| is serialized, the uninitialized variable:
|original1.expected_file_modification_time_| will cause this error.

A fix solving this problem is added in this CL. For convenience, PS#1
shows the original version of code, and PS#2 is the fixed version. What's
more the bug number is changed to 911930.

The original CL's description:

> This CL typemaps blink::FormDataElement with network.mojom.DataElement.

>Bug: 860403
>Change-Id: I7611e2cedf479f1da429fe8a026a290be447de5e

Bug: 911930
Change-Id: I47509fa0b84478e692ef221d3dc3829fbc5e5eea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1522554
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Richard Li <richard.li@intel.com>
Cr-Commit-Position: refs/heads/master@{#640668}