Create frame: replace proxy_routing_id by previous_routing_id.
When a RenderFrame is created, it keeps the id of the RenderFrameProxy
it needs to replace. With RenderDocument(https://crbug.com/936696): we
would like to implement same-process RenderFrame swap. It means the
RenderFrame can now be used to replace a RenderFrame.

The RenderFrameHost's id and RenderFrameHostProxy's id are coming from
the same integer sequence generator. It means the same id can be used to
represent any of them.

This CL renames |proxy_routing_id| by |previous_routing_id|. It is
still used to represent a remote frame, but it will soon be used to
represent a local frame as well.

This CL doesn't change Chrome's behavior.

Bug: 936696
Change-Id: Ifabde8db53bfed3429b882f0d40abf3c1a775b78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1515576
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640670}