Reset loader in NavigationRequest::OnRequestFailedInternal.
The loader used to be reset in NavigationRequest::OnRequestFailed().
This enforces the assumption: "If the request failed because of the
loader, no more calls are expected from the loader anymore".

It might worth doing it more broadly in OnRequestFailedInternal. The
request might failed at the initiative of the NavigationRequest first,
in this case, it shouldn't failed later at the initiative of the loader.
The NavigationRequest probably doesn't support failing while failing.

Bug: 936962
Change-Id: Id1b1174b3a907ffcc8516016ee83b9533c913649
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1517593
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640691}