DiceTurnSyncOnHelper: Call SetSyncRequested for the right profile
Previously, it called SetSyncRequested from TurnSyncOnWithProfileMode,
even if the mode was NEW_PROFILE - so it effectively set it on the
wrong profile.
Now it's called from SigninAndShowSyncConfirmationUI, where we know
we're in the correct profile.

This also switches to use SyncUserSettings instead of writing to
SyncPrefs directly. SyncPrefs are mostly an internal sync thing which
no clients should have to mess with directly; SyncUserSettings is the
proper interface.

Bug: 906034
Change-Id: I2787bc9ceabcff8ffb429f5fe612fa38de8d71d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1523399
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640739}