More CHECKs to help narrow down crashes in a closed RenderWidget.
Determine if the RenderWidget is for main frame or child local root,
and determine if we're Closing the RenderWidget without detaching the
frame, which in a child local root case would leave blink running with
a Document pointing to the closed RenderWidget.

The test being removed is testing a code in RenderFrameImpl that no
longer exists. It is also testing a scenario that we believe should
not happen - closing a child local root's RenderWidget from the
browser.

R=dcheng@chromium.org

Bug: 939262
Change-Id: Ib13c916eacb0aecd0394ba378ce69809eb7c9b89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1519066
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640764}