CacheStorage: Check content settings for storage permission.
This CL plumbs a content_settings path to check if the current window
is allowed to access cache_storage.  This plumbing is identical to
other storage types.

Currently this mechanism uses a sync IPC (again as done by the
other storage type) to determine if blocking is necessary.  It would
be nice to avoid this, but it would require perhaps a larger
refactoring of content_settings.  For now we mitigate any performance
penalty by caching the allow-vs-block decision on the first access
attempt.

A browsertest is added to verify the behavior and prevent regressions.
It was verified to fail without the accompanying code changes.

Bug: 940931,940902
Change-Id: Id0e9b00f6a66a5e9f849d1deeb0705bd871ae54c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1518592
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640778}