Fix sync not recovering if custom passphrase cleared
Sync datatype HISTORY_DELETE_DIRECTIVES should stop if custom passphrase
is set. Prior to this patch, that was implemented by issuing a datatype
policy error, but that also meant (in the USS codepath, i.e.
ModelTypeController) that the datatype cannot recover from that state
without a restart of the browser.

This was somewhat weird or asymmetrical because, during normal startup,
the condition is detected in the controller's ReadyForStart(), which
doesn't enter an error state.

A cleaner solution for this requirement is to treat the passphrase state
exclusively as precondition, and when it changes, report it via
ReadyForStartChanged(). This means no error states are involved, and
that the controller can resume normally if preconditions are met (i.e.
custom passphrase cleared), fixing the issue at hand.

Bug: 941733
Change-Id: Icf5ac07bd52260182c71ae6ca303c264699f3f83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1523387
Auto-Submit: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Jan Krcal <jkrcal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640830}