Revert "Add temporary simplified process lock check in the renderer process."
This partially reverts https://chromium-review.googlesource.com/1507195
since we don't need the instrumentation anymore. The parts that are kept
are plumbing of the network error code into Blink and some minor
refactoring in RenderFrameImpl - both of these at the request of the
reviewers of the CL as useful parts to stay.

Bug: 931895
Change-Id: I71b7cdf82f76d4dfe7f9110f6bd077cc673e14c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1524095
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640896}