Adapt omnibox to not use a SecurityInfo
This is part of a larger refactoring to get rid of the SecurityInfo
struct and change consumers to use a SecurityLevel and/or
VisibleSecurityState instead. See
https://chromium-review.googlesource.com/c/chromium/src/+/1500414 for
more of the reasoning/background.

This CL changes the LocationBarModelDelegate to expose a SecurityLevel
and a VisibleSecurityState instead of a SecurityInfo.

Change-Id: Idfcaf34180803b83574b4595ddeeb59022285c4b
Bug: 939588
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1506602
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640912}