Introduce delayed memory purge after freezing
Decoupling memory purging from freezing is necessary in order
to make freezing lower-cost. Currently, memory purging is performed
immediately after a page is frozen. This approach results in more work
for the CPU immediately following a freeze and makes freezing more
expensive than necessary. A lower-cost freezing functionality
enables freezing to be performed more aggressively, and allows for
better use of the CPU. Specifically, short-term freezing should be
performed within minimal cost, allowing us to better allocate CPU time
in times of higher workload.

However, completely removing memory purging from freezing is not an
option, since there is a measurable memory benefit (10-50 MB)
associated with freezing. Using a delay ensures we retain the memory
benefit associated with freezing, while making short-term freezing
lower-cost.

To support this effort, a timer responsible for triggering the memory
purge was introduced in MemoryPurgeManager. Furthermore, a memory purge
will only occur if all pages are frozen. This approach ensures that
once a purge occurs, there will be no need to purge the renderer again
and prevents unnecessary purges.

In order to measure the effects of the new behaviour, the new
functionality is being introduced as an experiment. A FeatureParam has
also been introduced to allow for the variation of delay time.

Bug: 926186
Change-Id: If497e4e8de914136179c75de0ac032c058a2a527
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1457023
Commit-Queue: Aditya Keerthi <adityakeerthi@google.com>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640923}