[DevTools] Roll inspector_protocol to 1e65f2e3a9759b799ca3a87a41dcc9f58fdd720f.
This is a reland, but I can't use the reland feature unfortunately.
The revert PR was
https://chromium-review.googlesource.com/c/chromium/src/+/1522494

This new roll lowers the stack limit for the json / cbor parser,
so I'm hoping it will work on win-asan. I have no reasonable
way to try this at the moment however.


(1) Make a source_set in content/browser/devtools for the encoding_test,
and add the test to content_unittest. This will ensure that further
changes to our test won't break after the try-bots pass a submission
to Chromium. I've broken the tree twice with that, previously.
(2) third_party/inspector_protocol/BUILD.gn is now a different file
than it is upstream. But the good news is it's really really simple.
I've updated roll.py accordingly.

Also: 
Change-Id: I935624d156b2ffc9df12fd5f1d745da8a10998c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1524917
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Pavel Feldman <pfeldman@chromium.org>
Commit-Queue: Johannes Henkel <johannes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#640981}