Print Preview: Move destination and invitation stores, user info
print-preview-app and print-preview-model owned properties used
exclusively by destination-settings. This CL moves these properties
and the user-info element into destination-settings. It also converts
recent destinations to a setting, so that sticky settings updates
can be done the same way they are for other settings, and so that
destination-settings can own this property as well.

This CL also adds a new destinationState for the print-preview-app to
correctly update its state based on destination events.

These changes will facilitate more easily refactoring print-preview-app
in future, by reducing the properties and logic it is responsible for.

Bug: 942345
Change-Id: I1cbad055b40a2ba1ca4e067284320c2752f47c6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1517045
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641005}