[Extensions Bindings] Remove JS-bindings C++ messaging code
The old JS-based bindings for extension messaging had a handful of
native helpers, which are now obsolete with the full native
implementation. Remove the JSRendererMessagingService and ExtensionPort
classes entirely, and remove the vast majority of MessagingBindings.
MessagingBindings::BindToGC() is used outside of messaging, and so needs
to remain (but we should rename the class or move the method in a
follow-up).

Bug: 938998
Change-Id: Ia697f4e19b8ac03ef87a62d9b8385cced8f08fbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1523981
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641013}