Use per-frame task runner at extnsions' GCCallback
This CL switches task runners at several places from per-thread to
per-frame. The purpose of this change is to make it possible to freeze
tasks for bfcache.

The callback of GC in extensions executes JavaScript, but this callback
is now executed on the default task runner. This CL replaces this with
a per-frame task runner.

Some tests hit this: e.g.
extensions_browsertests --gtest_filter=WebViewAPITest.TestDialogConfirmDefaultGCCancel

Bug: 870606
Change-Id: Ic1c2968a2e9e07629e7c117e1e1c4e97c2df0457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1514516
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641104}