[AF Wallet sync] Ensure proper order of notifications to metadata bridge
This CL fixes a speculative bug in interaction between the wallet data
bridge and the wallet metadata bridge. The bug occurs when the bridge
deletes and re-creates a particular entity (because it is different than
before). If these notifications are sent out in the wrong order (first
ADD, then DELETE), we end up without the metadata entity.

It is speculative because I am not sure that it can happen in reality
that payments sends the same card with different data. Anyway, it feels
safer to assume less.

Bug: 928758
Change-Id: I74a2d6c68107fffdde18a69810828d629ca4e50f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1518575
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641106}