Switch Bing search-by-image to upload image data instead of URL
The current Bing Search-by-image URL is the one for searching based off
of an image's url. However, we now have cases where we want to search by
the image data, which uses a different parameter. Also, in all cases
where we search with the url today, we also have the image data already.

This CL also adds support for converting the image data to a base64
encoding, as that is what Bing uses.

If there is some process for updating search engines, let me know. I
also thought about a way to support both versions of the image search
url (search by url and image data) but couldn't come up with a good
solution.

Bug: 938669
Change-Id: I978c7a45bcd37835cda15f43a93bec0b8737b6bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1508394
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#641153}