Dropping a stale command line switch.
We had a command-line switch to disable UKM reporting from the renderer
composistor. As denoted by the switch's name, it was only intended to be
used as a workaround for some browser tests. As noted in
content_switches.cc, once https://crbug.com/761524 was fixed, we
expected to be able to remove this switch. At least on my machine,
removing the flag doesn't break any of the content_browsertests.

Bug: 761524
Change-Id: If5835c6319186c2697afe2852563b07718ee2219
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1525001
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Tom McKee <tommckee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641219}