[AF Wallet Sync] Commit changes to WebDB on remote update
This CL adds the missing commit call to the wallet_data sync bridge.
The open transaction is normally closed on shutdown but it does not get
closed on Android when the process gets killed.

This situation was never dealt with in any of the autofill sync code.
Thus this CL introduces a CommitChanges() function that is accessible
by the sync code.
(The bug affects all USS and Directory autofill sync data types; Wallet
is the most affected one because the payments backend has a very
limited traffic capacity. Further data types will be addressed by
separate CLs.)

This CL is a merge candidate. The bug is causing a traffic increase to
the Payments backend on Stable so it needs to get fixed soon.

Bug: 942435
Change-Id: I48b88d3c3ed646b7dc8c6c3b0394bf33e342c097
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1526262
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641240}