Ensure clip changes are propagated in PrePaintTreeWalk
When a clip changes, PrePaintTreeWalkContext::clip_changed is set and
propagated to all descendants to ensure they update. There was a simple
bug where the clip_changed bit was not initialized from the parent
context in the PrePaintTreeWalkContext constructor. This was only
observable in an edge-case when
PrePaintTreeWalkContext::tree_builder_context was not present to handle
the propagation. This bug was masked in builds with DCHECKs enabled
because PrePaintTreeWalkContext::tree_builder_context is always created.

Bug: 936720, 933555
Change-Id: I0599e9f530f4db2c1a3566443cfece3ef319f2a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1526301
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641254}