Fixing references to CopyOnWriteBuffer which might copy data.
CopyOnWriteBuffer has two versions of data(). One is const the other is
not. The const version returns the data as read-only. The non-const
version returns the data as writable and could potentially make a copy
of the data before returning it (hence, copy-on-write buffer).
This code wants a const pointer to the data, but is using a non-const
buffer, so the non-const method is actually called.
The change is to use the cdata() method which is the same as the data()
const method.

Bug: None
Change-Id: Ie8c25a53d4a9691853cb31752a79059249067e16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1525163
Reviewed-by: Steve Anton <steveanton@chromium.org>
Commit-Queue: Amit Hilbuch <amithi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641375}