ozone/wayland: Fix Drag&Drop crash when dragging within chromium
With dcheck enabled, dragging any data between chromium
windows (e.g: drag a page link into omnibar) makes it to
crash just after drop. This CL fixes it.

That happens because when the drag session starts from
a chromium window, the (singleton) WaylandDataDevice's
source_data_ already holds the OSExchangeData instance
(which contains the data to be transferred), so in this
case an optimization was partially implemented to avoid
sending data through wayland pipe in this special case.
So, to fix the crash this change add the missing bits to
this alternative code path, so that the respective dcheck
does not fails anymore at aura level.

Additionally, this CL does some readability improvements
and cleanups in the DataDevice class implementation and
interface.

Bug: 875164
Change-Id: I34f53bdf5415997b15f27ba874da9968efba1e9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1524884
Commit-Queue: Nick Yamane <nickdiego@igalia.com>
Reviewed-by: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#641384}