Immediately complete synthetic gestures in VR
In https://crrev.com/b37e23cbd52ba447ab89c6d18f4ea0ecdc3fa5c0 I landed
a change to force synthetic gestures to cause a full redraw and wait for
the resulting CompositorFrame to be presented before invoking the
completion callback.

This patch broke VR telemetry. We never receive the presentation
notification so synthetic gestures never complete and cause the tests to
hang and timeout.

This CL causes synthetic gestures to complete without causing the redraw
and waiting for frame presentation if we're in VR mode. This was the
behavior prior to my change and gets the waterfall back to a green state.
I'll investigate a correct long term fix in a followup.

Bug: 940063
Change-Id: I6f6b9e9a5b9e369a1ec010cba93a1d952d3d0a18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1525150
Reviewed-by: Stefan Zager <szager@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641677}