[Chromeshine] Hook up in-memory components to persistence
This CL makes a large number changes to account for the
fact that UsageStatsService operations are now async. The major upshot
is that promises are used instead of actual values. If data has been loaded,
then promises are fulfilled immediately. Otherwise, they're fulfilled once
the database is loaded. There is a corresponding downstream change that
adjusts UsageStatsBrowserService to account for this.

There are also a couple bug fixes; usage stats database needs to split its
databases into different folders, so that is fixed.

Bug: 939456
Change-Id: I0869318b5733ea6cb496af31224c6e9c93030fb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1509103
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641699}