Do not coalesce wheel with phaseBegan to wheel with phaseEnded.
If no element in the scrolling chain is not scrollable in the given direction
(delta_hints) on GSB we latch to the viewport. This is not user friendly on
platforms that do not have any scroll affordance animations (e.g. Windows)
since if the user changes the scrolling direction and an element in the chain
is actually scrollable in the new direction, the user expects it to scroll and
has no idea that we have latched to the (possibly unscrollable) viewport for
the rest of the scroll sequence. To address the issue for timer based wheel
scroll sequence we end the sequence and start a new one upon arrival of a wheel
event with different delta directions if no GSU has been consumed since the
beginning of the sequence. This cl ensures that the GSE/GSB generated for
breaking the latching seqeunce are not coalesced with each other.

Bug: 882998
Change-Id: Icf703c935475fffd69b899c59abc03ec1b548784
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1524703
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Commit-Queue: Sahel Sharify <sahel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641719}