mojo: Fix cfi-icall failure
Control Flow Integrity indirect call checking [1] verifies that indirect
calls are only made to functions defined with a type signature that
matches the function pointer type. CFI-icall as it is used in Chrome
only works for functions that address-taken in the same DSO as it is
checked. However, the function pointer in InvokeWatchCallback can be
from another DSO, requiring CFI-icall to be disabled for this function
for now.

[1] https://www.chromium.org/developers/testing/control-flow-integrity

Bug: 941469
Change-Id: I19add8104b6ed3c1691e256c6ebbcdb6a6bba272
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1524713
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Ken Rockot <rockot@google.com>
Auto-Submit: Vlad Tsyrklevich <vtsyrklevich@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641722}