Plumb transparent values for MUS.
This is a follow-up fix to https://crrev.com/c/1504599. The occlusion
tracker also uses the transparent flag on each window to decide if it
can affect occlusion. Plumb this value through, and set the layer draw
type property for each remote layer. This will let the occlusion tracker
function the same for remote windows as it does for local.

Bug: 937323
Test: Bug 908629 does not regress.
Test: Android window can be occluded by Browser window in SPM.
Test: Added unit test
Change-Id: Ib7875355346563b2c83447bf8b685d58130ab74a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1506061
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641800}