Force numlock to be on when sending modifier state to sommelier
linux apps decide how to interpret numpad keys based on the modifier
flags. CrOS is designed so that numlock is a vestigial key (c.f.
crbug.com/212491#c19 ), but the implementation has that bit in the mask
as always off, despite that the behaviour is always on.

This fix hard-codes that the modifier sent to linux will always say
numlock is on.

Bug: 937102
Change-Id: Iffe91b180ad22de0d303ec404df30efc73ffc93b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1525241
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641825}