chromeos: gets mirroring in overview mode working with hidden windows
Specifically for the single-process-mash case.

In overview mode it's possible for many of the windows to be hidden. In this
mode we still want to see the content, so kMirrorWindowList was added.
kMirrorWindowList basically forces WebContentsViewAura to treat the content
as visible, even though the window isn't actually visible. This won't work
in mash because WebContentsViewAura and ash will eventually be in separate
processes.

This patch provides an alternative to kMirrorWindowList that will eventually
replace kMirrorWindowList. In particular, two things are necessary to get
the same functionality:

1. The remote client has to think the window is visible, otherwise the
compositor in the remote client won't generate any frames. To enable this
the window-service gets functionality such that ash can force the client to
think the window is visible regardless of local state. I experimented with
making the windows in ash visible, but this has a much worse set of problems.

2. Occlusion tracker gets the ability to have a window forced visible. This
way the remote client thinks its windows are not occluded, which is important
for the WebContentViewAura.

I am not removing kMirrorWindowList in this patch as I hope to merge this.

BUG=936674
TEST=covered by tests

Change-Id: I7d853dd31fd5f23292b45b09164e75aea196ee24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1512664
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#641879}