Move tab_renderer_data into c/b/ui/BUILD.gn
Because (for historical reasons) the file was in the very small
c/b/ui/views DLL, it could not access the default constructors or
destructors of any types in c/b/ui/views that were not in that DLL. This
would have made it impossible to, for example, share thumbnail image
data between a tab helper and the preview display in a flipbook or hover
card (adding a non-trivial member to the struct of a data type defined
in c/b/ui/BUILD.gn would cause linker errors).

Since the only files that depend on tab_renderer_data are in
c/b/ui/BUILD.gn, it should be moved to that package along with many of
the other existing files in c/b/ui/views/tabs which are already there.

Bug: 928954
Change-Id: If32406bb3363ffbb96fbea60e50a8452217f6a29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1529294
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642059}