Make SimpleKeyedServiceFactory thread safe.
crbug.com/701326 is a problem (being worked on) in BrowserContext
factories, but Simple factories are not restricted by the preexisting
usages in different threads. The DCHECKs can, therefore, just be
introduced.

There is already an override of GetKeyToUse() which does not
perform the check, this CL fixes that and future problems.

Change-Id: I160ada8a1edaccd0686794c0d018e768c54d4259
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1529576
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642098}