[DL]: Fix display lock composited layer paint bug.
Sometimes, we don't go through the PaintLayerPainter::Paint() function
and instead call ::PaintContents directly, such as the case in a
composited layer. In this situation, we weren't preventing paint
from happening (and effectively didn't clear the painted output) of
a display locked element.

This patch fixes this and moves the "stop paint" call to PaintLayerContents
instead.

R=chrishtr@chromium.org

Bug: 882663
Change-Id: Ia6087115480f34e5a78879408819eff0a6b5f3a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1531007
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642146}