[Extensions] remove iron-collapse for the stream
This CL is a preparatory CL for iron-list and removes the
expand/collapse animation from the stream item. To make the stream item
a more lightweight component while still maintaining functionality,
iron-collapse is replaced with a div with a hidden property and contents
of #activity-item-main-row are now in the slot for cr-expand-button.

Bug: 932768
Change-Id: Iad7bdcc0ce44d1f3b40600c2d953586da7e00d9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1525621
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Kelvin Jiang <kelvinjiang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642298}