Remove input field selection regression in Gallery
Change https://chromium-review.googlesource.com/c/chromium/src/+/1217747/
introduced a side-effect that causes the entire input field to be
selected on a focus event. In the gallery app this caused the Width
input field to be fully selected when doing an image resize. This
change explicitly sets the selection to the end of the field as it
used to work before the regression.

This CL is the second attempt as the original CL 1453696 was
reverted in
https://chromium-review.googlesource.com/c/chromium/src/+/1519947
due to https://bugs.chromium.org/p/chromium/issues/detail?id=931157.

Bug: 914741
Tests: Ran all browser and unit tests, manually checked as well.
Change-Id: I95bd7cc31ccbd0631201414142cfe9b4cf44073f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1525253
Reviewed-by: Trent Apted <tapted@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642304}