Add CryptAuthEnrollmentManager implementation for v2 Enrollment
Creates an implementation of CryptAuthEnrollmentManager to be used for
CryptAuth v2 Enrollment. This implementation,
CryptAuthV2EnrollmentManagerImpl, will replace the v1 implementation,
CryptAuthEnrollmentManagerImpl. While the API for the abstract class is
not ideally suited for CryptAuth v2, using it will allow for a smoother
transition from v1 to v2.

Bug: 899080
Change-Id: I3ba278cf7987810687fa199db61a258c36f70e69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1440201
Commit-Queue: Josh Nohle <nohle@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642305}