Histograms README - add note that histograms for single events are okay
This came up from a discussion at the metrics meetup last
December.

Frankly, I'm not sure this is the right place in the doc to put
this section, nor am I positive this is the right phrasing.  Or perhaps
it's too long and I don't need the example.

Change-Id: I12a42c713cda93a04616d6509475d8f34c4f26af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1531649
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Mark Pearson <mpearson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642322}