oopd: Land OverlayCandidateValidator for CrOS.
Add code necessary to instantiate a working OverlayCandidateValidator on
Chrome OS with OOP-D enabled. This includes code to create the validator
and DrmOverlayManagerGpu implementation that communicates with the DRM
thread by PostTask.

This CL only enables overlays if the --enable-hardware-overlays flag is
provided. A bit more plumbing is needed to turn on overlays by default
for devices that support it.

Bug: 930173
Change-Id: I4196f40525b3e8caccb96823de2dee3e1dbeceaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1504043
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642496}