Add password support in EditableCombobox.
The read-only Combobox does not support a password mode, leaving it to
the users to implement one in the model. This approach doesn't easily
work for EditableCombobox because it relies on Textfield, for which we
need to explicitly enable a password mode.

This CL therefore introduces EditableCombobox types, of which there are
only two for now. For the password type, we can hide or reveal
characters, which affects the items in the menu and sets the appropriate
mode in the Textfield.

This CL also modifies UpdateItemsShown to directly read the current text
from the model's owner (the EditableCombobox), and therefore does not
need to receive the latest content as argument.

Bug: 923660
Change-Id: If6ee2a3a419bd8e51b9853144876831688a07d36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1529359
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Edin Kadric <edinkadric@google.com>
Cr-Commit-Position: refs/heads/master@{#642497}