viz: Don't extend child's deadline to match parent
The problem with extending deadlines is that it allows lower-priority
clients down the hierarchy block higher-priority clients up the
hierarchy. For example:
1) Browser embeds the main frame with the deadline of 10.
2) Main frame embeds an OOPIF ad with the deadline of 2.
3) Main frame responds within a reasonable amount of time, but the
OOPIF is very slow and responds after 15 frames.

With deadline inheritance, the browser and main frame are blocked for
10 frames before something is shown, even though the only client that
hasn't responded yet is just a low-priority client (an ad). Without
deadline inheritance, the browser and renderer will activate after 5
frames and can pump 5 more frames before the renderer in the
alternative scenario shows something.

This CL also takes advantage of the fact that deadlines can't be
extended in order to simplify deadline inheritance. Rather than
actually propagating the deadlines down the surface hierarchy,
just activate the last surface in embedded SurfaceRanges when
the parent activates (note that the child's activation will trigger
grandchildren's activation and so on). The new implementation lives
in SurfaceAllocationGroup because it allows us to do this in a
simpler and more efficient way. Figuring out whether a surface is
"late", which means it inherits a deadline of zero from its embedder
that has already activated by deadline, is also implemented using
SurfaceAllocationGroup.

Bug: 931801, 942093
Change-Id: I0385ac79e742d0673168c0c64df30350de5140e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1518718
Commit-Queue: Saman Sami <samans@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642561}