Wire CryptAuth v2 Enrollment into DeviceSync service
First, the ClientAppMetadataProvider is now built in
ProfileImpl::HandleServiceRequest() and passed to the DeviceSync
service, then subsequently passed to the DeviceSync service
implementation.

Second, in the DeviceSync service implementation--specifically, in
DeviceSyncImpl::InitializeCryptAuthManagementObjects()--the enrollment
manager is chosen based on the feature flag |kCryptAuthV2Enrollment|.
If this flag is true, the v2 enrollment manager is created along with
the key registry; otherwise, the v1 enrollment manager is created.

The DeviceSync implementation unit tests now run all tests separately
for both v1 and v2 enrollment infrastructures.

Bug: 899080
Change-Id: I4a590b3bea3c22b8c72a169bb48288c31b0c1e32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1493394
Commit-Queue: Josh Nohle <nohle@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642628}