Implement |AutomationAXTreeWrapper::IsInFocusChain|
This fixes a problem where we sometimes would incorrectly report a node as having the focused state. This was because according to its local tree data, it is focused, but it isn't focused according to its ancestor tree wrappers (who's child trees don't point to that tree).

Bug: 933618
Change-Id: I534cb4137cd198b90254e10548fc9431d1825782
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1531522
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642665}