Focus sub-menu on parent menu keyboard navigation.
This implements the case when a sub-menu is opened using
the keyboard, causing focus to be moved to the first item
in the newly opened sub-menu.

This CL also hides the sub-menu if the top level menu is being
navigated with the arrow keys.

       navigation of the main menu. Changed existing unit tests to
       handle the changed focus from main menu to sub-menu.

Bug: 941968
Tests: Added unit test to confirm sub-menu gets closed on keyboard
Change-Id: I020db000e6bd547433aa44891519159ad5c20619
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1529498
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642709}