Perfetto: Ensure the MetadataSource is registered before the TraceEventDataSource
Needed because until the Perfetto Observer interface is ready, the logic used
to signal tests etc that tracing is "ready" relies on the TraceLog getting
enabled, which is what the TraceEventDataSource does. To ensure that all of
the other data sources are also ready at that point, we make sure the
data sources are registered in the reverse order of them being added
to the ProducerClient.

R=ssid@chromium.org

Bug: 839084
Change-Id: I4728afa1bf4fda72ff83778211138f2d45ded85f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1532987
Reviewed-by: ssid <ssid@chromium.org>
Commit-Queue: oysteine <oysteine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642747}