[BrowserSwitcher] Name the alternative browser chrome://browser-switch
On chrome://browser-switch, instead of saying "an alternative browser",
it will now say e.g. "Opening in Internet Explorer" if we can detect the
browser easily.

If AlternativeBrowserPath is left unset, or set to a preset value
(${ie}, ${firefox}, etc.), the new string is used. If it is configured
with an executable path, we fall back to the old string.

Bug: 939438
Change-Id: I7c5491f83d09779e78fbfb1c91e09ecc6f40c5b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1511709
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642776}