MacPWAs: Poll for Chrome's AppShimHostManager to be ready
Background:

Connecting to Chrome will fail until the function
AppShimHostManager::InitOnBackgroundThread in the browser process
completes.

For this purpose, there current exists a ReplyEventHandler class
instantiated in the app shim process that will signal the
AppShimController when the Chrome process' event loop has started.

This doesn't actually do what we want, because the function
AppShimHostManager::InitOnBackgroundThread may or may not have actually
run when the event loop starts.

Fix:

Delete the ReplyEventHandler class.

Change AppShimController to take as an argument the
NSRunningApplication of the Chrome process to which we will want to
connect.

Make AppShimController poll (at 100 msec intervals), checking to see
- if the NSRunningApplication is still running
- if the socket path to which we will connect exists

Once the socket path exists, go ahead and try to initialize the
mojo connection to Chrome.

Bug: 943807
Change-Id: I0dac4d5a9af95396231b6472aeea8972fb7d0a06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1532906
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642833}