Decouple browser/page_load_metrics state from mojom::PageRenderData.
To support subframe aggregation, we will modify the IPC to send a delta
instead of the cumulative score.  In preparation for this, introduce a
new structure to hold the total score for use in PageLoadExtraInfo,
AMPPageLoadMetricsObserver, and PageLoadMetricsUpdateDispatcher.

A follow-up will rename mojom::PageRenderData, to avoid confusion.

Bug: 581518
Change-Id: Iba25e4eddfdb1225d9da6f6b80d4dd9a2919e931
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1531290
Reviewed-by: Bryan McQuade <bmcquade@chromium.org>
Commit-Queue: Steve Kobes <skobes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642906}