tablet: Carry over snapped windows from desktop to tablet for all users.
This CL addresses the workflow where you enter tablet mode and then
switch to a user account that had an active window snapped in desktop
mode. This window should be carried over to split view, consistent with
the case where it belongs to the active user session when tablet mode
starts. The current behavior seems broken because split view mode is not
active but will become active if you then switch to desktop mode and
back to tablet mode.

All changes to the split view controller are just reverting changes from
CLs 1426922 and 1481799. The functionality from those CLs is moved to
the tablet mode window manager and extended to handle the workflow in
question. To facilitate that, code for carrying over snapped windows to
split view is factored out and modified to work for windows that are
already tracked (as well as for windows that have yet to be tracked).

Test: manual
Bug: 936610
Change-Id: I9c0153ba2bcba1473b5c684320a877299ea7eb1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1497477
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#642994}