Adding cpu metrics to AdsPageLoadMetricsObserver.
This CL adds cpu data to the AdsPageLoadMetricsObserver.  The observed
cpu metrics are load (percentage of active time over total time) on the
cpu for the page as a whole and for ad frames.  For activated frames,
this is split into the pre-activation and post-activation load.  For
unactivated frames, this is split into the pre-interactive and
post-interactive load.  All of these metrics are relative to the
overall time for the page as a whole, not for individual ad creatives
(for example, a creative made halfway through the page's lifetime
considers its time relative to the lifetime of the page).


Bug: 937298
Change-Id: I5923a624d4c18a219505e05a39a6efd8146e41d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1512494
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643013}