Reland "[DL]: Fix yielding budget to ensure it marks the next phase.""
This relands the yielding budget update to ensure that we process the
next phase after completing a phase.

This includes a fix to initialize |next_phase_from_start_of_lifecycle_|
variable in the yielding budget, since we can access it prior to normal
initialization in the following scenario:

e.dl.update();
e.appendChild(div);

The append would check if we can style it while inserting it, which
would check the budget before the lifecycle update phase begins (which
is what normally initializes the variable).

R=chrishtr@chromium.org

Bug: 882663
Change-Id: I6f8ad99ec0455769b3f16e630abcfd7a3dd9a3d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1534136
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643021}