Fix AXPlatformNodeWin::Navigate DCHECK after iframe load
The DCHECK was verifying that, on a navigate to an element's first
child, that child's parent is the node being navigated from. The check
was failing when navigating from a content fragment root to its first
child, which turned out to be the root node of an iframe rather than
the root node of the root frame.

The issue was that document load in the iframe was incorrectly
triggering a switch-out of the node tracked by the fragment root as its
sole child. The fix is to only perform that switch when the root node
of the root frame changes.

Bug: 928811
Change-Id: Ic4fe771f1dcf536426d4658ec778dbc9a939acb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1532893
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643050}