AccessibilityNotificationWaiter should listen to all frames
This tries to simplify AccessibilityNotificationWaiter in order to make
it easier to make some future changes that will reduce flakiness.

The main change is that now you always pass a WebContents and you always
listen to all frames.

Previously you could construct an AccessibilityNotificationWaiter with either
a WebContents or a RenderFrameHost, but either way you had to do extra work to
listen to all frames. It had five constructors and it still often required
extra loops to set it up.

Instead, all of the code to listen to all frames is now part of
AccessibilityNotificationWaiter and it's always run. It makes the
class easier to use, and it's harmless to listen for changes on
additional frames even for tests that don't strictly need it.

Bug: 944537
Change-Id: Ic72abe809348b3349cb461e673de4453f56a74a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1534677
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643103}