Implement a GetAllErrors method for InstallableManager.
This CL adds a way for a caller to initiate a "dry-run" installability
check that accumulates all errors, and returns a list of strings via a
callback. Tests are added to ensure the correct functionality.

Some of the error cases trigger on navigation away from the page. These
will not be returned to a GetAllErrors call and have been marked as
UMA-only.

BUG=915945

Change-Id: I732b81185c1c632935e73ac3cdae49e6960ecdb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1514011
Commit-Queue: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643205}