[omnibox] Make default fields hard to not initialize
There was a field in AutocompleteMatch's that was not getting
initialized, a bool |from_keyword|. This CL initializes it, and also
moves a lot of the initialization of other fields into the class. Thus,
if someone adds a new constructor, the fields will at least get a
default value.

Bug: 934316
Change-Id: Ib9f04def10a4e5afb7493ac01b98febc05f182af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1532243
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Kevin Bailey <krb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643218}