Support windows virtual desktops in occlusion calculation.
If a browser window is not on the current virtual desktop, it is marked occluded.
If a window is not on the current virtual desktop, we ignore it when deciding if
a browser window is occluded or not.

This unfortunately spams the VS output window - newer versions of Windows have fixed this.

Bug: 943534
Change-Id: Iee9654f494fb2f977688c93347f4171447fb4b58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1531460
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643347}