Support vulkan rendering without vulkan surface on X11
When chrome is running with remote desktop environment,
VK_KHR_xlib_surface is not supported. To support this desktop
environment, a flag --disable-vulkan-surface is added. With this flag,
chrome will initialize Vulkan without using swap chain, and render
to offscreen vk image and then copy the content of the vk image to
xwindow.

Note: --use-gl=swiftshader is needed to make it work.

Bug: 943583
Change-Id: I7c3d89108bef84dc0225201fdb0688356bc125d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1532177
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Peng Huang <penghuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643437}