Migrate AMP subframe metrics to use AMP loading behavior.
Historically, we analyzed URLs to determine if a page was an AMP
page. For example, we'd look for www.gooogle.com/amp/. This is a
suboptimal approach, as it makes assumptions about the content
being served based on the URL.

Now that we expose an AMP loading behavior, we use that as a more
general signal.

This is part one of a multi part change to deprecate URL analysis
for AMP metrics in UMA and UKM.

Change-Id: I2bfd254d696c77ffe6107826d0a4dcd42d77303c
Bug: 939333
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1512304
Commit-Queue: Bryan McQuade <bmcquade@chromium.org>
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643446}