Adding Offline Previews UKM
Offline Previews are a replacement for slow page loads. In particular,
when Chrome is on a slow network (2G-like speeds) and a copy of the page
has already been stored in persistent memory, the Previews code may
serve the Offline Page instead of fetching the page from the origin.
This is a direct replacement of a standard page load. In that way, these
loads are like any Preview; they represent an efficiency introduced by
Chrome to load origin pages more quickly. By not reporting UKM, we
cannot fully analyze the effect of offline previews, and we cannot add
it to the opt out pipeline used for the rest of Previews.

Bug: 922277
Change-Id: Ib8cdb2f40de2ae5ae1127a2e4b44c66c17765804
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1516016
Reviewed-by: Bryan McQuade <bmcquade@chromium.org>
Commit-Queue: Ryan Sturm <ryansturm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643503}