Make InstallableData GURL members references.
The members of InstallableData are explicitly marked as being owned by
InstallableManager. This ensures consistency of data and reduces
copying. This CL makes the GURL members into const references to further
reduce copying, and updates tests that assumed the data was not a
reference.

Comments on members of InstallableData are also clarified.

BUG=915945

Change-Id: I4d8ee872998686ced3d8bf8a38e005e7e8ae2065
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1536039
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Commit-Queue: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643765}