[Onion Soup] Move MediaStreamVideoCapturerSource to Blink
The header for this class is temporarily in public since
it is still referenced in content. Final destination will be
blink/renderer/modules.

The move was not completely straightforward due to restrictions
placed on headers in public and usage of render frame IDs.

This CL contains blink style violations with regards to the use
of base::Bind variations and helpers. They will be converted
to WTF in follow-up CLs.

Bug: 921006
Change-Id: I75fa61a73acaebaddeed6df7018e7ba80b0f87e4
TBR: haraken@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1536176
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643792}