[iOS] Speculative fix for accessibility crash in page control.
This is a speculative fix for a crash when acessing the
accessibilityElements property in the tab grid page control. The theory
behind this fix is that the page control implemented both the
acessibilityElements property *and* the UIAccessibilityContainer methods
that can be used instead of the accessibilityElements array.

This CL removes the (presumably superfluous) UIAccessibilityContainer
methods that don't seem to be needed.

to verify accessibility attributes were unchanged.

TEST: Manually tested page control, inspected using Accessibility Inspector
Bug: 932585
Change-Id: Ica0474c7fb5a743d26d2c1a94039d5eab8517d1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1517894
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643809}