Support windows virtual desktops in occlusion calculation (reland).﻿
This is a reland of crrev.com/c/1531460.

This is original patch. Patchset 2 will
have the fix for the win7 issue, which is to  say that only windows
that are definitely on another virtual can't occlude windows on
current virtual desktop. (!= false instead of == true).

Support windows virtual desktops in occlusion calculation.

If a browser window is not on the current virtual desktop, it is marked occluded.
If a window is not on the current virtual desktop, we ignore it when deciding if
a browser window is occluded or not.

This unfortunately spams the VS output window - newer versions of Windows have fixed this.

Bug: 943534
Change-Id: Idd0949ade3e51371c482df250292766ec84cd33c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1536441
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643834}