Change clipboard timeout to use separate Feature
Before, the clipboard timeout feature was in the OmniboxBundleExperiment.
To make it easier to change the timeout, this CL adds a separate feature
to control the timeout that can be set anyhwere. There are currently no
experiments going on in OmniboxBundledExperiment to change the timeout,
so this will not break any existing experiment.

Bug: 943164
Change-Id: If75c8caef62850cc081a6f82d35e4e68e25f77e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1535878
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643964}