[SendTabToSelf] Make Java code use the C++ "IsFeatureAvailable" call.
Previously, the logic was re-implemented in Java. With the CL, the
Java code calls into the C++ version of this function.

This CLs moves the SendTabToSelfShareActivityTest.java from
javatests to junit. That way, we can mock out the JNI layer.

It also splits the send_tab_to_self_util in 2 (same name + desktop
version) so that the former can be included in android builds.

Bug: 941074
Change-Id: Idbc4305d3807669ed42e17740c6b64a168e2ed94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1518463
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Jeffrey Cohen <jeffreycohen@chromium.org>
Commit-Queue: Jeffrey Cohen <jeffreycohen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#643967}