Refactor per-resource metrics in AdsPLMO to a separate observer
Refactor all of the Ads.ResourceUsage.* metrics to a separate
PageLoadMetricsObserver. Currently AdsPageLoadMetricsObserver is
recording all metrics that have anything to do with ads which complicates
the class and also makes these metrics prone to bugs.

The class is being renamed to resource_metrics_observer because these
histograms do not exclusively record metrics for ad resources.

This change is purely a refactor, it does not affect histograms or
output behavior, or change testing.

Bug: 936040
Change-Id: I3d340c151cce9366263f14c9a5f476b6721035c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1500074
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644009}