Fix the issue that the AppListView of PEEKING state shows in fullscreen
When AppListView with the virtual keyboard shown is dismissed, the
virtual keyboard is closed after the AppListView does. If the
browser is behind the AppListView, IME is updated due to the changed
focus. So the virtual keyboard is hidden for the wrong IME instance.
As a result, InputMethodBase::keyboard_bounds_ of the AppListView is
not reset although the virtual keyboard window is closed.

In this CL, during the process of closing the AppListView, check the
state of the virtual keyboard and close it if it is active.

Test: ash_unittests
Bug: 944233
Change-Id: I54639aa05eb6a81bc288d9a9c199c36571933a87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1532990
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644018}