[CrOS Wallpaper] Remove ripple effect on disabled Daily Refresh toggle.
Currently, when the user has the Wallpaper app open, and takes their
chromebook offline, the Daily refresh toggle is disabled but still
produces a ripple effect when the user clicks on the toggle.  This
CL adds the constraint that the ripple effect will only occur when
the Wallpaper app is online.

Bug: 889847
Change-Id: I7ccbe35a82787ff4515b8e1878a4408bde9f9816
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1539019
Reviewed-by: May Lippert <maybelle@chromium.org>
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644148}