[On-device adaptive brightness] Move daily metrics logging.
This cl moves daily metrics logging to OnUserBrightnessChangeRequested so that
we treat a sequence of rapid brightness adjustments as 1 user adjustment.
OnUserBrightnessChangeRequested is only called after user brightness adjustment
is completed.

If user adjustment effect is DisableAuto, then we disable the adapter when
OnUserBrightnessChangeRequested is called. OnUserBrightnessChangeRequested is
called as soon as user makes a brightness change, hence we can disable the
adapter immediately.

Bug: 881215
Change-Id: I86b47d3acd2ffd9711f7e00568ac9d796e4474f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1539316
Commit-Queue: Jia Meng <jiameng@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644222}