🦁 Fix organization of ios omnibox popup files and build files
Many of the protocols and files used in the omnibox popup were actually
stored in the omnibox directory. This lead to dependency cycles,
especially when trying to keep Showcase code from depending on too
much on components code.

To fix this, I moved
- autocomplete_match_formatter
- autocomplete_suggestion
- truncating_attributed_label
- image_retriever

from omnibox/ to omnibox/popup/. All of these were only used in popup/,
so it was easy to fix the dependencies.

This also allowed the creation of a new popup_ui target that can be
used in Showcase without bringing in too many dependencies.

Bug: None
Change-Id: I4aa69690f9d322d5105a26420e8992882204a3bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1532103
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644258}