Fix the issue that AppListView is moved downward in tablet mode
AppListView's bounds are stored as the property called
kVirtualKeyboardRestoreBoundsKey when the search box is activated in
clamshell mode. When the tablet mode is turned on, this property
is still kept in AppListView's native window. As a result, when closing
the virtual keyboard in tablet mode, the native window's bounds are
restored. It explains why we can see that AppListView is moved downward
when dismissing the virtual keyboard.

In tablet mode, AppListView should not be moved because of the virtual
keyboard visibility. So in this CL, clear that property when tablet
mode is turned on.

Test: ash_unittests
Bug: 944133
Change-Id: I714ebd50448a995f77bae759f6282efc8adc6df1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1538983
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644445}