Print Preview: Add a flag for phase 2 layout changes and resize
Print Preview's phase 2 layout changes will be added behind a flag. Add
this flag. If the flag is enabled, set the maximum dialog size on all
platforms to 1000x660, instead of only doing this on Mac.

No behavior changes expected if the flag is not enabled, or if using
Mac OS where the dialog size is already limited.

Bug: 945619
Change-Id: Ifcfb716842dbf9539032760d97a456fee2f4c60a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1538687
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644545}