Create WorkAreaInsets class
This new component will encapsulate information needed to calculate
work area for root window.
It is in 1:1 relationship with RootWindowController and owned by it.

Caching height on accessibility panel is restored in this change for
consistency with docked magnifier and for performance reasons (avoid
frequent calculation of this value).

Test: manually with 2 displays +  existing unit tests
Bug: 944685
Change-Id: Ib76f144a54408a769ec101e15f6a2f2512aa4ae5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1538689
Commit-Queue: Aga Wronska <agawronska@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644608}