Introduce ConflictsDataFetcher
As part of moving the ModuleDatabase to its own sequence, the
chrome://conflicts page still needs to access its content from the UI
thread.

This is done by the ConflictsDataFetcher, which lives on the
ModuleDatabase task runner and returns the resulting data to the
conflicts handler.

Bug: 943303
Change-Id: I73c1413afaa1555abc2ba3c3bc891851b75ad71a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1531090
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Sébastien Marchand <sebmarchand@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644669}