[signin] In Mice mode, do not allow unknown accounts in the Gaia cookie
With Mirror and Dice, unknown accounts in the Gaia cookie are not
actively removed, to prevent unnecessary calls to /LogOut.
In Mice mode, we want the accounts in the cookie to exactly match the
accounts in the system, and thus only accounts known to Android are
allowed in the cookie.

Bug: 944826
Change-Id: I6c80c6207679b7828a5716571e6b7bbdde12a1bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1538405
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644710}