[wasm] Handle FetchDataLoader::Client::Abort for streaming compilation
A wpt test showed that for WebAssembly streaming compilation, we did not
handle the abort of a {fetch} correctly. With this CL we throw a
DOMException in that case.

I am not an expert of the FetchDataLoader API, and of the blink GC.
Please take a closer look at these aspects.

R=ricea@chromium.org

Bug: chromium:943487
Change-Id: I3038415265928947233664653a6e64316d240405
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1535877
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644742}