[ios] Add a KeepRenderProcessAlive flag to prevent WebProcess suspension.
The KeepRenderProcessAlive webState flag allows the ios/web webView container
to track it's |window| status, and instructs the WebClient to attach the webView
back to a window if it's hidden. For ios/chrome, this is simply the keyWindow.

This prevents the WebProcess and NetworkProcess from suspending for active views.

ios/chrome set KeepRenderProcessAlive to true for the active webState with the BVC
(or sub-presented view controllers)), or for one-off webState's like DOMDistiller
and Preload.

See go/prevent-wkwebview-suspend for details.

Bug: 930123, 943569
TBR: noyau
Change-Id: I0e8e5cb0f13e21091a3fefcc4e79685d86fe352c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1529921
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644754}