[Tango->FCM] Make lazyness work for the new invalidations
Currently messages are not persisted for the FCM invalidations, because
the sender id for which invaliations are marked lazy mismatches the id
received in the incoming messages. In addition, messages are replayed
before app handler is knows, so they aren't passed to it.

This patch implements the fix.

Bug: 882887, 944484
Change-Id: I6f53d65488f196f76dedc0b8ae23f9889d5c13a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1533875
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Tatiana Gornak <melandory@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644773}