Chrome OS: Move AuthPolicyLoginHelper
The files were moved from chromeos/login/auth to
chrome/browser/chromeos/authpolicy.
The class was renamed to AuthPolicyHelper.

Helper functions are not login-related, so it doesn't make sense to
keep the file in login/auth.

This CL is a first step to add helper functions for
ActiveDirectoryPolicyManager.

TBR=xiyuan@chromium.org

Bug: 843177
Change-Id: I7e440e49dfc0d170a53a8cb424c51e78d2f634a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1541109
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Commit-Queue: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644837}