Keep RemoteFrameView::subtree_throttled_ up to date.
Previously, subtree_throttled_ was only set once, when an iframe was
attached to the layout tree, and then never updated. I haven't seen any
bugs resulting from this, but it's clearly wrong, and fixing it appears
to claw back some of the performance regression in the bug.

BUG=941116

Change-Id: I5a094a7c35d0c0a673de358f4cb095bb8f6de032
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1531680
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644875}