Move PageScaleFactor setters from WebLayerTreeView to WebWidgetClient (6/n)
The SetPageScaleFactorAndLimits() and StartPageScaleAnimation() methods
both move to WebWidgetClient, though are only called for the main frame
widget, when compositing is being used.

The HasPendingPageScaleAnimation() can move directly to RenderWidget
since it is only used inside of content.

The WebLayerTreeView is a shortcut past the WebWidgetClient that is
not needed now that we always composite. This works toward eliminating
this extra API layer.

More importantly this helps us Close/Restart the WebWidget parts of
WebViewImpl by not requiring plumbing 2 pointers with slightly different
lifetimes (WebLayerTreeView is created inside the Init of the
WebWidgetClient). This will help avoid a bunch of complexity in
creating WebFrameWidgets and other WebWidgets.

Also add a bunch of DCHECKs around expectations that page scale is only
a feature of compositing, and the code paths and VisualViewport are not
used when compositing is disabled. Secondly, the same holds when the
main frame is remote.

R=dcheng@chromium.org, wjmaclean@chromium.org

Bug: 912193
Change-Id: Ie8312a15ca515917fc80c75dc6371426dd3d4da4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1530825
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644946}