Add metrics provider for uploading background traces
Background tracing uploads metrics of larger size and it is better to
upload them as independent metrics rather than part of uma log, to track
upload size, and also to make sure the uma metrics are not discarded
from saving to disk due to large size. The background tracing manager
will make sure traces are only uploaded on WiFi, or the traces are
small when uploading over data, to make sure weekly upload quata for
UMA metrics is not affected on Android.
The CL also changes the MetricProvider API to pass in uma proto instead
of system profile for independent metrics collection. Background tracing
metrics provider will look for traces from background tracing manager,
when the metrics service asks for available metrics. If available the
uma proto will be filled with the trace buffer.

BUG=925151

Change-Id: I69f2400b2bdf73067e820c8031960390d4b8cbdb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1481737
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: oysteine <oysteine@chromium.org>
Commit-Queue: ssid <ssid@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644952}