[Media Notification] Derive background color from artwork
The media notification should derive its background
color from prominent colors in the artwork. This
uses as similar algorithm from Android.

Also breaks out some of the logic of CalculateProminentColors
so we can use it without the reset of the function. This
is because we want uninteresting colors and do not want
the color profile matching.

BUG=944598

Change-Id: I60df7f600f6fc881dfb38d2057c903a5679845df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1534923
Commit-Queue: Becca Hughes <beccahughes@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#644985}