ScriptProcessorNodes are not active source nodes
Previously, a ScriptProcessorHandler was placed on the
active_source_handlers_ list, as if it were like an
AudioScheduledSourceHandler.  (This is kind of true since
ScriptProcessorNodes can act as sources.)

However, once placed there, they are never removed until the context
goes away because there is no way to note that node has stopped
processing, which is unlike AudioScheduledSourceNodes which have a
stop method.

So don't add ScriptProcessorHandler on the active source list.  The
lifetime of the handler (and node) is managed by HasPendingActivity
which is true if there is an audioprocess event listener.

Bug: 892917
Change-Id: Ied8d3bf84249d3ca61f0736ef40e641234dd6b69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1540289
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645013}