[Cronet] Common UrlRequest should report metrics before callbacks.
We already do this with success and cancellation, we should do the same
for the error callback.

For Cronet native metrics, this prevents the following use-after-free
race that occurs if metrics reporting is enabled here [0] (it's
currently disabled):

1. A (Cronet native) Cronet_UrlRequestImpl::InvokeCallbackOn*() function
is run on the executor for the UrlRequest [1][2][3]. Importantly, these
do *not* run on the network thread.

2. With the exception of InvokeCallbackOnCanceled(), those functions
call DestroyRequestUnlessDone(), which via DestroyRequestDoneLocked()
calls the (Cronet common) function CronetURLRequest::Destroy() [4].

3. CronetURLRequest::Destroy() does its work via
CronetURLRequest::NetworkTasks::Destroy() on the network thread [5].

4. CronetURLRequest::NetworkTasks::Destroy() reports metrics from the
common UrlRequest to the native (and Android, etc.) UrlRequest object [6].

That is, it calls CronetURLRequest::NetworkTasks::MaybeReportMetrics()
[7], which calls OnMetricsCollected() on the CronetURLRequest::Callback
[8] to move metrics from net/ and common Cronet code to Cronet native
(or Android, etc.)-specific code.

However, since this happens on the network thread, the original native
Cronet_UrlRequestImpl::InvokeCallbackOn*() function (running on the
request executor) may have already called the the success / fail /
cancel listener.

The reason this is a problem is the app is allowed to free the native
UrlRequest object after one of the success / fail / cancel callbacks
completes, causing a use-after-free race.

(Technically, the native implementation of OnMetricsCollected() lives in
Cronet_UrlRequestImpl::NetworkTasks, the native implementor of
CronetURLRequest::Callback. Cronet_UrlRequestImpl::NetworkTasks is not
owned by its enclosing Cronet_UrlRequestImpl class but is intead owned
by the common CronetURLRequest class [9].

However, Cronet_UrlRequestImpl::NetworkTasks has a pointer to its outer
class -- we don't want to worry about pointing to freed objects.)

With this change, the metrics will have already been reported to the
native impl class *before* Cronet_UrlRequestImpl::InvokeCallbackOn*() is
run -- this was already true for the succeeded and cancelled cases, and
now it will be true for the failed callback as well.

Since metrics are only reported once [10],
CronetURLRequest::NetworkTasks::Destroy() should only report metrics in
the cancel case.

----

Note that OnUploadDataProviderError() [11] presents a similar complication,
however, it requires a different approach, as the error is produced by
the cronet API user and not by cronet internals.

Because of this, the OnUploadDataProviderError() doesn't run on the
network thread -- some logic is needed to ensure that the Metrics are
reported from the network thread and then *only after that* have the
user's error callback run on their executor.

The way this CL fixes the OnUploadDataProviderError() race (by adding
APIs accepting callbacks to the common code) is, admittedly, rather
ugly, but I'm having difficulty thinking of a cleaner way to do this.

[0]
https://cs.chromium.org/chromium/src/components/cronet/native/url_request.cc?l=338&rcl=7feadb2d0b65a8d09beee807cd48af727b0260c4

[1]
https://cs.chromium.org/chromium/src/components/cronet/native/url_request.cc?l=538&rcl=64ada39645d537bb838572865fde3f3c0875e4c6

[2]
https://cs.chromium.org/chromium/src/components/cronet/native/url_request.cc?l=547&rcl=64ada39645d537bb838572865fde3f3c0875e4c6

[3]
https://cs.chromium.org/chromium/src/components/cronet/native/url_request.cc?l=557&rcl=64ada39645d537bb838572865fde3f3c0875e4c6

[4]
https://cs.chromium.org/chromium/src/components/cronet/native/url_request.cc?l=455&rcl=fa25a7c629b94a9f1d67285105b582d9b4963f27

[5]
https://cs.chromium.org/chromium/src/components/cronet/cronet_url_request.cc?l=154&rcl=d2675bf10dc4afe5fbbd12c4fb71dca26386b2b7

[6]
https://cs.chromium.org/chromium/src/components/cronet/cronet_url_request.cc?l=331&rcl=37bb439f1e692357890d641119199dfa627aadbf

[7]

Note that in Java, and soon in native, there's also a similarly-named
method, called CronetUrlRequest.maybeReportMetrics() in Java, to report
metrics from Java Cronet to the app -- not to be confused for
CronetURLRequest::MaybeReportMetrics(), which moves metrics from Cronet
common to Cronet native / Java, etc.

[8]
https://cs.chromium.org/chromium/src/components/cronet/cronet_url_request.cc?l=369&rcl=2ecbb6045e164f0d72fb9c842790565dbc28b35d

[9]
https://cs.chromium.org/chromium/src/components/cronet/cronet_url_request.h?l=251&rcl=4d0fb14d4b6d4459569938a92caf8657074c9d31

[10]
https://cs.chromium.org/chromium/src/components/cronet/cronet_url_request.cc?l=366&rcl=37bb439f1e692357890d641119199dfa627aadbf

[11]
https://cs.chromium.org/chromium/src/components/cronet/native/url_request.cc?l=481&rcl=5f5ee515ff92bfbece2e840623a90a272404a9cb

Bug: 879208
Change-Id: I70c2e102b2ae3590f66c847c16f72f25c45024db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1529356
Commit-Queue: Caleb Raitto <caraitto@chromium.org>
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Reviewed-by: Misha Efimov <mef@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645016}