overview: Overview opens when pressing F5 with app list transient focused.
If the app list window is open, and it has a transient window opened and
focused, pressing F5 will only close the app list. THere may be a flicker
of the windows getting transformed, as it is currently attempting to
enter overview before being canceled.

This cl ignores windows that are parented to the app list container, same
as window cycle list.

Test: manual
Bug: 941945
Change-Id: I347e5ae50d3ad421e8a0fe853695589f9a831186
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1524093
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645112}