Fix item drag release animation.
Previously, when an AppListItemView was released from drag, the item's
icon would quickly disappear and then reappear. This was due to the
proxy icon disappearing and the drag_view_ icon fading in over the
default transition duration of 200ms.

Now, a transition duration of 0ms is used so that the drag_view_
appears exactly when the proxy icon disappears.

Bug: 946005
Change-Id: Icf03f3ee8e70a5eb2197b00bea6ed5d018b2d6bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1539269
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645117}