[android] Use ConnectivityManager to fetch proxy information.
On Android M+ we can use ConnectivityManager to fetch proxy information
rather than relying on system properties or intent extras.

In order to keep WebView applications that use reflection to trigger the
proxy change broadcast working, we must continue to register the old
ProxyReceiver object. To ensure that this is not triggered by real
system broadcasts, we register it with an empty intent filter.

Bug: 851995
Bug: 908085
Change-Id: Ibfa6c35693301868f8a63ea924649e4c2bc8a328
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1526268
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Commit-Queue: Tobias Sargeant <tobiasjs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645223}