Exposes affinity in text fields
Affinity is retrieved from the frame selection,
i.e. from the currently visible selection.
Text fields don't expose their affinity in Javascript,
so it doesn't make sense to cache its value so that it could be retrieved when the text field is not focused.
Any time the text field is not focused, we return
a cached selection that has a downstream affinity.
R=dmazzoni@chromium.org, yosin@chromium.org

Bug: 639340
Change-Id: If488edc29cd72e9835c7120466ea8dfa01d5841d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1535051
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645337}