Android: Fix Python deps for bundle wrapper scripts.
The generated bundle wrapper scripts don't change when the Python deps
of the wrapper scripts are modified. This means that targets that
depend on the wrapper scripts won't be rebuilt when Python deps of the
wrapper scripts change.

This CL changes create_bundle_wrapper_script.py to have a dependency
on apk_operations so that the generated wrapper will be regenerated
when Python deps change.

Change-Id: I34d8fec0c324b19b6ffa39977668656088714ad6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1543460
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Eric Stevenson <estevenson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645350}