Resolve ambiguity between ::features and ::printing::features
There is both a ::features namespace and a ::printing::features
namespace (which is kind of against the coding style), so
if you write just "features" inside the printing namespace,
the compiler will resolve "features" differently depending on
if it is aware ::printing::features exists.

In certain jumbo builds, the compiler will be aware that
::printing::features exists when compiling the code and
will resolve it incorrectly. This patch removes the ambiguity
by adding a :: prefix.

Bug: 945619
Change-Id: I569878be26debb77dea69679dd37414f602ec1a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1541226
Auto-Submit: Daniel Bratell <bratell@opera.com>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645422}