Remove synchronous composite APIs from WebWidget and WebWidgetClient (7/n)
Mostly these methods are called from test_runner, where it can call
directly to the RenderWidget now instead of going through
WebWidget -> WebWidgetClient -> RenderWidget.

This deletes the API from both WebWidget and WebWidgetClient.

test_runner also needs access to the RenderWidget behind the current
WebPagePopup which it can't get to without casting to internal blink
types, which it can't do as it lives in content/. So to get to that we
do need a ForTesting API method on WebPagePopup to get the client.

The other consumer was in web_test/ which doesn't get to access
RenderWidget directly (yet?) so it needs to go through content/public/.
To satisfy it, move the API from WebWidget over to RenderFrame directly
which is implemented by WebFrameWidgetProxy who will call over to
RenderWidget.

In this CL we learn that there are 4+ ways to request a synchronous
composite in web tests.
1. Through JS bindings which go to TestRunnerBindings. This now goes
over to the main WebWidgetTestProxy.
2. Through standard compositor frame scheduling, via ScheduleAnimation.
This now also goes over to shared code with 1 called from AnimateNow()
in WebWidgetTestProxy.
3. Through the browser asking to do a composite for a pixel output grab
which goes through WebTestRenderFrameObserver to that frame's
RenderFrame which gets picked up now by WebFrameTestProxy, who calls
WebWidgetTestProxy.

^ These are all valid and share a destination now.

4. CompositeAndReadbackAsync schedules a composite via its own path
instead of via ScheduleAnimation(). This is a legacy pixel test path
and I am going to work to remove it.
5. More things that schedule one-off sync composites in LayerTreeView
and I would like to move these all out to WebWidgetTestProxy if possible
in followup CLs.

R=dcheng@chromium.org

Bug: 912193
Change-Id: I4e676d25d10bfc7bbe63df7f0bd504ef864668cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1539900
Commit-Queue: Avi Drissman <avi@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645437}