[CrOS Cellular] Display the cellular connectivity type in quick settings.
Previously, when a cellular network was the only connected network, we
would display a sublabel in quick settings which included a description
(in words) of the signal strength (e.g., "weak" or "medium"). Our UX
designer has suggested that we instead replace this label with the
connectivity type (e.g., "LTE") instead.

This CL replaces this label with the connectivity type for cellular
networks and leaves all other network types the same. That is, Wi-Fi
sublabels will still use the signal strength descriptions such as
"weak".

Bug: 945949
Change-Id: Id550f58d287c2d7b01dc8a3d44cc2101e9b1e57e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1542721
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645449}