Changed how WindowReorderer::ReorderChildWindows work to be more efficient.
Currently, Window::Reorderer::ReorderChildWindows perform the following
steps:
1. GetOrderOfViewsWithLayers to get the View in z-order.
2. For each of those View, get the compositor Layer, and use the
   StackAtBottom method to place it as the bottom child of its
   current parent layer. StackAtBottom will use StackRelativeTo:
   a. Find the index of the child Layer that we are moving.
   b. Erase it from the children vector.
   c. Place it at the front of the children vector.
   d. Remove the cc_layer_ from its parent.
   e. Add the cc_layer_ to the same parent as before but different index.

This is performed for every single Layer of the Views. Step 2 can be
performed for the entire children at once since we are only reordering
the children layer, and we know the parent will remain the same.
Similarly, step 2d and 2e can be performed simultaneously for all
cc_layer_ of the children, and we do not have to remove from the parent
first since the parent stays the same.

This CL reduces the runtime of ReorderChildWindows from ~2-4 ms (for
~10 tabs open) to less than 0.3 ms per call.

Bug: 938298
Change-Id: I93712197bddbe608cf5c4c163a33b3dcd073bf81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1529580
Commit-Queue: Edwin Joe <ejoe@google.com>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645462}