[VK] Specialize PopulateKeyboardContents based on its boolean arg.
PopulateKeyboardContents is used twice: once with |show_keyboard| = true
and once with false. We inline PopulateKeyboardContents(..., false)
into LoadKeyboardWindowInBackground, where the code can be simplified
because the state is guaranteed to be INITIAL. We can then also simplify
PopulateKeyboardContents(..., true) as the state is guaranteed to be
not INITIAL.

This reduces the amount of branching the code so it makes it easier to
understand.

Bug: 845780
Change-Id: I34229b160b9dbb8c28bb29938331b93f6dcee79b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1530498
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645624}