Move content::ManifestParser with tests to blink.
This is a part of moving manifest implementation to blink.
This CL moves content::ManifestParser to blink and introduces
WebManifestParser in blink/public/web to expose the parser to
content::ManifestManager.

blink::ManifestManager uses still base::JSONParser but it will
replace with blink::JSONParse at the end of this task.

Bug: 704441
Change-Id: I6ecea28d15fbb92f81a80f68ae011b2f971b7e4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1535207
Commit-Queue: Miyoung Shin <myid.shin@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645637}