Fix the app menu's update notification icon on dev+canary.
The update menu item's icon was being chosen based on the state of the
UpgradeDetector, while the menu's itself was chosen based on the
AppMenuIconController's chosen severity level. With this change, both
icons are created by the AppMenuIconController using the theme
properties of the hosting ToolbarView.

BUG=941448

Change-Id: If77a999473ff732e95a2fc27afb9297883aeeaef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1536194
Commit-Queue: Greg Thompson <grt@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645713}