[css-properties-values-api] Calculate transitions on computed values.
We currently calculate the transition update for custom properties before
the (pre-animated) computed values are known. This is because font-size
isn't applied yet at the time of the transition update, which means 'em'
units can't be resolved correctly. We currently circumvent this issue
somewhat by producing a tentative "computed" value for registered
properties, which doesn't undergo conversion, and then perform the
animation update using that value. This means that instead of transitioning
from a computed value to another computed value, we are essentially
transitioning from a computed value to a specified value.

This causes issues e.g. when a value changes from from 10px to 10em. We
can't really know whether we should transition or not, since we don't
know what the 'em' unit represents.

The real solution to this is the Cascade Project (Issue 947004), but
this is a complicated task. Meanwhile, this CL addresses the most
important issues in another way:

 * Split out application of high priority properties in a separate
   function.
 * Converted values (e.g. em -> px) even when absolutize=false. This
   ensures that the call to ComputeRegisteredVariables produces actual
   computed values, rather than specified values.
 * Apply high priority properties before calculating the animation
   update. This means we can resolve 'em' units, and detect transitions
   correctly.
 * If we _do_ animate custom properties, then we apply high priority
   properties again. This is needed, because high priority properties
   themselves can refer to properties which are animated. The first
   application pass of the high priority properties produce the
   pre-animated values to calculate the update. The second pass produces
   the final values with animation effects of custom properties taken
   into account.

This CL also avoids applying the custom properties twice, like we did
before. I don't remember right now why that was necessary. Probably
because we needed the variables present to resolve any var()-references
in animation- and transition- properties, but, since variable resolution
destroys the var()-references, we had to "restore" these var()-references
again, so they could be resolved against the animated values. (Similar
to what is now necessary with high priority properties). This is not
necessary anymore, because var()-resolutions that take place before
CSSVariableResolver::ResolveVariableDefinitions are non-destructive
(absolutize=false), meaning that we don't lose the var() references.
This is kind of a drive-by change, though. It could have been made as a
separate CL, as the non-destructive resolution behavior was already on
master.

Another slight improvement is the elimination of the
ComputeRegisteredVariables step, if we don't have animations. We also
avoid doing that more than once. I suspect this could have been done as a
separate step as well.

Change-Id: I10d549b110b22e1c21ab755cb52fbb7fb84bae0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1541103
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645725}