Update SelfSizingTableView to support other row layout types
When the table view is using Autolayout to automatically calculate row
heights, rowForHeightAtIndexPath doesn't actually return any height.
Instead, contentSize should give the size of the elements in the table
view.

However, the contentSize is incorrect when intrinsicContentSize is
called. At that point, it uses estimatedRowHeight for the contentSize
instead of the actual row heights. Because intrinsicContentSize
uses contentSize, it can be recalculated whenever contentSize changes.

Bug: 945317
Change-Id: Ie4343ea2e8fa811117c8a3efcefd73ba6a4f7a1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1541255
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645758}