Remove unneeded canvas code from mirroring.html
The canvas that we're trying to set the property on doesn't exist, and
isn't needed for the code to mirror properly.  Additionally, we were
getting a null access error because we tried to access the base layer
on the same frame that it was updated.  None of this is needed for the
mirroring to properly work, so removing it.

Bug: 930271
Change-Id: I8a876f092543cb453b6d2770d14553840f028b34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1545701
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645798}