[android] When ConnectivityManager returns null ProxyInfo, clear proxy
Prior to using ConnectivityManager, we treated a null proxy info object
as meaning that we should look up the proxy settings from the system
properties. However ConnectivityManager is authoritative, so we should
treat null in that case as indicating that proxy info should be cleared.

Bug: 851995
Bug: 908085
Change-Id: Ia285a01fa0683dbc39c852774a051e42cf05fe01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1541247
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Commit-Queue: Tobias Sargeant <tobiasjs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645799}