Remove ContextFactoryObserver::OnLostVizProcess().
This observer function is only handled in DelegatedFrameHost to clear a
saved frame bit for FrameEvictor. I talked with samans@ he doesn't
think this is necessary. The only change I see here is we might try to
evict a surface that doesn't exist in FrameSinkManagerImpl, which is
already a case we handle.

Bug: 947279
Change-Id: If49a193dee717c179694f7d69ea1bad3f4782215
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1544619
Reviewed-by: Saman Sami <samans@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645806}