sensors: Make SensorDeviceManager use UdevWatcher rather than DeviceMonitorLinux
DeviceMonitorLinux is a singleton that happens to be created in a task
runner that can block at the moment. Make SensorDeviceManager use
UdevWatcher instead, as it allows SensorDeviceManager to run in any
sequence (that can block).

This, in turn, allows us to completely rid //services/device/generic_sensor
of the dependency on the services-global "FILE task runner" and lets us
clean up our code a little bit.

One significant change in the unit tests is that we now create one
PlatformSensorProviderLinux per test, as ScopedTaskEnvironment
invalidates the provider's |blocking_task_runner_| when it is destroyed.
One consequence is that we can remove the Shutdown() method from the
provider and get even further with the code clean-up.

Change-Id: I7080e712106bac1a3a406e1f392487e5ca346971
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1541090
Commit-Queue: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645874}