mac: Actually use DWARF4.
clang defaults to DWARF2 on macOS 10.10 / iOS 8 and lower. Our iOS deployment
target is iOS 11, but our macOS deployment target is still 10.10, so explicitly
opt in to DWARF4.

The reason for the conservative default is that tools like atos don't support
DWARF4 on macOS 10.10, but we require a newer version for building chromium
anyway (just not for running it).

Bug: 946404,942155
Change-Id: I35a5417fce823d8788823ac123b3aef5e0e95c84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1545694
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645889}