Minor refactor of ValueForContextTime
This function actually returns two values: the value from the timeline
and a bool in the parameter list indicating whether the value was
computed from the timeline. (This can be false if the apprpriate lock
could not be obtained.)

Make these two return values explicit by returning a tuple for the two
values.

Bug: 938550
Change-Id: I421036f8770c355a29c722fd7c6758082446c6b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1546079
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645920}