Added temp UMA logging for renderer hangs.
Adding temporary UMA statistics to determine what the
browser was waiting for when it determined that the
renderer was hung.

This is similar to logging added in r406039 and removed in
r455764, but the hang monitor is now different enough that
the UMA values cannot be reused.

Bug: 947559, 938647
Change-Id: Ia4d44e09ec653f23fcc9d8ea0fcde05a887d21da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1546250
Commit-Queue: Chris Mumford <cmumford@google.com>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645979}