Pass file size of RAR entries to UpdateArchiveAnalyzerResults
Since it's not possible to truncate a file from within a sandbox, we
need to track the file size independently, then pass this information
from unrar to SetLengthAndDigestForContainedFile.

TESTED=Downloaded a test RAR and displayed the hash entries. The hashes
were wrong without the change, and correct with this change.

Bug: 944633
Change-Id: I34bd06aa98330e3d3750910780e80246f60da2ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1540418
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645991}