Changing Print Preview errors and fixing some bugs
In destination settings:
- Do not hide the dropdown if the destination is ready or expected to
be ready imminently (i.e. capabilities are already known, and just
waiting for the event from DestinationStore). This causes a loss of
focus.
- Disable the dropdown only for fatal errors which should disable the
entire UI, or when there is a validation error and the destination is
updated and the state is not NOT_READY. Need to check state explicitly,
as the destination needs to be updated before updating the state to
READY. If we only check destinationState and disabled, the dropdown
will be momentarily disabled when the destination updates, causing a
loss of focus.
Also:
- Introduce a new error property. This simplifies propagating
printer errors between the preview area and the destination settings,
so that the dropdown will be disabled correctly.
- Remove some redundant states in destination settings and preview
area, and use the error property to distinguish where necessary.

Bug: 943072, 942444
Change-Id: Ib2ff3f65971ace8f5db71ed1bf2a33b9bf28dbea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1539189
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#646062}