Calculate the overlay occluding damage rect in surface aggregation.
This is for underlay video power optimization. The damage rect on top of a
video quad is calculated in surface aggregation and saved in the share quad
state. The Overlay Processor can use this info to reduce the damage rect to
zero if possible.

The function ProcessOverlayOccludingDamage() is called at each surface in
CopyPasses(). Each surface damage rect is transformed to the frame root target
space and added to |damage_rects_union_of_surfaces_on_top_|. To get the
occluding damage rect, just intersect |damage_rects_union_of_surfaces_on_top_|
with the overlay in root target space. Because surfaces are processed from top
to bottom in CopyPasses(), we only need to look at
|damage_rects_union_of_surfaces_on_top_| for this calculation.

When NeedsNonZeroOverlayOccludingDamageRect() is false, no transformation is
performed. It only checks if damage_rects_union_of_surfaces_on_top_ is empty
and returns this case. Currently, NeedsNonZeroOverlayOccludingDamageRect() is
true for Windows only.

This CL will fix an issue from the previous implementation with
|has_surface_damage| where damage info cannot be passed down to the overlay
processor when there is no quad in a render pass. |has_surface_damage| will be
deleted when no one uses it.

Bug:868561

Change-Id: I00bfaa13a4b6c374cc42e9bbe6b7f64c350ef76d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1510075
Reviewed-by: enne <enne@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#646102}