Remove disable-features=GrantNotificationsToDSE from ClearBrowsingDataPreferencesTest.java
This flag is now enabled by default and should be removed. Previously
some tests failed with it enabled. This CL fixes important_sites_util.cc
so that those tests pass.

More specifically, sites are usually marked as important if they have
the Notifications permission. However the default search engine is
granted the notifications permission by default. It should not be added
to the important sites list in this case. Doing so results in the
default search engine always being listed as an important site even if
there is not stored data associated with it. This was causing existing
tests to fail.

Change-Id: Idca4b5e6e5749b507e4e1cffa412ee47734dc0a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1545435
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Raymes Khoury <raymes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#646162}