desktop-pwas: Split create_shortcuts into desktop shortcut and pin app
For policy installed Web Apps, the option to create a desktop shortcut
is separate from the option to pin the app to the shelf e.g. an admin
could choose to create a desktop shortcut on Windows and Linux but
not pin the app to the shelf on Chrome OS.

To support this use case, we separate the "create_shortcuts" option
into "add_to_applications_menu", "add_to_desktop", and
"add_to_quick_launch_bar".

Bug: 844278
Change-Id: I1c99c693fabd759c8319304d4b7212c7913e605c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1534907
Commit-Queue: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#646170}