Rename DeviceLoginScreenAppInstallList policy
This CL renames the "DeviceLoginScreenAppInstallList" to
"DeviceLoginScreenExtensions". The plan is to also allow whitelisted
extensions on the sign-in screen since apps will be deprecated soon.
This CL just renames the policy and doesn't perform any logic changes
(will be coming in a seperate CL).
This policy does not have server-side UI integration yet.

Bug: 938943
Change-Id: Id81826e0e50a2a3862b1569375720e4d6dd90248
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1505448
Commit-Queue: Alexander Hendrich <hendrich@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#646227}