Fix crash in UiControllerAndroid::UpdateActions
Before this patch, when, during an AA prompt action, Chrome would crash
when users click on a link that opens a new window from within CCT.

This happened, because the Controller would call the UIDelegate during
the shutdown procedure, after calling WillShutdown, which the contract
of UIDelegate explicitly forbids.

This patch guarantees that no UIDelegate is called after WillShutdown
and ignores attempts to switch the controller state away from the
STOPPED state.

This is a workaround. It is meant to be easy and safe to cherry-pick.
The proper fix would be to shutdown immediately as part of b/128300038

Bug: 947403
Change-Id: I828eb7e53a1c789a867052b9df95b90d2b253cc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1546093
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#646256}