reland: InvalidateLayout now results in a later call to Layout
Patchset 1 is the original patch. There are two additional fixes here
. ChromeNativeAppWindowViewsAuraAshBrowserTest was expecting a sync
layout. It already had logic to wait for view to be in position, so I
just enabled that path always.
. HostedAppBrowserController::GetAppShortName() may be called when the
extension is null. See 947391 for the exact stack. This is likely
necessitated by layout being called at a different time then it was
before.

Prior to this it was up to consumers of View to know when the right
time to call Layout() was. This is error prone, and easy to get wrong.
This patch makes it so a call to InvalidateLayout() results in a later
call to Layout() (starting from the root). The call to Layout()
happens right before the damage rects are sent to the compositor.

This only cleans up views side. No doubt other cleanup can happen in
other places. Additionally, there are future changes that should happen:
. add/remove should really InvalidateLayout().
. It isn't always necessary to start the Layout from the root. Some
views size *never* changes (such as the BookmarkBar), in which case
layout need not start from the root. This is a future optimization.

BUG=627663, 947391
TEST=covered by tests

Change-Id: I84c057b3df078659734d8a73bae06cd76cd60575
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1547051
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#646391}