cros: BaseWebUIHandler can now record incoming/outgoing messages
- Rework how JSCallsContainer stores pending calls; store the actual values
  directly instead of storing a closure
- Rework BaseWebUIHandler::CallJS so it is a single template function with no
  overloads
- BaseWebUIHandler::CallJS can now optionally now record the outgoing JS call
- BaseWebUIHandler::Add(Raw)Callback can now optionally record the incoming JS
  call

The recording can be used by a test to validate what was actually sent to JS as
an alternative to creating a *View class that is then mocked.

Deferred JS execution is now built on-top of the recording capabilities instead
of being an isolated/independent feature.

TBR=michaelpg@chromium.org
Bug: 928555

Change-Id: Ie3e7d9dff3cd9224810124960b9ba481aef9f0fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1534710
Commit-Queue: Jacob Dufault <jdufault@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#646510}