devtools/tracing: Support perfetto backend + proto output.
Adds the ability for TracingHandler to use the new Perfetto Consumer
interface to control a tracing session. For now, we still support the
old TracingController backend via a new TracingSession virtual class.

Also adds a proto output format option to the protocol.
Further adds necessary missing methods to the perfetto ConsumerHost
interface.

To support startup tracing with perfetto + retrieving of the trace via
DevTools, TracingHandler needs to become the consumer that initiates
and controls the startup session, instead of TracingController.
Otherwise, the trace data can't be read via DevTools. For this, we add a
new command line flag.

Background: We're replacing the old tracing backend (TracingController +
TraceLog) with Perfetto! In this process, former clients of
TracingController become consumers of the Perfetto service instead.

Bug: 925142, 938879
Change-Id: I2ac389e4b9a6d258b784acaf76043417be4f4f3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1505934
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Alexei Filippov <alph@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: oysteine <oysteine@chromium.org>
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#646895}