[BGPT] Use visual viewport for fixed NFSRs.
NonFastScrollableRegions (NFSRs) are set on the scrolling contents layer
of the layout viewport scroller so the compositor can tell when events
occur over uncomposited scrollers (or plugins) which requires these
events to be handled on the main thread.

BlinkGenPropertyTrees broke this behavior for uncomposited scrollers
that are contained by a position: fixed layer. When an event arrives on
the compositor and hits a ScrollNode, we walk up the scroll tree
checking each ancestor ScrollNode's layer to determine if the event
falls inside a NonFastScrollableRegion. Prior to BGPT, the scroll tree
hierarchy looked like this:

inner scroll
     |-> outer scroll
             |-> position:fixed

However, enabling BGPT produces the following hierarchy:

inner scroll
     |-> outer scroll
     |-> position:fixed

That is, position: fixed layers, which do not scroll with the
layout/outer viewport, are now parented to the inner/visual viewport.
This is semantcally more correct as the pre-BGPT hierarchy implied that
a fixed node will scroll with the layout viewport. The reason fixed
scrolling worked pre-BGPT is the PropertyTreeBuilder produced this
hierarchy in the TransformTree which determines how the layers are
drawn.

However, the algorithm for checking NonFastScrollRegions assumes we'll
visit the outer/layout viewport's ScrollNode, which is where all the
NFSR are currently stored (even though we check each ancestor
ScrollNode's NFSR, Blink only sets it on the layout viewport). In BGPT,
if we start at the position:fixed ScrollNode we'll miss that there's a
NFSR.

This CL fixes the issue by placing all NFSRs that are fixed w.r.t the
layout viewport on the inner viewport, leaving non-fixed ones on the
layout viewport. This actually fixes a correctness issue (the
scrolled-doc test cases) since placing fixed regions on the layout
viewport means they're incorrect when the layout viewport is scrolled.

Additionally, I found several cases where the NFSR implementation breaks
so I've added several new tests, here's a summary:

                                       !BGPT      BGPT      w/this patch
for non-composited-scroller-in:
pos-fixed-scrolled-doc.html             FAIL      PASS*     PASS
pos-fixed-scroller-scrolled-doc.html    FAIL      FAIL      PASS
pos-fixed-scroller.html                 PASS      FAIL      PASS
scrolled-absolute-scroller.html         FAIL      FAIL      FAIL
scrolled-position-fixed-scroller.html   FAIL      FAIL      FAIL
sticky.html                             FAIL      FAIL      FAIL

The w/patch behavior is identical for both BGPT enabled and disabled on
all these tests. All these tests pass when run without threaded input.

*This passes with BGPT because we deem the hit test unreliable due to
the fact the fixed layer's first scrollable ancestor != first hit
scroll layer in LTHI::IsInitialScrollHitTestReliable but that's
not really a fix, making the fixed layer a scroller breaks even with
BGPT as shown in the test case below this one.

Bug: 938572,817600
Change-Id: I5b03efe1f36521ed55b664c9e41c86a1a81a42dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1547084
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#646907}