Omnibox: Remove LocationBarModel::GetEVCertName().
This method is only called internally, and so I just inlined the (short)
body into its caller. All external callers have already transitioned
to using GetSecureDisplayText() which is more general.

This is part of a larger refactoring to get rid of all calls to
LocationBarModel::input_in_progress() or calls that depend on the user's
edit state.

And that's part of a larger cross-platform omnibox UI architecture
project.

Bug: 912722
Change-Id: Ida340d462d384d7814448df0bd0794ad921c7fad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549205
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#646973}