[PE] Include visual overflow of self-painting LayoutTableCells for rowspan.
If a table cell spans multiple rows, then the row needs to expand its
visual overflow to include the visual overflow of that cell, even if
the cell is self-painting. This is because the table row is the delegated
painter of the background and collapsed borders of such cells.

Bug: 947713

Change-Id: I690991de15ad4c527a8647ab11e5fb302269a2fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1548868
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647008}