[FCP++] The queue should stores pointer instead of node id
The main change in this patch is to change the type used in
image/text's queue for paint time. The type used to be node-id, this
patch changes it to be the record pointer.

The purpose of this change is to avoid exposing node id after turning
text/image into record. This will lay the ground work for supporting
the tracking of background image individually. Avoiding node id will
help to make background-image and non-background-image handling more
consistent.

Others:
* insert --> emplace
* push --> emplace

Bug: 936149

Change-Id: Ia71af7100e84bcd24c29bc55e4b9ef6d1f24b7ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1544877
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647016}