Add katie, jamescook to accessibility owners files
Promote katie to an owner of nearly all accessibility code. She's always
free to delegate or decline if she's not the best reviewer.

Add jamescook as an owner of the aura and views accessibility code,
because we need an additional reviewer for mash accessibility work
who's familiar with both mash and accessibility.

Bug: none
Change-Id: I1f820bf186538a5a3eb92d05603497f21bb53bd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549883
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647070}