[serial] Implement tab indicator triggered by open()
As long as there is an active connection between a frame in a tab and a
serial port the tab should display an indicator.

This is accomplished by adding an additional Mojo pipe that is kept open
between the device service and the browser processes for as long as the
render process has a connection to the port.

Bug: 917204
Change-Id: I6441070f3473bd6b4a98dae98844a89ee4eb2329
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1546502
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647109}