CSS Alt text for Content property
The content property should accept alternative text to be supplied after
a forward slash after the last content list. This alternative text can
then be used for non visual mediums.

This change extends the CSSParser to accept the alternative text syntax,
and introduces a new subclass of ContentData, AltTextContentData, which
does not require the creation of a LayoutObject but is used to store the
alternative text, until the Accessibility Tree is built and it can be
retrieved to use in name calculation.

This change only applies alt text to pseudo elements, other elements.
Making the alt text work for other elements that support the content
property will be included in a follow up CL.

Spec: https://drafts.csswg.org/css-content/#alt
BUG=860016

Change-Id: I4891483e24531da0f42f800377af6ff77d3c8d89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1412094
Commit-Queue: Meredith Lane <meredithl@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647143}