Permit "unresolved" pseudo CSS property only on Custom Element V0 flag.
This patch makes the property only working on Custom Element V0 flag
because the property is a part of the V0.

There is a case where a style sheet doesn't have a document,
which is needed to check if the feature is enabled on origin trials.
However this is rare and for feature conservativity, we permit the property.

Bug: 937111
Change-Id: I9dbec21c3bcda93383d90eb7c34e1ce28ea2a418
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1496258
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Yoichi Osato <yoichio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647159}