[AF Wallet Metadata] Rescale sync metric for last use date of entities
This CL fixes metrics introduced in CL 1515533 and records the time
diffs in minutes (instead of in ms).

Prior to this change, most recorded values were in the last bucket so it
was impossible to draw any conclusions from the metrics.

Bug: 853688
Change-Id: I68d0d200e2ead24ae20be0a6ce28fbadbeda6968
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549076
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Auto-Submit: Jan Krcal <jkrcal@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647187}