Virtualize TransformStream
Make TransformStream delegate to a new class TransformStreamWrapper for
the creation of the readable and writable pair. This will allow adding
in a native implementation in future. TransformStream itself remains a
concrete class as after the initial creation it has no behaviour
specific to the underlying implementation.

Most of the code in transform_stream_wrapper.cc has been moved unchanged
from transform_stream.cc.

Rename TransformStreamDefaultController to
TransformStreamDefaultControllerWrapper and make it inherit from a new
abstract interface, TransformStreamDefaultControllerInterface. This will
permit plugging in an alternate implementation.

Modify TextDecoderStream and TextEncoderStream to use the new interface
name.

This is preparation for addition of the new C++ implementation described
in the design doc:

https://docs.google.com/document/d/1n0IIRmJb0R-DFc2IhhJfS2-LUwl6iKSBNaR0klr3o40/edit

BUG=902633

Change-Id: I5ad376a951dd426e3a8dffbda18643f44dc8daab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1545952
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647191}