Add ComputeTargetOcclusionForWindow to WindowOcclusionTracker.
This also makes finishing a visibility animation on a layer trigger a
recomputation in the occlusion tracker, as this can affect the result.

ComputeTargetOcclusionForWindow is used for ARC windows to send an
occlusion fraction taking into account animated windows. This is
necessary because we need to deliver the occlusion fraction as if
animation had completed before activation is sent. See
crbug.com/939258 for more details.

This CL fixes auto-PIP triggering for showing new windows. Some cases
will need to be done in a follow-up CL. These are the cases that exist:
1. Occlusion by showing a new window - handled by this CL
2. Occlusion by activating a new window via overview
  - Not handled. The overview activates an overview specific window
    upon entry. To fix, we will need to remove the behaviour
    (crbug.com/948492) and compute target occlusion upon window
    stacking change. I will do this in a follow-up CL.
3. Occlusion by activating a new window via Alt-Tab
  - Not handled. This is a similar issue to overview, where a alt-tab
    specific window is activated upon alt-tab.
4. Occlusion by activating a new window via shelf icon
  - Handled via the regular occlusion pathway.

Bug: 939258
Bug: b/126665003
Test: added unit test
Change-Id: I71f3b822636e7559f1b6ae81189cd69521471a57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1540898
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647194}