Make top level menus scrollable if too tall.
There are menus available at the top of files app that can have
varying height. If the window is resized so that the window is
smaller than the menu height, it should be possible to scroll
the menu to pick entries that are out of view.

This CL checks if the menu fits inside the height of the viewport
and limits its height and enables scrolling if so.

Bug: 947910
Tests: Added unit test.
Change-Id: I8d3daf900ff1c5a8a2c852088fab319e7b0448dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549297
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647236}