Replace blink::WebRTCCertificateCallback by a base::OnceCallback instance
This is part of the effort to minimize the amount of APIs exposed by
Blink, where classes and enums are being assessed, and its usage audited.

At this time, it was verified that blink::WebRTCCertificateCallback (and
blink::WebRTCCertificateObserver) inherit from blink::WebCallbacks, a class
designed to bridge callbacks passing from Chromium to Blink (and vice-versa)
prior to OnionSoup.

Similarly to [1], [2] and [3] this CL replaces the use of blink::WebCallbacks
in WebRTCCertificateObserver and WebRTCCertificateCallback by a
base::OnceCallback instance. Note that this callback takes an optional
(base::Optional) parameter, so it is possible to handle success and error
cases.

No functionality change is expected.

[1] https://crrev.com/c/1545931
[2] https://crrev.com/c/1544529
[3] https://crrev.com/c/1548294

TBR=dgozman@chromium.org
CC=​​​blink-reviews-vendor@chromium.org

BUG=919392,945851

Change-Id: I34de1ecb7d78f90cd5bdd21e545a4eec0a571d81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549353
Commit-Queue: Antonio Gomes <tonikitoo@igalia.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647261}