base: Fix Value layout for GCC
It turns out that the previous changes to the base::Value
layout broke GCC compilation (see [1] for details).

This CL fixes the situation by using a new DoubleStorage
type that will store double values in a 4-byte aligned
struct, with bit_cast<> being used to convert between
double and DoubleStorage values in the implementation.

This ensures that base::Value remains as small as possible
in all cases. The small penalty is that loading/storing
double values on 32-bit ARM is slightly slower due to
the fact that the value is no longer 8-byte aligned.

+ Fix the ValuesTest.SizeOfValue test to work correctly,
  and disable it for debug builds, so it doesn't fail
  because debug versions of the internal containers
  are larger on certain systems.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1472716

BUG=646113
R=dcheng@chromium.org, pasko@chromium.org, lizeb@chromium.org, jdoerrie@chromium.org, jose.dapena@lge.com

Change-Id: I9a365407dc064ba1bdc19859706f4154a495921e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1550363
Commit-Queue: David Turner <digit@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647271}