gpu: Disable SurfaceControl on low end devices.
Video underlays need the display buffers to support alpha, which
requires reallocating the buffers in display's BufferQueue when these
overlays are present. This will need some heuristics to avoid constant
reallocations if a video is being scrolled in and out of the viewport.

Since this is only needed for low end devices and the combination of
those with Android Q is going to be lower, disable SurfaceControl on
these devices for now.

R=piman@chromium.org

Bug: 946360
Change-Id: I3c3019e6f4e44df4c34242bc34d97865e0b34de6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549208
Auto-Submit: Khushal <khushalsagar@chromium.org>
Commit-Queue: Antoine Labour <piman@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647318}