[GCPW] Fix mdm enrollment check on Windows 10 version 1703
The function GetDeviceRegistrationInfo does not exist in
mdmregistraion.dll on version 1703 so we have to fallback to the
less accurate IsDeviceRegisteredWithManagement in this case.

This function may return a false positive (enrolled to MDM but not
on the correct server) but the behavior is better than always
forcing the user to go through online sign in.

Bug: 949150
Change-Id: I072cdce78fcad3d1be24da7dcbf112ce50b17237
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1551099
Commit-Queue: Tien Mai <tienmai@chromium.org>
Reviewed-by: Roger Tawa <rogerta@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647350}