DL: Add size option to acquire().
This patch adds the size option to acquire. The current implementation
isn't quite correct. It's just using the same implementation as we would
before with the locked frame rect. However, this is the first step to
it working correctly, and in theory it should already be preventing
layout when we acquire something.

R=chrishtr@chromium.org, rakina@chromium.org

Bug: 882663
Change-Id: Ib09177ad0c3d4f732458a249023ca189d85ec90c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1544577
Commit-Queue: vmpstr <vmpstr@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647352}