Change CSS properties on text track list items to prevent overflow
Checkboxes for subtitle tracks were appearing on a different line than
their corresponding label in the VR-Browser. This also reproed on Clank.

This CL changes the CSS properties that caused them to be drawn on the next line.
This also allows the text-track list to be somewhat wider, since sometimes
items in the list would be cut off.

Bug: 876353
Change-Id: I90bfabd335836c95516de5d6a92d1908a4ef2a17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1545112
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Commit-Queue: Will Cassella <cassew@google.com>
Cr-Commit-Position: refs/heads/master@{#647381}