Use placeholder image for tabs which have no preview data (yet).
This affects the preview images shown on the new tab hover cards
feature only.

Colors are added for the background and foreground of the "no preview
available" image. The colors are chosen based on UX mocks and are
designed to work properly in both light and dark UI themes.

The image itself is a larger version of the vectorized "no favicon"
globe icon, taking up 64x64 DIP.

Bug: 928954
Change-Id: I4bb5c88bf16a78007f89e0e79a37480c34361e26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1548775
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647460}