Fix invalid surface embedding on Android.
We were occasionally embedding invalid viz::LocalSurfaceIds on Android.
This in the past has led to serialization failures. Though we've been
mostly lucky in that subsequent events on Android led to needing new
surfaces before the serialization.

This was occurring because we always embed a surface in response to
RenderWidgetHostViewAndroid::SyncrhonizeVisualProperties. What was
happening was that the surface was being evicted, before we received
an acknowledgement from our child to a previous surface synchronization
notice.

Eviction would invalidate the current ID. The child, having no changes
itself, would not advance the id. So we'd embed the invalid one.

This change adds an early exit to SynchronizeVisualProperties to avoid
this. Subsequent work to begin showing again will lead to proper
embedding.

This change also adds a CHECK in DelegatedFrameHostAndroid to ensure
we don't try embedding invalid ids.

TEST=RenderWidgetHostViewAndroidTest.NoSurfaceSynchronizationWhileEvicted
Also most of chrome_public_test_apk had tests which were hitting
the CHECK in DelegatedFrameHostAndroid

Bug: 934905
Change-Id: I5cb976c81125200484880ff89a02e8e740d85ba3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1538639
Reviewed-by: Eric Karl <ericrk@chromium.org>
Reviewed-by: Saman Sami <samans@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647523}