Collapse file descriptors in the AHB sync
Previously, a vector of file descriptors was used for the multiple
readers that could exist at a time. This meant needing to wait of
multiple FDs.

This CL merges those FDs as they are added so we only ever need to
operate on a single writer-group FD.

Bug: 938594
Change-Id: I1e1baac2cf2508bb4a5bde31ec56d31b5f25d446
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1550012
Reviewed-by: Eric Karl <ericrk@chromium.org>
Commit-Queue: Chris Blume <cblume@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647571}