[Reland] Virtualize TransformStream
Make TransformStream delegate to a new class TransformStreamWrapper for
the creation of the readable and writable pair. This will allow adding
in a native implementation in future. TransformStream itself remains a
concrete class as after the initial creation it has no behaviour
specific to the underlying implementation.

Most of the code in transform_stream_wrapper.cc has been moved unchanged
from transform_stream.cc.

Rename TransformStreamDefaultController to
TransformStreamDefaultControllerWrapper and make it inherit from a new
abstract interface, TransformStreamDefaultControllerInterface. This will
permit plugging in an alternate implementation.

Modify TextDecoderStream and TextEncoderStream to use the new interface
name.

This is preparation for addition of the new C++ implementation described
in the design doc:

https://docs.google.com/document/d/1n0IIRmJb0R-DFc2IhhJfS2-LUwl6iKSBNaR0klr3o40/edit

This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/1545952.
The only change is remove an include of "trace_wrapper_member.h".

BUG=902633
TBR=jsbell@chromium.org

Change-Id: I0ac25fa2332016cbdc1c9e4266b3594ed93868fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549810
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647647}