OOR-CORS: Update a layout test to verify a fixed credentials issue
img-redirect-to-crossorigin-credentials.html test was added with
a TODO that says Blink has a bug, and one of test cases is modified
so that it passes with the wrong implementation.

Once OOR-CORS is enabled, this test case passes. So, I'd modify the
test case to check expected behavior by the spec, then allow this
test runs only with OOR-CORS.

Bug: 646277
Change-Id: Ifd957261c3c0f2d4218ebeb071b2340df4559838
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1550182
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Auto-Submit: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647656}