Jumbo compile fix: Too many LoggingEnabled and DebugLogging
The temporary logging methods were copied to several
files and in jumbo builds, where those files are compiled
into the same translation unit, they clashed and
compilation broke.

This patch deletes the extra copies and exposes the one
in direct_manipulation_helper_win.cpp. The name
is not great for an exposed function but since this is
just very temporary code (right?), it should be fine and
we need the build to get going and I don't want to
revert anything.

Bug: 914914

TBR=chaopeng@chromium.org,sky@chromium.org

Change-Id: I35d96a655bbaeede01d081e96426bd7b8358f295
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1552799
Reviewed-by: Daniel Bratell <bratell@opera.com>
Commit-Queue: Daniel Bratell <bratell@opera.com>
Cr-Commit-Position: refs/heads/master@{#647677}