Use a per-frame task runner at mojom::SafeBrowsing binding
This CL switches task runners at several places from per-thread to
per-frame. The purpose of this change is to make it possible to freeze
tasks for bfcache.

We found an actual case that JavaScript is executed:
./tools/perf/run_tests benchmarks.system_health_smoke_test.SystemHealthBenchmarkSmokeTest.system_health.memory_desktop/load:media:soundcloud:2018

As this is a non-standard JavaScript API execution, we adopted
kInternalDefault as the task type.

Bug: 913912
Change-Id: I512355bda609c82be436e1fa1c162a74ab8b046a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1536036
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647684}