Make PlatformNotificationServiceImpl a KeyedService.
All methods in this service need a profile anyway, so this converts
PlatformNotificationServiceImpl from being a base::Singleton to be a
KeyedService instead to make the API simpler.
This will also enable the service to listen for content settings updates
to trigger notification dismissals in future.

Bug: 905352
Change-Id: Ia16d300e34ce3c386c1f1748b7423e9f6eab2138
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1336150
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647703}