Clear directory-redundant prefs in even more codepaths
We still observe discrepancies in UMA Sync.DirectoryVsPrefsConsistency
that suggest the relevant prefs outlive the directory content (bucket
CACHE_GUID_EMPTY_IN_DIRECTORY__BIRTHDAY_EMPTY_IN_DIRECTORY), at least on
Windows (but not on Android).

We now mimic the same old behavior in a codepath that is believed to be
the last diverging logic: browser restarts for users with sync the
transport enabled without sync the feature.

R=treib

Bug: 92328
Change-Id: Id1c06c87a989cafc0bc5c7852c9bebe12f602e0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1552805
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647711}