[omnibox]: Reuse classification for document and historyUrl providers
This is the 5th refactoring CL aimed at reducing duplication and
inconsistency for classifying omnibox results.

For classification, document suggestion contents and historyUrl
suggestion descriptions currently use
a) bookmarks::CleanUpTitleForMatching,
b) AutocompleteProvider::TermMatchesInString, and
c) HistoryProvider::SpansFromTermMatch.

Document suggestion descriptions (e.g. 'Google Drive') are not
classified, and historyUrl suggestion contents have unique
classification logic which this CL does not address.

This CL changes document suggestion contents and historyUrl suggestion
descriptions to use FindTermMatches and ClassifyTermMatches defined in
autocomplete_match_classification.cc instead. These methods were created
in the previous refactoring CL's and are based on the methods (a), (b),
and (c). But unlike the previous methods, these are located adjacently
in a single file and are more generic to allow reuse by ideally all
providers.

Additionally, this CL removes the
AutocompleteProvider::TermMatchesInString method as it's no longer used.

Migrating from the previous to the new classification methods has a few
behavior changes as well:

1) The document and historyUrl suggestions now check for prefixes
matching, where if the user text is an exact prefix of the suggestion
text, further occurrences of the user text are not styled. E.g., the
user text "the" will bold only the first occurrence in the suggest text
"[the] dog ate the moon".

2) The document and historyUrl suggestions do not restrict matches when
the user text contains symbols. Previously, the user text 'the-' would
match the suggest text '[the-]dog' but not the suggest text 'the dog'.
With this CL, symbols in the user text are disregarded unless prefix
matching, and 'the-' matches '[the] dog'.

Bug: 366623
Change-Id: I48a7d06e05f5c77551033bc085a263b57840cd57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1552140
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647753}