[iOS] Fix popup shortcuts sizing.
Uses the collectionViewContentSize as the cell size for the shortcuts.
In order to do this, ShortcutsViewController now is a UICollectionVC.
To avoid issues with rotation, configureLayout: method now takes
targetSize: parameter that allows pre-calculating the insets when the
ShortcutsVC's view transitions to a new size.

Change-Id: I9a740289bb60c43e0b443217fb94798e0ac3bc92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1552846
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#647776}