DL: Move disabled test into a different section.
This test is failing because we're not processing "self" lifecycle
phases yet. Specifically, if the timing is right we don't run prepaint
on the locked elements which causes a crash later on since in the
composited layer state, we can't paint because we don't have a layer
state, which would have been set by prepaint.

This should be fixed when we process self lifecycle phases.

R=chrishtr@chromium.org

Bug: 949403, 933544, 949326
Change-Id: I0c5b897ace1ed4aec56aa752210ab8a10c93bba8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1553249
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647811}