Fix VR entry pixel check flakes on older devices
Changes the way that the VR entry tests check that the requested content
is actually being rendered. Previously, we took an actual screenshot,
which was prone to catching dialogs and such being drawn over the
screen. Instead, switch to checking the dumped frame buffer.

Bug: 947252
Change-Id: Ia6809ead471aef205de95e8d8f80b7a0f4c14f92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1552406
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647813}