Make VizCompositorThread TYPE_DEFAULT on Windows.
When Chrome is launched by a service in a console session, which seems
to how Chrome is launched from some automation frameworks, creating a
message only window for MessagePumpForUI fails in the GPU process. As a
result VizCompositorThread fails to run any tasks and nothing is ever
drawn, see https://crbug.com/942023.

The display compositor thread was originally made TYPE_UI on Windows in
https://crrev.com/c/1025954. This was to allow using a child HWND
created in the GPU process for software composited output. That code was
deleted in https://crrev.com/c/1450399 so the thread shouldn't need to
be TYPE_UI anymore.

Bug: 942023
Change-Id: I2a7dbf46a3a772b5ab1a9a20367d3e7f08116cbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549763
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647967}