[Cast dialog] Anchor dialog to browser window if it exists
On Mac, there is an issue where if the Cast dialog isn't anchored to a
view, it loses focus immediately after opening, causing itself to close.
This change [1] made the dialog not anchored when centered, even if
there is a browser window to be anchored to. This CL reverts that.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1519469

Bug: 946388
Change-Id: I9bbe4fb6e915181b33cf4613d8eb5957a9e8f522
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1554105
Reviewed-by: Bret Sepulveda <bsep@chromium.org>
Commit-Queue: Takumi Fujimoto <takumif@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647970}