Reland "Reland "Use NetworkChangeNotifier to initialize the connection type in NetworkChangeManagerClient.""
This is a reland of 5db69e59c9c38230e5b2223de0de0e42998a5763 with a flaky test fixed.

The original version of this CL added a unit test that recently started
flaking due to crrev.com/c/1538608. That CL made it so the
DnsConfigService was created on the UI thread, but would run and be torn
down on its own SequencedTaskRunner. However, there are a couple
SerialWorkers owned by the DnsConfigService that make sure they're
destroyed on the same sequence they're created on, which isn't the
correct sequence anymore. This CL makes the DnsConfigService be
created on its SequencedTaskRunner so that when it gets destroyed,
everything will happen on the correct sequence. See crrev.com/c/1336066
for related information.

Original change's description:
> Reland "Use NetworkChangeNotifier to initialize the connection type in NetworkChangeManagerClient."
>
> This is a reland of d0310b0c0221e3f44da813443ca11c9869620cb4
>
> Original change's description:
> > Use NetworkChangeNotifier to initialize the connection type in NetworkChangeManagerClient.
> >
> > Bug: 942782
> > Change-Id: Ia2971348153d54cd87a48e8045efc9d31948db8a
> > Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1529354
> > Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
> > Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
> > Cr-Commit-Position: refs/heads/master@{#641801}
>
> Bug: 942782
> Change-Id: Ie2331fe506b7b8839d9ae998bb3fd338747e0b24
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1529664
> Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
> Reviewed-by: Paul Jensen <pauljensen@chromium.org>
> Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#642597}

Bug: 942782
Change-Id: I72aec3991a421ee29dddf05334eb87e3b972484a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1548355
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#647991}