TaskManager: Show ServiceWorker tasks in TaskManager
This patch creates ServiceWorkerProvider/ServiceWorkerTask to track all
service worker status changes. During the initializing stage, the
ServiceWorkerProvider will iterate all service workers' running status
and add itself as an observer of the ServiceWorkerContext.
Then, the provider will create tasks according to the running status.
Through the invoking of the callback, the provider can create/delete the
ServiceWorkerTask dynamically.

As the type of the ServiceWorkerTask is SERVICE_WORKER, a new type of
porcess which is service_worker is added accordingly.

BUG=716609

Change-Id: Id6c153794c9e2d3d569a367d46710613f6b176a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1406147
Commit-Queue: Minggang Wang <minggang.wang@intel.com>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648022}