Send VK bounds changes even when ARC++ app is not focused.
We need to send VK bounds change even when ARC++ app is not focused
to update the size of the dummy IME view in Android.
If we don't update the size, WM can move up views in the apps not to be
covered by the dummy IME view.

Bug: b/124401771
Bug: 943370
Change-Id: I5a2f9a32f6c30d636e6aaad79bc04a36451166c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1553048
Auto-Submit: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Commit-Queue: Yuichiro Hanada <yhanada@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648105}