[Autofill] Show the name before the phone in labels.
Focusing on an email field in a contact form or an address field in an address
phone form  currently formats the disambiguating info as follows:
(774) 800-7474 • Pilar Flores

Now it will be formatted as Pilar Flores • (774) 800-7474.

It feels a bit more natural to see the name first.

Note that this change was implemented in address contact forms in this CL:
https://chromium-review.googlesource.com/c/chromium/src/+/1542559

Bug: 936168
Change-Id: I7f5e81ec3f4c7374589161fccce0f4cef41e953b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1554046
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Cr-Commit-Position: refs/heads/master@{#648158}