[AF] distinguish [un]supported fields for validation
Some field types are not supported for the validation, meaning that
the loaded rules do not have data for that level to validate the
users' data. They need to be distinguished from the supported ones.
Otherwise, the validation process would be inconsistent. For that
purpose, they'll be treated like VALID fields.

Example: For US, if state is invalid, city would be considered
as  invalid too, whereas if state is valid, it would not be validated
because it's not supported. Therefore, UNSUPPORTED should dominate
INVALID validity type.

Change-Id: I4c291488b1d016f9afc54463b742f2d6afc88111
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1531655
Commit-Queue: Mathieu Perreault <mathp@chromium.org>
Reviewed-by: Mathieu Perreault <mathp@chromium.org>
Reviewed-by: sebsg <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648197}