Clean up sandbox code in renderer_main_platform_delegate_mac.mm.
With the V2 sandbox architecture on Mac, processes are sandboxed as
part of process launching. This removes a conditional branch that
was used during the rollout of the V2 sandbox. It also removes the very
hacky DisconnectCFNotificationCenter(), which is no longer needed since
the static initializer discussed in the (deleted) comment can no longer
acquire the Mach resource.

Bug: 689306
Change-Id: I2e099f640ee2dff7ce0e91cf211bcef4e14db1ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1553799
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Auto-Submit: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648251}