[remoting][FTL] Separate callbacks in StartReceivingMessages()
This CL separates callback being passed to StartReceivingMessages() into
an on_ready callback that is called whenever the stream is ready, and
an on_closed callback whenever the stream fails to start or is closed.
Previously there is no way to get notified when the stream is first
started then get closed. This allows FtlSignalStrategy to handle stream
failure after the stream is started.

Bug: 947337
Change-Id: I65220ad98cd188a5e9cedf918b8a38beb90e8438
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1553385
Reviewed-by: Joe Downing <joedow@chromium.org>
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648294}