[Autofill] Adds ability to show new label formats.
This happens when the autofill-use-improved-label-disambiguation flag is
enabled on the following kinds of forms: address, address plus phone,
address plus email, and contact.

This does not happen on address plus contact forms.

When I delete a few fields so that the form is one of aforementioned types,
this is what appears when the name field is focused:
https://screenshot.googleplex.com/nPMuzDRpo3S.

Bug: 949957
Change-Id: I200645396a37cd09811129f3e1fdbcc2afc45791
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1554047
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648322}