[omnibox]: Clean up classifications for historyUrl "url-what-you-typed"
This is the 6th refactoring CL aimed at reducing duplication and
inconsistency for classifying omnibox results.

The historyUrl, zeroSuggest, and clipboard providers can provide
"url-what-you-typed" suggestions; e.g. the input text 'google.com' can
trigger a historyUrl suggestion with contents 'google.com' and URL
'http://google.com/'. The zeroSuggest and clipboard providers explicitly
clear the classifications of their suggestions (probably because there's
no user text associated with the suggestions). The historyUrl
"url-what-you-typed" suggestions classify their descriptions as other
historyUrl suggestions (which was cleaned up in the 5th CL
crrev.com/c/1552140). Therefore, this CL only addresses historyUrl
"url-what-you-typed" suggestions' contents' classifications.

Previously, the contents were classified using
AutocompleteMatch::ClassifyMatchInString and
AutocompleteMatch::ClassifyLocationInString. Specifically, if the user
input began with 'view-source:' and the colon was not followed by a
protocol (e.g. 'http'), then ClassifyMatchInString was used and the
first occurrence of the (slightly modified to be a valid URL) user input
was bolded. Otherwise, if the input did not begin with 'view-source:' or
the input began with 'view-source:<protocal>', then
ClassifyLocationInString was used to bold a number of characters equal
to the length of the user input, offset by the length of the prefix
(e.g. 'http://www.') in the destination URL but not the user input (e.g.
if the user typed 'google.com', the URL would be 'http://google.com/',
and the missing prefix length would be 5) minus the length of the prefix
trimmed from the URL when setting the contents text.

This CL simplifies "url-what-you-typed" classifications to simply bold
from the start of the contents text for the length of the user input. In
most scenarios, this is effectively the same as before. In the few cases
where the new behavior differs from the old, the new behavior is the
more desirable:

1) Typing 'chrome:', 'chrome:/' or 'chrome:// will suggest
'chrome://version. Previously, 'chrome://' would not bold the
suggestion, though the other inputs would bold the entire suggestion;
this was because the URL would be fixed to chrome://. With this CL, all
3 inputs bold the entire suggestion.

2) Typing 'view-source:x' or 'view-source:x/' will suggest
'view-source:x'. Previously, 'view-source:x/' would not bold the
suggestion, even though 'view-source:x', 'view-source:x/y', and 'x.com/'
would. With this CL, all those examples will bold their suggestions.

The reason these cases failed previously was because they have

Worth noting, the "url-what-you-typed" classification could be further
simplified to simply bold the entire suggestion. However, this would
produce less desirable bolding for inputs such as 'file:/' or
'view-source:/'. Both before this CL and with this CL, those inputs bold
only the typed number of '/'s, i.e., '[file:/]//' and
'[view-source:file:/]//' respectively.

Bug: 366623
Change-Id: I5df1b55db8d76408b0d3fb4d6017c6497b17a59b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1554058
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648333}