WebFrameTest does not need to override time functions
This test defined FakeTimerSetter class which was overriding the time functions
for a timeout related test. It turns out that the test does not really need it
as it was just added for a find text operation that turns out to be synchronous.

This is similar to https://crrev.com/c/1549086

Bug: 919383
Change-Id: I55d040c44cdf29d5cf1779a4d11b321f00fee18a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1554345
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Commit-Queue: Sergio Villar <svillar@igalia.com>
Cr-Commit-Position: refs/heads/master@{#648605}