[Autofill Assistant] Experiment ids extra intent.
This patch adds the ability to provide a special extra intent string
with a comma separated list of experiment ids.

It is owned by the controller and kept in the ClientContext provided to
the service when the latter is being created. Each request to the
backend will contain the experiment id string as part of the client
context field.

Bug: b/129302131
Change-Id: I4891e822694aebb96015d12d8b50803f7a3d8f49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1552835
Commit-Queue: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648616}