[Autofill Assistant] Add visibility condition to the reference.
This change moves the definition of the visibility check into the
Selector, both internally and in the protocol. This better matches the
implementation and allows looking either for an existing or a visible
element in WaitForDom, but also in preconditions.

This works well for WaitForDom and preconditions, but there are a few
actions that only make sense on visible elements, such as click and
focus. For these, the client action forces the visibility check.

This change also replaces DCHECK for empty selectors in actions with
code that returns INVALID_ACTION. This makes it easier to detect
possible script versioning issues happening in production.

Primary motivation for this change is allowing scripts to run either
visibility or existence checks from WaitForDom.

Change-Id: I252d5b4ce8250b863f531cecc2cacfb83da51a64
Bug: b/129666439
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1553549
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648620}