Throttle Pinch-Zoom Re-Rastering in OOPIFs
This CL implements pinch-zoom re-raster throttling for OOPIFs by making
OOPIF compositors aware of when the main-frame is active in a pinch
zoom gesture.

It achieves this by receiving |is_pinch_gesture_active| via the
FrameVisualProperties sent to the FrameConnectorDelegate and appending
it to the VisualProperties synced from the browser process to subframes.
Then is is plumbed into the compositor along the same route as
|external_page_scale_factor|.

An existing test for propagation of PageScaleFactor to subframes has
been enhanced to verify the value propagates from mainframe to
subframe, and from subframe to sub-subframe. Another test verifies
the flag makes it into the compositor when VisualProperties are synced
to the RenderWidget.

TBR=danakj@chromium.org
TBR=alexmos@chromium.org
TBR=kenrb@chromium.org

Bug: 924336
Change-Id: I3eeec35077291aef945791ce49743b90b55d0247
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1557879
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648648}