[Tab Capture] Simplify reuse of already captured content
Before this CL, class FrameSinkVideoCapturerImpl uses a somewhat
hard-to-follow protocol with InterprocessFramePool for resurrecting
video frames in case that the source content did not change.

This CL changes this protocol to a simpler-to-follow mechanism, which
also can reduce the cases where content is re-captured even though
it is already available in a previously captured buffer.

Change-Id: I635e3cf0dce61f704a6b613702791f03ff0c217a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1479154
Commit-Queue: Christian Fremerey <chfremer@chromium.org>
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648688}