Send none attestation if user decline to share
This CL changes the behavior when user decline to share attestation on
attestation prompt. Instead of canceling the request,
OnRegisterResponseAttestationDecided completes the request with
none attestation. Specially for internal (platform) authenticators, we
do not erase the AAGUID from authenticatorData even if the user
declines to share attestation.

Bug: 929423
Change-Id: Ib50c0ab076c483ba708c48595bb833c6ecd08c2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1552248
Commit-Queue: Suzy Li <suzyli@google.com>
Reviewed-by: Kim Paulhamus <kpaulhamus@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648820}