[css-grid] Call SetSelfNeedsLayoutForAvailableSpace where possible
Instead of SetNeedsLayout. This allows NG to do a better check on whether
relayout is needed.

large-grid improved by 26.7% in NG:
https://pinpoint-dot-chromeperf.appspot.com/job/104ea3ae0c0000

This did expose a bug in our caching logic, where we can reuse a
cached fragment even though a percentage margin has changed. This is fine,
except that we did not recalculate the margin for legacy layout; this
patch also fixes that.

Change-Id: I41054fa6014ae7098eb07ca2b63bd7ede14c33a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1554113
Reviewed-by: Manuel Rego <rego@igalia.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Commit-Queue: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648882}