Swap GPU Mac standard and optional OS
Swaps the standard and optional Mac versions for the chromium.gpu.fyi
Mac Release 10.14 bots. The versions were configured assuming that
the "optional" dimension would be tried first, falling back to the
standard if the timeout was met waiting for the optional dimensions.
However, it looks like the opposite is true, so swap the versions to
get the correct behavior.

Bug: 950098
Change-Id: I9de396274165d7767d18e47973ff1eaaa3446dea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1557865
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#648889}