[signin] Add a SCHEDULED state for AccountReconcilor
When the reconcilor does multiple cycles, we need to be able to wait
until all the cycles are finished.
To enable this, this CL adds a SCHEDULED reconcilor state: when the
reconcilor does multiple cycles, it will go to the SCHEDULED state
without going through the OK state.
For example, if there are two reconcilor cycles, the state will be:
RUNNING -> SCHEDULED -> RUNNING -> OK

Bug: 948140
Change-Id: I145c30eb38fdc06d72e88ddd4ecbc568440fa62c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1550430
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649055}