Add OfflinePageModel::GetPagesWithCriteria
This CL introduces the concept of PageCriteria which describes some selection
criteria for an offline page. This criteria is used to query for pages in
GetPagesTask, but could be used for other tasks like deleting pages later. This
criteria can be evaluate in C++, and additionally used to create sql for
filtering a SQLite query.

This CL also uses the new method to fix a bug in feed where we would sometimes
attempt to open a tab-bound page, and end up failing.

Because GetPagesWithCriteria is very general purpose, it replaced the other
page query methods.

In terms of binary size and efficiency, I think this is a good trade-off. We're
replacing several smaller queries with a single larger one, shaving about 4KB
off the binary size. Depending on the query, it may be faster or slower --
querying for multiple client_id or namespaces is improved.

Bug: 949162

Change-Id: I58eb97014d8403ed248c1e38635cc0665ca5c18e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1540063
Commit-Queue: Dan H <harringtond@google.com>
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649122}