Use TargetBounds to compute RestoreBounds
Toggle maximize should use the target bounds so that it can restore
the correct orginal bounds. (although this is probably less important
because it's near impossible for a user to trigger this while animating)

WindowResizer shoud use just bounds(), because that's where a use captured
the window and started dragging.

Bug: None
Change-Id: I5e25805de92cc2afc616fda6068ec6ac50ecf700
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1559350
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649128}