CrOS enrollment: announce enrollment in progress
This informs the user that enrollment is on-going. The 'alert' role is
probably not perfect for this use case but it's the best I could find,
and it's used in nearby code for similar scenarios.

Testing by running through the remora enrollment flow and observing that
"Enrolling device..." is read out.

BUG=b:80126209

Change-Id: Iee7388dae29458412cb44f9d3578cfeee2d234cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1559182
Auto-Submit: Felix Ekblom <felixe@chromium.org>
Commit-Queue: Alexander Alekseev <alemate@chromium.org>
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649228}