arc: Limit locale/preferredLanguages for autotest.
In case arc-disable-locale-sync is used this limits ARC container to
start with 'en-US' as a locale and 'en-US,en' as preferred languages.
This stabilizes tests that expects UI resources be fixed in English.
There is an alternative solution that could be implemented when needed
to pass required combination as a switch. However this is not required
at this time once we don't have such auto-tests.

TEST=Set locale for local test account to 'ru' and confirmed that
cheets_AppCompatTest.clamshell.com.google.android.apps.books is failing
similar as described in bug. Apply this patch and test passes.
BUG=b:130142125

Change-Id: I2219207d2d75c00c6ee04590e430a929389c7319
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1559463
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Yury Khmel <khmel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649236}