[NTP] Improve middle-slot promo metrics
Fix NewTabPage.Promos.RequestLatency to properly capture if a request
fails or succeeds and differentiate between succeeding with and without
a promo present.

Add NewTabPage.Promos.ShownTime to log the time from when an NTP is
opened to when the promo is inserted into the page.

Add NewTabPage.Promos.LinkClicked to log clicks on links embedded in
middle-slot promos.

Bug: 937505, 934604
Change-Id: I5f4fdb464929485ba38a7a134de4f4f5112e7670
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1504221
Reviewed-by: Ramya Nagarajan <ramyan@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Kyle Milka <kmilka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649286}