[ios/web] Always keep transient views on top.
In some cases, an transientContentView may be displayed when the
CRWWebControllerContainerView has no window.  When the
CRWWebControllerContainerView is put back in a window and the main
webViewContentView is restored, it will obscure the transientContentView.

Instead, always keep the transientContentView on top.

Bug: 949204
Change-Id: Ifa992620958ac3b7cfeb85d4451e2d13beee7810
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1559894
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Auto-Submit: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649297}