BlinkMemoryMgt: Annotate Blink's renderer/core with the macros of Allocator
This CL finishes annotating the rest of the classes of //third_party/blink/renderer/core
with the macros of Allocator.

After some analyzes, it was concluded that for the case of MediaQuery and Entry classes,
it makes sense to use USING_FAST_MALLOC because it has been used as a member variable
with new() or smart pointers.

In LayoutObjectCounter and CompositingInputsRoot class, they make sense to use DISALLOW_NEW
because it has been used as a member variable with a reference type.

Lastly, the rest of the classes this CL annotates make sense to use STACK_ALLOCATED.

Bug: 919389
Change-Id: I6c81513e12b2c100170a8a7d02a0e3ea5eda8b40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1559417
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Gyuyoung Kim <gyuyoung@igalia.com>
Cr-Commit-Position: refs/heads/master@{#649405}