Give LoadIconFromExtension a fallback icon
This code path is contingent on the App Service being enabled.

If, for whatever reason, the underyling extension system can't find an
icon, we now default to the IDR_APP_DEFAULT_ICON instead of an empty
icon. This matches the existing behavior with the App Service disabled.

BUG=826982

Change-Id: I1f79a4a02513871ba955e6bdbd1a91107b230b45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1558568
Commit-Queue: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649428}