Fixed crash in android due to NullPointerException
In the new UI in Chrome Android, one can toggle the permissions even
when they are blocked for Chrome or for all apps in Android.
Geolocation is a special case where the permission can be both modified
per app and per device (for all apps).

In one scenario (described + video in the bug), switching the permission
to blocked causes chrome to crash, due to attempting to setTitle() on a
null object. This happens as when you toggle the binary toggle for
permission category exceptions the name changes from "Allowed" to
"Exceptions" and back.

Bug: 949951
Change-Id: If8e7d2b58a977b42c10e5b18e064e79e84d76a2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1560194
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Kamila Hasanbega <hkamila@google.com>
Cr-Commit-Position: refs/heads/master@{#649503}