ExpandToEnclosingUnit (TextUnit_Page, TextUnit_Document)
This adds an implementation of ITextRangeProvider::ExpandToEnclosingUnit
for TextUnit_Page and TextUnit_Document.

Behavior for TextUnit_Page is identical to TextUnit_Document, as web
pages are not paginated.

TextUnit_Document behavior is to always expand to the beginning and end
of the top-level document. iframes will jump out of the current iframe
and will also expand to the beginning and end of the top-most document.
There currently is buggy behavior for the iframe case, and a follow up
task will address this issue.

Bug: 928948
Change-Id: I9790b2200d61b3449d905488596b7c731ba8c61b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1559762
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#649537}