Use shader for rounded corner when animating
When the unified system tray is animating it uses a separate layer to
to perform blur and animation. This layer does not have a mask applied
to it for performance bottlenecks and thus does not have any rounded
corner clipping. This patch modifies this and applies the shader rounded
corner on the layer.

Bug: 898329
Change-Id: I9fed724e4659960710c3daee5e33ec92e2223fb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1562624
Reviewed-by: Tim Song <tengs@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Malay Keshav <malaykeshav@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649695}