[remoting] Remove oauth_helper.h
It appears this code is no longer used anywhere. It contained a
misleading method: GetOauthScope() which was described as getting the
scopes for the host's refresh token. It simply returned a hard-coded
list (which included the 'googletalk' scope), rather than actually
querying the token's scopes. Now that we are migrating away from
Talk/XMPP, it doesn't make sense to keep this code around.

Change-Id: Ia9041f45f0cfffd56ccfa9af0cab11e72e69dc56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1560468
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649871}