Run WebGPU tests only when there is a supported backend.
Otherwise we might be using an unsupported backend or run tests against
the Null backend that doesn't actually do anything.

The WebGPUDecoder didn't have a separate initialization step that
returned a success value but we need to detect when we fail to find a
supported Dawn backend. WebGPUDecoder::Initialize is added for this.

WebGPU tests are also modified to be skipped when WebGPU initialization
failed.

BUG=chromium:852089

Change-Id: I0610047faf917fb3f53fa8e8d45a0c60ef814897
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1564075
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649945}