chrome/browser/ui/webauthn: can't select from just one resident credential.
When there's only a single resident credential on an authenticator it
sadly does not return the user information. Therefore we cannot show a
UI for the user to confirm their account selection.

On the plus side, this does solve the issue of showing double-UI when
Windows does the account selection.

Change-Id: I7000af96085ce852b7176fcfa06993ac2c6b6b55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1558711
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#649967}