[Autofill] Adds icons to profile suggestions for an experiment.
Also, adds the group bitmask to LabelFormatters. This is necessary for adding
icons. The bitmask indicates to what kind of form a LabelFormatter's labels
correspond.

If the labels correspond to a form with an address, then one icon should be
shown. If the labels correspond to a form without an address, then a different
icon should be shown.

Screenshots: https://screenshot.googleplex.com/weaC2oSksZe

Bug: 951174
Change-Id: Ief7952243d8df3a1a4b67bd1ccd629002000044f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1560464
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#649971}