Increase height of bottom toolbar in immersive mode
Increase the height of the bottom toolbar in immersive mode, and
restructures the bottom toolbar XML so that the browsing mode and
tab switcher mode toolbars inherit their height from their common
parent container.

The restructuring is required because the background color for
the browsing mode and tab switcher mode toolbars is set on
the browsing mode and tab switcher mode toolbars (not the parent
container) and controlled by each sub-component individually. In
order for the background color to fill the entire parent container,
the browsing/tab switcher mode toolbars must get their height
from the parent container.

BUG=949819

Change-Id: If8099848ab487abba8500b9ed1709a24bc8dc312
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1554190
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Theresa <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650002}