[iOS] Adding sync error when sign-in was previously interrupted
When sign-in is interrupted in the Advanced sign-in settings, an infobar
needs to be displayed in a new tab page.
The infobar opens the Google services settings. The Google services
settings shows a sync error to invite the user to review the sync
settings.

Screenshots:
https://drive.google.com/open?id=1BTpeBg-uhFL7rmZvaYGbewHANSVVqDTE
https://drive.google.com/open?id=1bur57se74fzCkrVsZZx7hzg0TO7jfMe8

Bug: 951162, 951199
Change-Id: Iaa7c3959c3b26b4385fcc102e4086e5778bbf778
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1560152
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650091}