Ensure accessibility extensions are unloaded on sign in profile
This change ensures profile switches (AccessibilityExtensionLoader::SetProfile) triggers the unload of the extension on the previous profile.

In practice, this means that there should only ever be one profile in which an accessibility extension ever gets loaded into assuming SetProfile is appropriately called. Furthermore, if SetProfile is called with the current |profile_|, this would amount to a reload.

Bug: 951990
Test: in ChromeVox, load on login screen and when signed in. Ensure braille does not emit from login screen ChromeVox (as reported by users), and that there is only one ChromeVox extension process.

Change-Id: I1d702a78583748e8adcffa19dd39eda151a6dd94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1564567
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650114}