[remoting][FTL] Properly handle send message failure in FtlSignalStrategy
Previously when FtlSignalStrategy fails to send out a message, it will
simply disconnect the streaming channel. This can happen when the
client tries to connect to a peer that doesn't exist, and disconnecting
the channel is unnecessary in this case.

This CL makes FtlSignalingStrategy synthesize an error IQ stanza if it
fails to send a message, which will be passed to JingleSession and
translated into a PEER_IS_OFFLINE error.

Bug: 947337
Change-Id: I3deb74148a515633a55812043f9664c312bc6a30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1563650
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650117}