Populate page ID and session key in data reduction proxy pingback
Page ID and session key are not sent in the pingback when network-service is
enabled. This CL moves the page ID generation to earlier in the navigation
and populates it in NavigationUIData, which is then used by the PLM
observers. This new logic is gated by a feature and enabled by default.

Bug: 951913
Change-Id: I8b1e8ff672742ebef661cc3f948d2b749a06ca87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1565130
Commit-Queue: rajendrant <rajendrant@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650157}