GWP-ASan: Add stack trace test
Add an end-to-end test (collection to decoding) that verifies that stack
traces are correctly recorded. This is not possible to do in the
crash handler unit tests because they are multi-process and ASLR might
change the addresses we need to confirm the expected values.

This change is in preparation for an upcoming change that will update
how stack traces are stored.

Bug: 951410
Change-Id: I38a5def08e6ff8572fc4acf0f16931a779aefdc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1564803
Commit-Queue: Vitaly Buka <vitalybuka@chromium.org>
Auto-Submit: Vlad Tsyrklevich <vtsyrklevich@chromium.org>
Reviewed-by: Vitaly Buka <vitalybuka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650189}