[Payment Handler][Android] Close PR sheet on PH dismissal.
Before this patch, tapping the Android back button in the navbar or the
"X" button in the CCT toolbar dismissed the payment handler activity,
but did not notify the service worker, which caused the Payment Request
overlay to continue blocking interactions with the web page. A second
tap on the Android back button finally closed the Payment Request
overlay as well. This behavior was caused by null tab in
PaymentHandlerActivity.handleFinishAndClose(). The
PaymentHandlerActivity used the tab to get the web contents to determine
which service worker to notify of the window closing.

This patch saves a references to the web contents at the time of opening
the payment handler and uses that reference when handleFinishAndClose()
is called.

After this patch, tapping the Android back button or the "X" button will
notify the service worker, so the Payment Request overlay no longer
blocks user interactions.

Manual test:
1) Open https://rsolomakhin.github.io/pr/skip-ui/ (or /pr/bob/).
2) Tap "Buy" on the page.
3) Tap the Android back button.

Bug: 951427
Change-Id: I4754f4487a9f371b298b1d24aaa9155f43f870c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1562225
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Pavel Shmakov <pshmakov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650315}