[Sync::USS] Refactor re-encryption conflicts in ClientTagBasedProcessor
Before this CL:
The processor would assume if the commit data are missing, it must be
a re-encryption request. However, it could also be the unsynced
metadata loaded on the startup without data.

After this CL:
This case is handled separately and added as a TODO because this
requires substantial changes in the processor.

Bug: 951752
Change-Id: I8fef3cc8ce33c958f8801e382b5d108a43823421
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1564080
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650324}