H2 session reuse refactor 2: Move observer logic to HttpStreamFactoryJob.
Previously, the HttpStreamFactoryJob would tell the
HttpStreamFactoryJobController to have the SpdySessionPool to inform
the HttpStreamRequest when an appropriate SpdySession was created.
When that happened, the HttpStreamRequest would then pass the resulting
stream to the JobController, which would then destroy the original job.

This CL simplifies that by having the Job watch the SpdySessionPool
directly, though it still informs the JobController when it gets
the SpdySession, and the JobController then immediately deletes the Job.

Bug: 912727
Change-Id: I9585f037e2028159a24064111d670faddfe2bdba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1551562
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Bence Béky <bnc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650406}