SkiaRenderer: Make DrawQuadParams mutable/hold tex coords
This consolidates some signatures, allows code to update
params->opacity when it figures out how to apply it
as a filter or paint instead.

This also adds texture coordinates to the DrawQuadParams,
which will be used in a subsequent CL to emulate
independent axis clamping of texture coordinates. It's a
lot simpler to just modify the parameters in-place vs.
making a copy to override, or having functions take even
more arguments.

Change-Id: I8e40de7d047926ae86fb3e428130ca108930ea66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1557428
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: weiliangc <weiliangc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650824}