Implement autoincrement storage keys in FakeModelTypeSyncBridge
For the case where a bridge cannot infer a storage key from specifics,
it's almost by definition the case that the storage key is dynamic and
is an autoincrement-like attribute of the insertion.

This patches implements this behavior in FakeModelTypeSyncBridge, used
to test the central ClientTagBasedModelTypeProcessor.

Bug: 948217
Change-Id: I39e7a7c7de1b23fc95bd5acc394501079fe51fc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1566345
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650873}