Make tweaks to EditableCombobox for the password bubble use case.
Add option to not show the menu on the next focus. This is useful since
the EditableCombobox is focused right away when the bubble opens, and
without this it brings up the menu and potentially hides the password
field.

Expose SelectRange from underlying |textfield_|.

Rename OnMenuClosed into CloseMenu.

Bug: 947187
Change-Id: I2f21ec5d916b30d9810694c5a9cf79be1758eb46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1567448
Reviewed-by: Michael Wasserman <msw@chromium.org>
Commit-Queue: Edin Kadric <edinkadric@google.com>
Cr-Commit-Position: refs/heads/master@{#650887}