Remove suiteSetup in polymer test base
The import of polymer is only needed for a few tests that are testing
behaviors that don't import Polymer already. Import Polymer in the
pages being tested so we can remove this code and more easily override
suiteSetup in future in order to wait for the HTML imports polyfill.

Bug: 925517
Change-Id: I18e7982bf46952235430da594cab2fe636a8a6f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1565782
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#650959}