media/gpu/v4l2vda: make sure InitializeTask() signals completion
InitializeTask() is supposed to inform the child thread that it may
proceed by signaling a WaitableEvent. However in some rare situation the
WaitableEvent would never be signaled. This was hard to catch due to
some sloppy programming from my end.

Fix this by managing the WaitableEvent in an outer method so we don't
have to signal it at every return point.

BUG=chromium:951426
TEST=VDA unittest passes on Hana.

Change-Id: Icdfbbc3ae6b30e93087218b467a046d820adf791
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1567552
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651113}