fix: close notifications for revoked domains
This closes notifications when a user revokes permissions to show
notifications for a domain, or if the service worker gets unregistered.

Note: This will only handle persistent notifications, as we have their
ids in the notification database. We can not do the same for non-
persistent notifications, as we do not have a reliable way of getting
them from all platforms.
This will not change the behavior on Android as we're not tracking site
permissions and Android already closes notifications when revoking
permissions for a site.

Bug: 905352
Change-Id: If05967e7adb9e6c765c3cd92d874fa3a08c8a803
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1342817
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651204}