[Sync::USS] Simplify the logic for re-encrypting entities in CTBMTP
This CL simplifies the logic that handles the re-encryption of received
entities that have out-of-date encryption. Due to recent changes
in EntityData being a std::unique_ptr<> it became tricky to reason
about when the data is available. In order to avoid confusing, and
potential bugs in case of overlooking some corner case, we should
always read the data from the bridge. This may incur an extra IO
but this should be OK because this scneario is already very rare.

Before this CL: commit data were cached from the incoming updates.

After this CL: commit data are always read from the bridge when required

Change-Id: I4af16657f91004bc00013cd68dbd349e060c2d39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1567773
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651258}