[ MimeHandlerView ] No FrameContainer for Navigations
When a frame is navigated to a MIME type handled by MimeHandlerView, an HTML
document is injected by MimeHandlerViewAttachHelper which contains an <iframe>
that is later on used to attach the GuestView. For this type of MHV, the
container class on the renderer side does not send resource request (already
handled through navigation) nor does it expose a ScriptableObject as there are
no plugin elements involved. The only role of the container is to implement the
BeforeUnloadControl API.

This CL modifies the procedure such that:
 1- MimeHandlerViewContainerManager will implement the BUC for the page.
 2- No MHVFC are created.
 3- MimeHandlerViewGuest is created on the browser side after the BUC is
    installed. This means the ExtensionsGuestViewMessageFilter is no longer
    involved.

Bug: 659750, 896679
Change-Id: Ia6d85afa4a5a39f7c3f2a38de733fdab6565ca44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1567100
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Commit-Queue: Ehsan Karamad <ekaramad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651288}