Deflake pixel-test harness for web tests.
1) LayerTreeHost would cause a full composite on an animation/lifecycle
update when a presentation request had been made, stealing control over
when to do the composite away from WebWidgetTestProxy. This is resolved.

2) TestFinished() would:
a) schedule a full composite (post a task) and
b) insert a presentation callback and
c) cause full viewport damage then
d) request browser to grab pixels which would
e) tell renderer to sync composite, consuming b) and c)
f) run the composite task from a)

However since f) is now after the test is done, it would end up
capturing the test tearing down if there was any damage in the
compositor.
A previous change made all composites fully damage which caused f) to
show up as causing flake.

So instead we remove a) b) and c) when doing compositor pixel tests,
and just have e) insert damage and do the composite as requested.
So, now the order of operations for a compositor readback is the
TestFinished() will:
a) request browser to grab pixels which will
b) tell renderer to sync composite which will
c) insert damage and o the composite

So there is no extra compositing task present to cause flake! Thus
we can remove the RequestPresentationForPixelDump() path in
WebWidgetTestProxy entirely.

As part of 2) we split off renderer-side pixel-grabbing things from
end-of-test-pixel-grabbing-via-the-browser things more explicitly in
the renderer test harness code. We leave a TODO about going further
and splitting these apart more explicitly instead of thru the
testRunner state machine.

R=avi@chromium.org, vmpstr@chromium.org

Bug: 951933, 952399
Change-Id: I7b443bf2d795a0dc642d2d861b8d7bd0674e43a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1565729
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651308}