BGPT: Add a fast-path for simple value transform changes.
This patch adds a fast-path for updating composited transform changes
if those changes are "simple", meaning that both the before and after
transforms do not change axis alignment and are 2d transforms.

In those conditions are met, we update the transform directly on
the cc node, instead of requiring a PAC update.

Change-Id: I0ef9c89b64c46cc8e5e89d015bda3eb10d6869ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1565467
Commit-Queue: vmpstr <vmpstr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651338}