[TTS] Support Translations and card tags.
Adds support for Translations and card tags to be returned from
the Resolve request.  Now we pass a "Translation Integration" code
to the server to return Translate results in the resolve request.
This level of integration also supports a code for the primary card
tag to be returned so we know what we're showing, and that's piped
all the way through to Java and logged in UMA.

BUG=952401

Change-Id: I89eda77e131915e7f2bb1187ab2c26b17d558de0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1566362
Reviewed-by: Theresa <twellington@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Auto-Submit: Donn Denman <donnd@chromium.org>
Commit-Queue: Donn Denman <donnd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651411}