device/fido: truncate pinAuth parameters at 16 bytes
The SetPin and ChangePin subcommands to the authenticatorClientPin
command previously sent 32-byte pinAuth parameters, when they should
have truncated the output of the HMAC operation for deriving the
parameter at 16 bytes. Some keys were tolerant of this, but others are
not, and the spec clearly says to truncate.

Also fix VirtualCtap2Device's expectation about the correct length for
pinAuth.

Bug: 870892
Change-Id: I32199a0014da60434145b82c8f1a689c754091d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1568791
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651420}