Create SimpleKeyMap, a map from BrowserContexts to SimpleFactoryKeys.
Accessing this map allows BrowserContextKeyedServiceFactories
(BCKSFs) to access SimpleKeyedServiceFactories (SKSFs) by converting
their key (BrowserContext) into a SimpleFactoryKey.

This conversion was already possible in //chrome by casting
BrowserContext to a Profile and calling GetSimpleFactoryKey(), but it
was not possible in //components.

Code paths starting from //content or //components may not go through
an embedder that is able to produce a SimpleFactoryKeyedService
(or the SimpleFactoryKey itself) from the BrowserContext.

For example:
    BrowserContext::GetVideoDecodePerfHistory() -->
    VideoDecodeStatsDBImpl::Create(...) -->
    leveldb_proto::ProtoDatabaseProvider::CreateUniqueDB
        <DecodeStatsProto>()

These code paths can now use SimpleKeyMap to get a SimpleFactoryKey
for the BrowserContext.

Change-Id: I6b0833e316b17085d9300ba00c770f559259b2ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1559019
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Mohamed Heikal <mheikal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651438}