ServiceWorker: Write script data directly into target vector.
Previously when service worker script text and code cache was streamed
from the browser each received chunk was stored in a separate vector.
When the worker thread was launched these chunks were then processed to
produce the final data necessary to start v8.  In the case of cached
metadata this essentially involved combining the chunks into a single
vector.  This essentially meant copying the data at least twice in the
renderer process.

This CL removes one copy by pre-sizing a vector to the total expected
size and directly filling it as each chunk is received over the pipe.

Bug: 946676
Change-Id: I319bf72152de045bee7d1f6a13ebcd99e3ec0932
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1560770
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651591}