Add handling of cancelled PluginVm image downloads
PluginVm image downloads that are tracked by download service when it is
initialized are cancelled once client receives information about them.
Before this CL information that these downloads were cancelled was
propagated to PluginVmImageManager which we do not want. So this Cl
stops this information for being propagated.

Bug: 951338
Change-Id: I40b1a7251241cc99e7886dbb0437d6671e461ca1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1564126
Reviewed-by: Alex Oldemeier <aoldemeier@chromium.org>
Commit-Queue: Olya Kalitova <okalitova@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651681}