[Autofill Assistant] Delete controller while a script is running.
With this change, the AA controller is deleted immediately, even when a
script is running, no matter what action is it. This simplifies the
shutdown procedure and makes it overall more robust.

This solves an immediate issue with wait for navigation preventing
shutdown.

There used to be ownership issues that prevented it from working
properly. These issues seems to have been fixed.

Bug: b/128300038
Change-Id: Ieeeccc9bcf3cc996072ef87b97b9fe741842adb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1569116
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#651743}