Account for inverted selection when determining if the user has backspaced.
The omnibox disables autocompletion if the user has backspaced - detected
by checking the current text size against the previous text size. An exception
to this is trailing selected text, which is not considered when comparing the size.

Previously only the selection_end field was used to determine the size of the selection,
which is incorrect in the case of an inverted selection (where the selection start is
greater than the end). This fixes that.

Bug: 951375
Change-Id: Ie245db06570e3b6df8d41a7645d60429973eecf6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1570751
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Will Cassella <cassew@google.com>
Cr-Commit-Position: refs/heads/master@{#651814}