Remove DisableFontSmoothing which is causing flakiness on unittests
The class DisableFontSmoothing is modifying a system wide setting
that affect rendering. Unfortunately, that setting is also affecting
other unittests.

The call to SystemParametersInfo is modifying a system-wide setting:
see: https://docs.microsoft.com/en-us/windows/desktop/api/winuser/nf-winuser-systemparametersinfoa

It seems to have first appeared in
  https://codereview.chromium.org/100255
to cater for the test -- RenderViewTest.PrintLayoutTest.

Chrome tests should not modify the global state of the system.
Unittests should be able to handle different configurations.

Every tests are passing correctly. If some tests are failing or disabled
because of this class, we should fix it by detecting if fonts smoothing
is activated or not.

Bug: 953293, 759870
Change-Id: Ib1899f00ba18ae530da2a078933ff8b0a0983f45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1569969
Reviewed-by: Wei Li <weili@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651835}