Track external pinch gesture active state separately.
This CL corrects a logic error in the pinch re-raster throttling logic
for OOPIFs. Prior to this CL, it's possible for pinch_gesture_active_
on the main-frame's compositor to be incorrectly overwritten to |false|
while a gesture is still active. This CL corrects that by tracking
local and internal pinch gesture activity separately, and using a
logical-OR of the two to control the re-raster logic.

Bug: 924336, 951861, 951924
Change-Id: I4ef7a699038cd523abc671b23bb0a94f4b28a849
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1572099
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651895}