wallpaper: Cache a downsampled image on first paint.
This will make entering overview wallpaper animation more performant as
that paints the wallpaper multiple times a second.

I tested this by making the wallpaper animate while the overview windows
are animating and using the AnimationSmoothness uma stats. It's not the
best way and the results have some noise sometimes but I think it's
good enought to verify that this is an improvement.

Test: manual
Bug: 953890
Change-Id: If472e417e667840707ec55a79178d569cda4ca40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1573085
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#651994}