Update profile menu edges for overflow.
Top corners of profile menus are sharper than bottom corners, probably
due to an overflow of the content on the edges. A 2 pixel margin is
added to the top (which was already added to the bottom) to over come
the issue.

See https://screenshot.googleplex.com/x66vzNYUO7z for screenshots of
current state.

Bug: 950854
Change-Id: I4398d0483243fde8bdac71fa3986da9bb6a07da8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1570169
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652035}