Ensure not to save PIP or MINIMIZED to |kPreMinimizedShowStateKey|
If PIP is triggered by minimize, the window state changes, for
example, from maximized, through minimized (1), and to PIP. If you
dismiss the PIP window it transitions to minimized (2).
If you unminimize the window at this point, we need to ensure not
to save PIP or MINIMIZED to |kPreMinimizedShowStateKey|

when the window gets unminimized in the above scenario.

Bug: 949006
Test: Confirmed Chrome didn't send the minimize state to Android
Change-Id: I00085444ae1cdab41c66aef507251ffc4d706466
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1567575
Commit-Queue: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652086}