desktop-pwas: Replace stop_if_window_opened with wait_for_windows_closed
If wait_for_windows_closed is true, the task will be taken out of the
queue until all windows for the placeholder app are closed. Once
all windows for the app are closed, the task will be added to the
beginning of the queue.

Replaces stop_if_window_opened because that option is no longer needed.

Bug: 844278
Change-Id: I76725d353a5d195ce00e2fed2167d14e1e0d13c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1572760
Commit-Queue: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652121}