vr/android: Disable SurfaceControl when in VR mode.
The android surface used by the display compositor when in VR mode is
backed by SurfaceTexture, which is incompatible for use with
SurfaceControl. Disable it when in VR mode.

Fundamentally, SurfaceControl can only be used when the consumer end
of the display's frame updates is SurfaceFlinger. With SurfaceTexture,
the consumer end can be any arbitrary process which created the
SurfaceTexture.

TBR=tsepez@chromium.org
R=piman@chromium.org

Bug: 953953
Change-Id: Ie132d143aeffcf1bb022f6c691a0738515b19a1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1573164
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Auto-Submit: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652288}