[Cronet] Enable metrics reporting when there are listeners.
A RequestFinishedListener can be set on the UrlRequestParam, and any number of
listeners can be attached to the Engine.

The common CronetURLRequest only needs to report metrics to the native
UrlRequest class if that UrlRequest has a listener, or if its engine has
one or more listeners.

There are many code paths by which metrics can be collected (success,
fail, cancel, upload failure), so instead of just adding a few new
testcases to url_request_test.cc, cross-product the test
parameterization with these cases: 1) no RequestFinishedInfoListener 2)
listener on UrlRequest 3) listener on engine.

This CL should land before any CL that modifies OnMetricsCollected() as
it ensures that OnMetricsCollected() actually runs in a test and at the
very least doesn't crash.

Bug: 879208
Change-Id: Ib7b8a252102e2ff86310c42e2db0a339b39efaa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1546517
Commit-Queue: Caleb Raitto <caraitto@chromium.org>
Reviewed-by: Misha Efimov <mef@chromium.org>
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652346}