MPRIS: Add title/artist/album metadata
This CL adds title, artist, and album names to MPRIS metadata. These
were left out of crrev.com/c/1562628 since there was a failing perf
unit test. However, it turns out the issue was just that the test
happened to use non-ASCII characters and this hit a DCHECK as we were
converting to ASCII. It turns out this conversion is unnecessary, as
MPRIS can handle wide strings. So this CL switches MPRIS to intake
base::string16s for the metadata so we can send the media session
metadata strings directly.

Bug: 951584, 952410
Change-Id: I3723554369d47041ac48e39a92bbd91c484eda21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1566224
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652356}