Skip fuzzing accessibility trees with too many nodes.
Clusterfuzz tries running this code with random inputs, and
we got a bug that one input made it time out - but the issue
was that the input was super large. It's not surprising that
it would time out if you give it enough nodes.

I set the limit at 500,000 nodes for release mode, which is larger
than any web page I've ever seen. For debug mode I set it at 50,000
nodes because larger than that can time out.

Bug: 944999
Change-Id: I50b57153876a65f92c5a88677bdffe64b8c7ae7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1574622
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652364}