Download home : Show/hide prefetch tab correctly
This CL fixes the prefetch tab visibility on download home.
The prefetch tab is visible only when prefetching is enabled
which requires the feature and the settings to be enabled.
Also adds an observer to the user setting pref, so that the
tab shows/hides as soon as the setting is toggled.

Bug: 951821
Change-Id: I85bc2d0f346d01b1a0f7f9d8b42430f3614ed373
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1564902
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652425}