media: Ignore EOI and beyond in JpegParser for size computation.
This CL makes JpegParser not include the EOI marker and any data after
that when computing the |data_size| for a JPEG. The assumption is that
the |data_size| should only include scan data: the documentation for the
field in jpeg_parser.h states that |data_size| is "the size of
compressed data of the first image." The |image_size| field is also
affected: it includes everything up to (and including) the EOI marker
and ignores anything after that.

This is done so that camera captures that come with trailing zeros don't
get the size misreported to the hardware acceleration driver.

Test: added a simple test to check the behavior.
Bug: b:129728310
Change-Id: I68557f38527256e5f6d1116866ae56964ec40da8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1554044
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652434}