Load urls in a window on the current desktop, on Windows 10.
When Chrome is launched with a URL on the command line, and
Chrome is already running, it looks for a browser window to load
the URL in. This CL makes it ignore browser windows that are
not on the current desktop, on Windows 10 (or higher).

Bug: 757853
Change-Id: Ib251619cbcef18809b886b96e21034507ea14654
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1538834
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652693}