Deduplicate `MockCursorEvdev` between unit tests
The copies in `EventConverterEvdevImplTest` and
`InputInjectorEvdevTest` were exact duplicates, and I'm working on a CL
that also needs it.

Also renamed it to `FakeCursorDelegateEvdev`, since it doesn't log calls
to its methods and so isn't really a mock.

No behaviour changes should be caused by this CL.

Bug: none
Test: run `EventConverterEvdevImplTest` and `InputInjectorEvdevTest`.

Change-Id: Iad54935e675bb2b253dcf6aa5c2bb0d9d9cae1be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1559207
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652718}