Fix status bubble text low contrast with some themes
Status bubble text was drawn with certain transparency. When it was
changed to be drawn on a label, the color would be automatically
computed based on foreground and background colors, which makes the text
color even lighter with some themes. We fix that by disable auto color
readability on that label.

BUG=952061

Change-Id: I25ec54aa82c4d62e7fa67f146f0e7c9e39511eac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1565504
Commit-Queue: Wei Li <weili@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652879}