SocketParams refactor 2: Split up CreateSocketParamsAndGetGroupId().
One method creates the group ID, one takes raw parameters and returns
a SocketParams object, and one takes a group ID (and other values) and
calls the second method.

This is part of an effort to replace the current SocketParams callback
with an object (And eventually remove all fields of the object, removing
SocketParams entirely).

Bug: 533571
Change-Id: I518d6a2e62f160879a0094aa4772359c56815e47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1561845
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/master@{#652970}