Migrate from RVO::Navigate to RFO::ReadyToCommitNavigation
These two callbacks are called at approximately the same time,
and we can use the more common one (ReadyToCommitNavigation) instead
of an outlier (Navigate).

This leaves a single client of RVO::Navigate, which can be removed
in a follow up. ChromeRenderViewObserver can be removed as well.

Bug: none
Change-Id: I0e2a046f11f451c0a1e952a44a0e84236bf93c53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1576012
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653175}