Always set a stable_id on mask isolation effect nodes.
Even if we omit a mask layer, the content isolation node
needs a stable_id if it gets a render surface.

Also, the stable_id needs to stay stable, because the
id is used to track damage and re-use render surfaces.
For this reason, we need to continue storing a list of
SynthesizedClips, even if they don't have layers.

Bug: 953933

Change-Id: Ied9b508072ee82f592561517bd8c067c927eabe1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1576231
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653262}