Fixed FileSystemURLLoaderFactory tests.
Fixed the browser tests to use the same task runners for
both the FileSystemContext and FileSystemURLLoaderFactory.

Additionally three more cleanup tasks were performed:

1. StartRequest was merged into TestLoadHelper as that was
   the only caller.
2. SetUpAutoMountContext returns the mount point instead of
   accepting an "out" parameter.
3. SetFileUpAutoMountContext renamed to SetUpFileAutoMountContext.

# Background

The FileSystemContext (used by the FileSystemURLLoaderFactory)
uses IO & file task runners provided to the constructor.
Previously these tests would provide the current thread's
task runner for both params. The FileSystemURLLoaderFactory is
hard-coded to do some of it's setup on the IO sequence because,
on some platforms (Windows), the FileSystemOperationRunner needs
to be run on the real IO runner because net::FileStream::Context
(on Windows) calls base::MessageLoopCurrentForIO::Get(). This
change provides the proper task runners to the test
FileSystemContext created in test_file_system_context.h as well
as moving the use of storage::FileSystemFileUtil to the file task
runner.

R=mek@chromium.org

Bug: 860547, 861841
Change-Id: I329dd2145488f9e4366fd80aa8b9a7ec0400ec9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1535182
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Chris Mumford <cmumford@google.com>
Cr-Commit-Position: refs/heads/master@{#653276}