PM: Remove last PageSignalGenerator user.
LocalSiteCharacteristicsWebContentsObserver (L41r) is the last user of
the PageSignalGenerator (PSG). This CL replaces that functionality with
a dedicated graph observer in preparation for retiring the PSG.
The storage of load time performance estimates is also removed. The PSG
maintains state for this that will be rewritten as a graph decorator in
due course.

Bug: 931407
Change-Id: I9840b6e9833702efde4d1587a990108c655e3b44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1579363
Commit-Queue: Sigurður Ásgeirsson <siggi@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653367}