Remove tid suffix from name in SimpleThread
This is not really needed, as tracing will keep track of the tid in a
dedicated field. Also it makes it harder for test ifrastructure to find
a given thread (prefix match instead vs exact match). Some current
infrastructure uses exact match and thus threads created with
SimpleThread might not get accounted for correctly (e.g. see bug)

If it turns out that that client code really needs this we could add
support for a placeholder %tid in the name.

Bug: 954579
Change-Id: Ibd595463d8421891616000c1d518f5bbcf95e046
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1578539
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Carlos Caballero <carlscab@google.com>
Cr-Commit-Position: refs/heads/master@{#653527}