[SendTabToSelf] Show STTS infobar when a WebState becomes active.
If we receive a shared tab when the user is not on a visible WebState
(e.g. Tab Grid or Settings Page), we register as observers to 2 things:

1- The WebStateList in case the user creates a new tab, or activates
a different tab.

2- The active but not visible WebState, in case the user comes back to
it.

When either is triggered, we'll show the infobar on the relevant
WebState and then stop observing these objects.

Bug: 944602
Change-Id: I33d1505fd639b37ed82ef22c24151b6bc8a1c371
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1574118
Commit-Queue: sebsg <sebsg@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653570}