[Mfill Android] Ensure hiding the sheet trigggers relayouting
When a keyboard disappears, the compositor waits for a relayout as a
signal to proceed with a closed keyboard.

With https://crrev.com/c/1503574, this was removed and caused tasks that
were deferred to be executed when the next relayouting happened - see
linked bug for an example.

Bug: 954125
Change-Id: I925cb2b0f4e6fab7158bf7eb16348116e5c09b1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1578611
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#653174}(cherry picked from commit daf5197ead07dbc3fb01cd4391710643b1c8c8eb)
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1581963
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/branch-heads/3770@{#45}
Cr-Branched-From: a9eee1c7c727ef42a15d86e9fa7b77ff0e63840a-refs/heads/master@{#652427}