ServiceWorker: Create resume type ServiceWorkerNewScriptLoader for resuming paused download
If ServiceWorkerImportedScriptUpdateCheck is enabled, there are two
phases in updating: byte-to-byte checking and launching a new service
worker for install/activate events. If script change is found in phase
one, the download of changed script is paused. In the second phase,
we need to resume the paused download to avoid making new network
request again.

This patch is for the second phase: Creates a resume type
ServiceWorkerNewScriptLoader in loader factory to continue the download.

Bug: 648295
Change-Id: I3720629c04fe853ef5d3ebc81996035662d93a96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1547411
Commit-Queue: Ting Shao <ting.shao@intel.com>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653607}