Remove all unnecessary heap allocation of gfx::ImageSkia
ImageSkia is a thin wrapper around a refcounted smart pointer, so there
is no need to allocate one on the heap (especially if the instance that
is being created is just a shallow copy of an existing instance).

This CL removes all of the pointer-to-ImageSkia (i.e. pointer-to-smart
pointer) cases except those needed for compatibility with the class
property system (which only allows complex types to be heap-allocated).

Bug: 953845
Change-Id: I743b41ef19012517fe0dfc3bb78f3408e14f7a54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1562762
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653716}