CupsProxyService boilerplate
DDoc: go/cups-plugin

This sets up the CupsProxy Chrome Service. This service is a singleton, managed
by the ServiceManager and lives in the browser process. The service is currently
started with the Profile; there will be future work to lazily initiate it on first
printing request. Service implementation (ProxyManager) landing in following CL.

Bug: chromium:945409
Test: Builds succeed + service unused so far. Unittests + tast tests landing in future CLs.

Change-Id: I1257f32dbff8a3bdbd5690651b7ccba9e693eb23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1547813
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Luum Habtemariam <luum@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653723}