Add new names for existing main CI bots
These bots are getting the 'archive' descriptor added to the middle of
their names. This is to both be more accurate to their actual purpose,
and distinguish them from the existing CQ bots, which will soon be
renamed to match their existing names. See the bug for more background.

Bug: 888810
Change-Id: Ib5df417ba69e806166dd17babd5214ff03418dd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1357628
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653823}