Stop using IsPasswordOverridden on iOS.
When we have multiple credentials we always allow user to select the right one.
It matches the behavior on desktop where username is always editable.
The CL cleans up some code related to the selector. It was crashing when switching
to a shorter username. ConfirmInfoBarController already supports links, there is
no need to reimplement it.

Bug: 831123
Change-Id: I877949b804257f216915b9a0b858de75f9b102ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1581631
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#653971}