Add UMA metric to track the percentage of hidden tabs that are frozen
Tracking the percentage of hidden tabs that are frozen will help us
better interpret changes in other metrics due to tab freezing. It is
suspected that many hidden tabs are not frozen due to opt-outs and the
use of web platform features that are not available when frozen.

Bug: 844426
Change-Id: Id0bab0f7e9b9c81449312ea70cdfabaa64b02242
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1570107
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Aditya Keerthi <adityakeerthi@google.com>
Cr-Commit-Position: refs/heads/master@{#654219}