ClientSocketPool:  Make cancel_connect_job close connected sockets.
In the case there's a connected socket and a pending callback,
CancelRequest would not close the socket when |cancel_connect_job| was
true. This CL fixes that, when there are no other pending requests.
If there is a pending request, the pending request gets the socket.

Bug: 912727
Change-Id: I9eedc65f2b206ea0db8e9221a18f374c5bb7f124
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1585102
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654244}