chromeos: merge ToplevelWindowEventHandler and WmToplevelWindowEventHandler
The functionality was split early on in mash when we had separate types. Now
that we don't, the split doesn't makes sense anymore. So that you can see
the diffs I kept the code in wm_toplevel_window_event_handler. I will send
a follow on patch that renames wm_toplevel_window_event_handler to
toplevel_window_event_handler.

ash/wm is not consistent in it's usage of namespaces. I went with 'ash' vs
'ash::wm' purely as it results in less churn (see patchset 2 if you want to
see some of the random places I had to add ::wm if this code lives in the
ash::wm namepsace).

BUG=none
TEST=none

Change-Id: I3081e47c1649998369542f3487e5636d0d16540e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1584744
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654251}