Encode the inspector agent state in CBOR.
This is a stepping stone for adding fields that are byte arrays.
Without CBOR, I'd have to base64 encode such fields, which is ugly.
And I do want byte array fields so that I can CBOR encode the agent
state for V8.

This change requires whitelisting the
third_party/inspector_protocol/encoding
dependency. It shouldn't be a problem because that
library doesn't depend on anything in Chromium.

Change-Id: Id84fdd1732e9857891550f433954580994bf8d5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1580779
Commit-Queue: Johannes Henkel <johannes@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654274}