Remove redundant args from OnEngineInitialized()
We are about to make the values in preferences the authoritative cache
GUID, birthday and bag-of-chips. As a prior step, simplify a few
signatures that are practically redundant (although, technically, the
values read by ProfileSyncService directly from Directory could have
changed and be more recent after this patch).

Bug: 923285
Change-Id: Ic71e12f1c265356ed7c98e91d654bc4ff9cae308
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1583879
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654362}