Add browser test for perfetto file tracer
Adds a browser test to perfetto file tracer in privacy mode. This
currently just tests if the output proto only contains the whitelisted
toplevel fields in TracePacket proto. It does not correctly verify the
submessages.
Adds a new library for checking privacy filtering in services/tracing.
Currently this is marked test_only. But in future we might want to
enable this in production on dcheck enabled builds.

BUG=925148

Change-Id: Ib85e0e67ce9716cd839aeda2158ecf0dcfe9ea13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1570235
Commit-Queue: ssid <ssid@chromium.org>
Reviewed-by: oysteine <oysteine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654379}