Shard VR Daydream View tests
The VR instrumentation tests have always taken a while to run with a
Daydream View paired due to the number of tests. For whatever reason,
they take even longer on Android P, causing them to fairly regularly hit
the 1 hour timeout on the bots. So, enable sharding for the Daydream
View version of the tests.

Bug: 953990

Change-Id: Ife442148e4799318f0665a4dfa345fbeff47a361
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1585251
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654481}