[Bluetooth] Expose BluetoothTransport via chrome.bluetooth.
We plan to implement metrics which capture pairing success in
pairing dialogs, and need to know the transport type of the
pairing device at that moment, in order to understand if any
transport type is more problematic than others. This CL
exposes the transport type to the pairing dialogs.

Bug: 953149
Change-Id: Iea54a04e9f1828e872c5ae6d5e1a072275a22582
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1582968
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654528}