Initial implementation of the Windows Memory Pressure Monitor.
This implementation is only based on the amount of free memory and the
disk idle time for now, more signals will be added later as needed.

The signal isn't used yet, this only implement the logic to build it
and a follow-up CL will put this behind a feature and add some
histograms that will help evaluate the performance of this signal.

Bug: 771478

Change-Id: Ic850c878482eb8637ec614490666f72e6146701c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1529267
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654541}