Add a README for the CompositorFrame fuzzer.
A quick README describing some of the quirks and gotchas of this
particular fuzzer, and providing a very high-level overview of how it
functions. Isn't meant to provide an introduction to fuzzing in general,
but links to further documentation available.

R=kylechar@chromium.org, riajiang@chromium.org

Bug: 923088
Change-Id: I1ca073906292216f9a92c1ecaf62e5da5d850815
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1586236
Reviewed-by: kylechar <kylechar@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Auto-Submit: Céline O'Neil <celineo@google.com>
Cr-Commit-Position: refs/heads/master@{#654559}