[ozone/wayland] Factored the clipboard logic out of WaylandConnection.
The ui::WaylandConnection class is somewhat overloaded with various logic,
so we move out things that are not directly connected to it.

Here we move the PlatformClipboard implementation out of WaylandConnection
into the standalone class, an instance of which will now be owned by the
connection object.

Bug: 942203
Change-Id: I2c36894141c220569d82266a0693b0157ce2b23d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1583700
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Reviewed-by: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654807}