Allow crowdsourcing password requirements to be used by origin.
Crowdsourcing is based on signatures not domains, but when Chrome receives
server-response, it knows the domain. Since in most cases all forms on the
site have the same requirements, it's possible to use crowdsourcing
password requirements for other forms. This CL implements this.

 1.It makes applying of password requirements more robust: as soon as
requirements found for one form on some site, they would be available
for all forms, this makes generation more robust when the form signature
changed after form load.
 2.As a result of this CL crowdsourcing password requirements will be
used on iOS.

Bug: 886583, 956982, 846694

Change-Id: Ib18f072913ec3ac4e25c855c62f6f8772e294454
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1584290
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Vadym Doroshenko <dvadym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654840}