[Autofill Assistant] ScrollIntoViewIfNeeded before click.
Before this change, WebController would attempt to scroll an element to
be clicked using the same logic as used to focus on an element. This
produced strange behavior on some websites, when zoomed, with the
element to click appearing and disappearing from the viewport.

This is a likely cause of reported issues with clicking when zoomed in.

With this change, WebController calls ScrollIntoViewIfNeeded instead of
trying to scroll nicely. It results in stabler behavior when zoomed: It
removes the flickering of element to click that was observed.

Bug: b/123958732
Change-Id: Id890737ffe48ee6ed21061222a65c2488041b73c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1587027
Reviewed-by: Ganggui Tang <gogerald@chromium.org>
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654847}