Clean up of DetermineAllowedClientPreviewsState
Removed is_reload and url from the arguments of
DetermineAllowedClientPreviewsState as they can be pulled out of
the navigation_handle also passed in.

To fix all tests using this function cleanly, the testing code has
an additional function called CallDetermineAllowedClientPreviewsState
that converts between the two inputs.

Changed GetReloadType() in mock_navigation_handle to be an fake
implementation instead of a MOCK_METHOD0 to facilitate testing of
the modified method.

Since testing the method now relies on using a mock_navigation_handle
instead of a nullptr, an extra check was added to
ShouldAllowRedirectPreview to prevent execution on the mocks.

Bug: 934400
Change-Id: I0ec63ee53d0e0c16a15592b060143d1ab49d4f0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1578061
Commit-Queue: Sean Harrison <harrisonsean@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/master@{#654965}