[Declarative net request] Allow relative redirect urls
This CL adds relative redirects for redirect rules. If a redirect rule's
URL starts with a slash "/", then it is treated as a path relative to
the extension which defined the rule. Setting the URL to redirect to
happens in indexed_rule.cc.

E.g. a rule belonging to an extension with id "asdf" which matches
"goog*" and redirects to "/abc/def" will redirect the request
"google.com/bcd" to "chrome-extension://asdf/abc/def".

Bug: 953379
Change-Id: Ia650cfcda66700dd1444d88f85982bcfb0043998
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1574947
Commit-Queue: Kelvin Jiang <kelvinjiang@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655363}