Skia GL representation for SharedImageVideo.
1.Add a new representation of the shared image for the android video.
This representation is used by SkiaRenderer in GL mode.

2. Add ProduceGLTexture() method for used by GLRenderer.

Overlay currently does not work with shared images for both GLRenderer
and SkiaRenderer and hence there is no overlay support or representation
in the SharedImageVideo backing as of now.
Once GLRenderer/SkiaRenderer supports overlay, we will add overlay
support in this shared image as well.

Bug: 956741
Change-Id: I03852eca0edc4cb6f74d61a491bce0de1bd094c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1584898
Commit-Queue: vikas soni <vikassoni@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655369}