Reland: [SpatNav] Account for clipping in z-ordering checks
The spatial navigation algorithm attempts to prioritize targets fully
contained within another target. However, the current implementation has
a bug because it uses unclipped rects to determine containment.

This CL fixes the issue by using MapToVisualRectInAncestorSpace which
does account for clipping.

Small cleanup: Remove the |ignore_border| option since it's always true.

[Reland Note]

This CL originally caused a Mac-only web test failure in:
fast/spatial-navigation/snav-stay-in-overflow-div.html

This revealed an actual issue: due to the line height difference, the
root_frame_rect of an interested node was clipped out. However, we
didn't consider the rect to be "Offscreen" because we use the
RectInViewport function for that. This function also clips but starts
with a slightly different rect: VisualRectInDocument vs
LocalBoundingBoxForAccessibility. This means that the node is considered
visible for some calculations but not-visible for others.

Clearly, we should be using one notion of "rect" across the spat-nav
system. We previously kept these reparate rect types due to layout
tests and compat concerns. However, I believe they've all been addressed
between this CL and
https://crrev.com/efea025461c496341145b5fb062c04b71f8c85a7.

This reland replaces the rect created in RectInViewport with
RectInRootFrame and simply transforms that up to the visual viewport,
making the two functions consistent. It also removes the hacky
"ensure not empty" check in RectInRootFrame that inflated empty boxes.
This also appears to be covered by the changes above and would
otherwise interfere with calculating whether an element is
visible/onscreen.

Base patchset contains originally landed/reverted CL. Changes to reland
in newest patchset for easy diff.

Bug: 957240
Change-Id: I9341ab05fb4f5f8c46fe6d43d3334e702395bcea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1589233
Commit-Queue: Emil A Eklund <eae@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655388}