Fix the presubmit check for files that are too big.
We added a check that no one is trying to add a file > 20MB
to the repo in r655002. That check depended on stat'ing files
using the path relative to the repo root.

However, it turns out that changes in a nested repo can also
cause the presubmit checks in this repo to run, and the local
path stat doesn't work in that case (it uses the wrong git repo
root).

This CL switches the check to using full paths, which doesn't
have that problem.

Bug: 958106
Change-Id: I929473efba38d0a17eb37025abb8030cfb9fafb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1590446
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Daniel Bratell <bratell@opera.com>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655421}