viz: Change OutputSurface and BeginFrameSource creation order
Before this CL DisplayProvider would take a BeginFrameSource and create
an OutputSurface, DisplayScheduler, and Display.

For GPU surface based begin frames we want to check if output surface
supports it and create an appropriate BFS based on that.  This CL
changes the order in which BFS and OutputSurface are created:

1) DisplayProvider (renamed to OutputSurfaceProvider) returns an output
   surface instead of Display.
2) RootCompositorFrameSinkImpl doesn't need two phase initialization. It
   creates the BFS after getting the output surface and then creates the
   Display.
3) FrameSinkManager supports creating RootCompositorFrameSink with an
   external begin frame source for compositor frame fuzzer.

Bug: 953970
Change-Id: I78d1700b24dc87a35096bde3993410a5da98a9ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1580859
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655478}