Dark Mode: Allow native themes to track dark mode status from another theme
The implementation for the `prefers-color-scheme` media query reads dark mode status from ui::NativeTheme::GetInstanceForWeb().

On non-Mac platforms, this is *not* the platform native theme, so we need to manually
set it for Windows in order for correct dark mode status to be picked up.

We don't want to expose setting dark mode status publicly, so letting a theme ask another to explicitly track it via DarkModeObserver seems like the cleanest solution.

Bug: 956873
Change-Id: Ia668e178084b1ddcc184635705c1d29ab17caa95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1588766
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655637}