PM: Tweak chrome://discards graph layout a little.
This change positions all nodes and "cools" the simulation on first
resize to avoid the crazy bouncing that would occur. Also moves the
node clamping to a force in the simulation, as that way the nodes can
be dragged anywhere.
New nodes are also given a default position in the middle of the graph
and their respective Y clamping range.

Change-Id: Ideb2158f321d79c7f5c4ccd3f8d68b7eef0ea317
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1590409
Commit-Queue: Sigurður Ásgeirsson <siggi@chromium.org>
Auto-Submit: Sigurður Ásgeirsson <siggi@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655646}