Add CSS specific to desktop devices.
The existing CSS works well for Chrome on mobile. However, it leaves a
large amount of whitespace to either side of the central text column,
which is much narrower than it needs to be for a desktop display. This
CL adds some styling to increase text width and focus attention on the
center of the page when on desktop, as well as generally make it look
more polished.

Bug: 925965
Change-Id: I0cc0b4a74b57622f0196790bb920e1db4f0fc433
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1572276
Commit-Queue: Aran Gilman <gilmanmh@google.com>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655717}