SessionTerminationManager added
A Chrome OS device might become locked to single user if it is forced by
"device_reboot_on_user_signout" device policy. In that case, only the data
of a single user become accessible on the device until it reboots.
Internally that is done by extending one PCR in TPM.

The policy is applied inside a user session (either at login or when ARC++
session is started). Then the device can be used only by that user. If the
device is locked to single user, the policy effect should be that when user
does a sign out, device reboots instead. After reboot the device is in
unlocked state and can be used by other users too.

The new class (SessionTerminationManager) manages the session termination
requests, deciding if reboot is needed when the policy locked the device
to single user, or a sign out is enough. It keeps the information if the
device is locked to single user in a local variable. The value of variable
is set to false initially.

When the policy is applied by LockToSingleUserManager class, the variable
inside SessionTerminationManager is updated to true.

At user sign out, if the value of variable is set to true, it will force
a reboot.

At login screen, if the policy is set, a request for TPM status is sent
to cryptohome to check if the device needs to reboot. The response from
TPM status request contains information if the device is locked to single
user. This might happen if Chrome crashed while in user session (e.g. on
lock screen) and was returned to sign in screen. This also covers the case
if Chrome crash caused the cache from SessionTerminationManager to be lost
and the device did a sign out instead of reboot before.

BUG=b:77799573
TEST=chromeos_unittests

Change-Id: I3d8ca16ff4728c3a58cb8c214932a304e538a55e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1508413
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#655924}