[Background Sync] Detect PWAs on desktop.
Chrome will only make Periodic Background Sync available to PWAs
on Desktop. When a PWA is installed for the origin for which the
periodic Background Sync permission is being requested, we also
ensure that content settings for both one-shot and periodic
Background Sync are set to ALLOW.

This adds that check, and a unit test.

Subsequent CLs will add similar checks on Android.

Bug: 925297
Change-Id: I7531f699b8ed8bcfc2c9883651d66ea95645bf0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1549070
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656005}