Improve Chrome's DLL prefetcher to save memory
Change the in-proc prefetcher to load the code into Image pages and not MapFile pages
to save CPU and improve memory usage, and do nothing on OS builds that enable OS
PreFetch of larger files. Note this require the PreFetchVirtualMemory API, so we'll
still read the file as data/MapFile on Win7.

Bug: 547794
Change-Id: I0e9ec9615131daa065a8d0ccbbb897de059524c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1534983
Commit-Queue: Joe Laughlin <joel@microsoft.com>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656187}