Inhibit tree scroll unless really needed.
CL:1546804 added scrolling for when directory tree items get
focus (to allow items that are tabbed with the keyboard to
come into view).

A side-effect of this caused selection and window focus to scroll
the directory tree even if the item was within view, causing
non-focussed items to be scrolled out of view.

This change uses the non-standard scrollIntoViewIfNeeded() method
to scroll the item only when necessary and then checks the element
to see if it moved. If it has moved, the scrollIntoView() method
is called to horizontally position the element since the latter
method honours the 'inline' positioning property whilst the former
method does not.

Bug: 955172
Change-Id: I61e04cdd2529274ee19cfd9b8e7a0e387132af35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1593885
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656302}