Implemented Swap Chain methods at service side.
CreateSwapChain method creates DXGI Swap chain and then wraps its front
and back buffer D3D11 textures into GLImages. For each buffer, a GL Texture
is created which is stored as gles2::Texture or gles2::TexturePassthrough
in SharedImageBackingDXGISwapChain that was created. The backing also holds
reference to created swap chain.

PresentSwapChain is implemented in SharedImageBackingDXGISwapChain. After
flipping the swap chain, we also have to rebind the surface representing
the back buffer. This is implemented in the BindTexImage implementation in
the GLImageDXGISwapChain subclass.

Bug: 939655
Change-Id: I646206f532728be4ae85af54f3eae5cc5fc4e74a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1573092
Reviewed-by: Antoine Labour <piman@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Ashitha Santhosh <ashithasantosh@google.com>
Cr-Commit-Position: refs/heads/master@{#656305}