[Chromecast] Ignore decoded sample rate if bypassing decoder
If the decoder were being bypassed in AudioDecoderForMixer and
SetConfig() were called with a new sample rate while a task to call
OnBufferDecoded() was pending, OnBufferDecoded() would set the sample
rate back to the old sample rate when the task was run. This changes
OnBufferDecoded() to ignore the sample rate when bypassing the decoder
because the sample rate is always based on the config, not the decoded
buffer.

Bug: internal b/131527230
Test: None
Change-Id: I784ac7b8fa936cdfaa86f6925a01c6478db3189f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1594237
Commit-Queue: James West <jameswest@chromium.org>
Reviewed-by: Kenneth MacKay <kmackay@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656312}