Misc improvements for loopack audio support on Windows.
Given recent work in https://bugs.chromium.org/p/chromium/issues/detail?id=938938
it was found that the code related to loopback support on Windows could contain
parts where it was not clear if the device role was taken into account properly.

This change is an attempt to clean up parts code where usage of a loopback device
might be broken. Basically, the change is an attempt to make the code more readable.

BUG=956526

Change-Id: Ia4f4ea177f3cd50805677790dbf0bdbd743561b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1583708
Commit-Queue: Henrik Andreasson <henrika@chromium.org>
Reviewed-by: Henrik Grunell <grunell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656375}