Load/fullscreenchange events may fire in any order. Fix test.
Don't assume that the fullscreenchange event arrives before the load
event in backdrop-object.html . It actually used to be safe to assume
this, prior to CL:1459542, because, without that fix, we'd always reload
the SVG upon layout object reattachment (even if it had already loaded).
A reattach occurs in that test when the OBJECT goes fullscreen. With
the aforementioned CL in place, we'll attempt to retain the SVG through
reattachment, so if it actually manages to fully load before we go
fullscreen, we'd fail to catch the load event at all in this test, since
we required the fullscreenchange event to fire before the load event,
for the test not to time out, and this suddenly caused flakiness.

Also set up the event listeners *before* requesting fullscreen, for good
measure (it may be unlikely that the events fire synchronously from the
request, but, then again, why take the risk?)

Bug: 957916
Change-Id: I01484b932520aba698df784c9057b86d0bbe7230
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1591848
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656654}