Quantify effect of prediction on AverageLag
Once scroll prediction is turned on by default, it will become
increasingly difficult to understand the impact that prediction is
having on this metric over time. For example, if the AverageLag metric
regresses it could be due to increased latency or a reduction in
prediction effectivness.

Because of this, we'd like to log a pair of histograms for what the
effect prediction is having on the AverageLag metric.
This information is already available to the AverageLagTracker as the
scroll_delta field on the LatencyInfo. Currently this is used as the
actual position at time=t, but we know that it would have been used
as the rendered offset in the absence of prediction.

I implemented this as two histograms because count-based histograms
don't appear to support negative numbers. Perhaps I'm overlooking
another type of histogram that would be better suited for this.

Bug: 836352

Change-Id: Ief387d0fc087e53cd4137dd95d91bb331b592cc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1586104
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Timothy Dresser <tdresser@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Reviewed-by: Ella Ge <eirage@chromium.org>
Cr-Commit-Position: refs/heads/master@{#656907}