Add presubmit check for test expectations with globs
Add presubmit check for collisions between test expectation patterns
which have opposing expectations. One such collision this CL tries to
prevent is when a test expectation with a glob has the failure expectation
and another test expectation whose pattern matches with the glob does not
have the failure expectation nor the skip expectation. Another collision this
CL tries to prevent is when a test expectation has the skip expectation and
another test expectation whose pattern matches with the glob does not have
the failure expectation nor the skip expectation. These presubmit checks try
to prevent unexpected regressions by making sure that more specific test
expectation patterns with the pass expectation are not over looked when
a test expectation with a less specific pattern is added and specifies
the failure or skip expectation.

Bug: chromium:948891
Bug: chromium:698902
Change-Id: Ia6587b87b78c6692552c6b6b57c00916f34364d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1597638
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657093}