Does not extend or shrink the AX selection when the DOM endpoints are unavailable in the AX tree
If he anchor or focus of a DOM selection are
either ignored or unavailable in the
accessibility tree, we extend or
shrink the DOM selection until we find
valid selection endpoints in the
accessibility tree.
However, in the case of a collapsed selection in the DOM
this causes problems with clients that expect both
the anchor and the focus to be the same.
To resolve this, we move both anchor and focus to the right
until we find a valid selection endpoint in the
accessibility tree when dealing with a collapsed
selection, while following the previous logic in other cases.

R=dmazzoni@chromium.org, aboxhall@chromium.org

Bug: 639340
Change-Id: I1e35bf8f08e5dc82628ecfae18b66424d2719b11
Tested: web tests
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1597495
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657154}