a11y: implement text control protocol in //ui
It turns out that the documentation is telling the truth about these
methods being required :). Not implementing them (especially the
accessibilityInsertionPointLineNumber and
accessibilitySelectedTextRange methods, for unclear reasons) causes
VoiceOver to ignore AXValueChanged notifications for the control.

The new implementations are largely ripped off from the implementations
of the old API methods, and roughly the same set of methods are left
unimplemented.

Bug: 957355
Change-Id: Iec2f9bda651f885378de46e121eb13a106656033
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1593396
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657212}