Adds C++ SamlPasswordAttributes class to match JS
During the SAML flow, information about the password - such as
when it may expire - can be provided by the SAML IdP. This is
extracted using javascript during the SAML signin process.
This adds a C++ class to matches the JS information - this
class is tested but not yet used.

In a follow up CL, this will be wired up so that the
information extracted during login is saved to a users's prefs,
and a BrowserTest will be added to test the whole flow.

Bug: 930109
Change-Id: I0fde7e5808c0c3995ffed765457e788b9e701876
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1594586
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Commit-Queue: A Olsen <olsen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657267}