Use higher priority to record Mac power stats
Seems like in practice, when the system is under load, BEST_EFFORT priority tasks can have *significant* leeway. This means that just when the system is most likely to be using a lot of power, we are sampling at a lower frequency, which skews the Power.Mac.* stats low.

This change moves recording up to the next highest priority level.

Bug: 956241
Change-Id: If46b3bf91ef6cef6b52370195aef51bd8357567a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1593400
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Reviewed-by: Sidney San Martín <sdy@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657268}