[Mfill Android] Enable keyboard accessory v1 by default
Since we launch the keyboard accessory in M74, set the feature to
enabled by default with M76 (to keep as kill switch).

With M77, drop it then - this is already reflected in the
flag-metdata.json.

Automatic password generation is already enabled and also targeted to
removed with M77.

Bug: 853794
Change-Id: Iaf0dfe51753445eb868c27d5401636d4f0b1c42d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1598821
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657270}