Remove ResourceType usage in NetworkServiceClient::OnSSLCertificateError.
The browser only needed to know if the request is for a main frame navigation or not. It can already
deduce this by looking at the requesting factory and request's IDs.

This is a step towards the TODO to remove this from network::ResourceRequest, since network service
shouldn't know about this enum.

Bug: 960143

Change-Id: Id01b2f654b9af8d2b42c2e861eead1675c6fe7ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1597198
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657304}