[merge m75] Mark ProfileKey context live before it is first used in TestingProfile.
Most unit_tests get the invoke
SupervisedUserSettingsServiceFactory::GetForKey with the TestingProfile
key in this stack:

  DependencyManager::AssertContextWasntDestroyed  [*]
  KeyedServiceBaseFactory::AssertContextWasntDestroyed
  SimpleKeyedServiceFactory::GetContextToUse
  KeyedServiceFactory::GetServiceForContext
  SimpleKeyedServiceFactory::GetServiceForKey
  SupervisedUserSettingsServiceFactory::GetForKey
  TestingProfile::Init
  TestingProfile::TestingProfile

Before it is marked live in this stack:

  DependencyManager::MarkContextLive
  DependencyManager::CreateContextServices
  SimpleDependencyManager::CreateServicesForTest
  TestingProfile::Init
  TestingProfile::TestingProfile

When a single unit test is run, the key is not in
DependencyManager::dead_context_pointers_ when [*] is called for the
first time, so there is no failure. When multiple tests are run, it
is possible for the key to already be in
DependencyManager::dead_context_pointers_ when [*] is called for the
first time in a test, so there can be failures.

This CL fixes the issue by marking the context live early in
the constructor of ProfileKey.

(cherry picked from commit c1a89979c6ee905a5cade3c2e15018a94e4f1b05)

Bug: 956718, 957124
Change-Id: I0476567497d95f0ccaf65a1f6362ae0def8fbcf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1590418
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#656793}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600073
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/branch-heads/3770@{#446}
Cr-Branched-From: a9eee1c7c727ef42a15d86e9fa7b77ff0e63840a-refs/heads/master@{#652427}