Restore original tail-processing for ScriptProcessor and AudioWorklet
While the idea in
https://chromium-review.googlesource.com/c/chromium/src/+/1544880 was
correct, we forgot to consider what happens when a source stops and is
disconnected to the ScriptProcessorNode (AudioWorkletNode).  In this
case, the output of the node is disabled.  But that's incorrect in
general because the node's processing must continue because it has
user-defined outputs so a silent input does not imply a silent output.

Rather than adding a special case in DisableOutputsIfNecessary, let's
just revert the change.  The internal properties aren't exposed to the
developer.

Bug: 959125
Test: the-scriptprocessornode-interface/simple-input-output.html
Change-Id: I0f217be24eacff6031a537086df535e6c1e8d9d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1598427
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657447}