Switch GetDeepLinkParam() to using UnescapeBinaryURLComponent().
It was using UnescapeURLComponent(), which leaves certain characters
escaped ("%00" through "%1F", and certain UTF-8 sequences which are
unsafe to display in the omnibox, like lock icons and 0-width spaces).
This behavior is generally not useful in most contexts - it's almost
always better to fail on bad input than to use partially unescaped
strings, and the UTF-8 logic isn't useful here.
UnescapeBinaryURLComponent(), on the other hand, unescapes input
unconditionally.

This CL is part of a project to clean up uses of
UnescapeURLComponent(), as it's been used for wildly disparate
purposes, and much of its logic is only suitable for cases where
we want to format a URL for safe display to the user. We want
to make it leave more weird characters escaped, but need to
eliminate usage of it in other contexts before we can do that.

UnescapeBinaryURLComponent() should be used instead in most cases,
particularly where URL_SPECIAL_CHARS_EXCEPT_PATH_SEPARATORS is in use,
as all consumers of that option seem to either be slightly buggy, or
don't actually need to leave any characters unescaped. The primary
cases where switching to UnescapeBinaryURLComponent() may cause
problems (without additional checks), are cases where unescaping
path separators or funky characters (like %01) is a problem. In those
places, we should generally fail instead of partially unescaping a URL.

Bug: 959213
Change-Id: I88cd0aff55c7a5cefd5224d011562810d7c3d150
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1598785
Reviewed-by: Yue Li <updowndota@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657480}