Send cancel-reset user action on reset error screen
An error message is displayed when user requests a factory reset and
the reset fails. The error screen shows a button that returns the user
to the OOBE flow when clicked.

This CL fixes the user action event sent to the C++ handlers. The
previous event we were sending no longer has any handlers.

This CL also modifies existing tests to verify that the event is being
handled.

Change-Id: I2871ba19283325fc7535cd107d3d27719e8a18be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1598090
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Commit-Queue: Tony De Luna <tonydeluna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657482}