Make OutOfProcessInstance::SaveToBuffer use net::GetSuggestedFilename().
It was just using net::UnescapeURLComponent(), which both has some
behaviors not suitable for this purpose (Like deliberately leaving some
characters we can't safely display to the user, like locks and \x01,
escaped).  Also, net::GetSuggestedFilename() is specifically designed
for this purpose, and has some behaviors that are desireable here
(Like replacing characters that shouldn't appear in file names, and
trimming whitespace).

It turns out that most usages of UnescapeURLComponent() are wrong, and
this CL is part of an effort to fix that.

Bug: 849998
Change-Id: Ib98e178e7f7cc002e49986026a6799b7c581caa4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1596362
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657486}