[OT-PW] Push a saveLayer op when rastering PaintWorklet results
In the off-thread paint worklet case, the drawing operations are captured
in the DrawImageRectOp::RasterWithFlags. While in the main thread paint
worklet, the drawing operations are in the PaintGeneratedImage::Draw,
this CL makes the operations in the DrawImageRectOp::RasterWithFlags
match the operations in the PaintGeneratedImage::Draw.

This change makes 64 off-thread paint worklet layout tests pass.

Bug: 957459
Change-Id: Ie33b32c9f1a15a2cec3886c66486bae147100015
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1596176
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657497}