Change aapt2 and bundletool install mode to copy
Changes aapt2 and bundletool to be installed from CIPD via copying
instead of symlinking. This is a temporary workaround for swarming
isolation not working properly with symlinks that point outside of the
src/ directory, which is the case for things downloaded via CIPD.

This can be reverted once symlinks are properly supported.

Bug: 950727
Change-Id: I3b92959cf6e33e0d2edfd5adfc05d4b04ad7b909
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1599683
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657576}