Updates ChromeVox and Select to Speak to work with the new AXSelection in Blink
1. Changes the Cursor class in ChromeVox to allow tree positions on richly editable nodes
Due to bugs in the old selection code in Blink, a ChromeVox Cursor
could not be set on a content editable root or any of its descendant nodes, other
than a text node.
The new selection code does return selection endpoints that are tree positions on
richly editable nodes, and so ChromeVox should both handle as well as set such endpoints.

2. Completely eliminates the use of anchor and focus objects in ChromeVox editing handler.
This patch uses the newly added attributes selectionStart/End which are
guaranteed to be in document order.
To determine if selection has been altered from the start vs. the end, which is
important for determining what needs to be spoken or brailled,
this patch relies on a boolean "isSelectionBackward" that is sent
directly from Blink, instead of relying on checking which endpoint has been modified.
3. Modifies Select to Speak based on a patch provided by katie@
and augmented by me.
Uses "isSelectionBackward" to switch anchor and focus around so that they will always
be in document order.

R=dmazzoni@chromium.org, dtseng@chromium.org

Bug: 639340
Change-Id: I072f9c9dfff5b6c9a107e66207fb9cc80f8d4cad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1572523
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657593}