Switch ArcFileSystemBridge to UnescapeBinaryURLComponentSafe().
It has been using UnescapeURLComponent() instead to unescape file names.
UnescapeURLComponent() leaves various UTF-8 characters that are unsafe
to display in the omnibox, but valid in other contexts escaped, as well
as "%00" through "%1F". In addition, it was being told to keep path
separators escaped. The UTF-8 logic was unnecessary here, and it's
better to fail on unexpected input rather than to partially unescape
it. Switching to UnescapeBinaryURLComponentSafe() gets us both of
these behaviors.

This CL is part of a project to clean up uses of
UnescapeURLComponent(), as it's been used for wildly disparate
purposes, and much of its logic is only suitable for cases where
we want to format a URL for safe display to the user. We want
to make it leave more weird characters escaped, but need to
eliminate usage of it in other contexts before we can do that.

UnescapeBinaryURLComponent() should be used instead in most cases,
particularly where URL_SPECIAL_CHARS_EXCEPT_PATH_SEPARATORS is in use,
as all consumers of that option seem to either be slightly buggy, or
don't actually need to leave any characters unescaped. The primary
cases where switching to UnescapeBinaryURLComponent() may cause
problems (without additional checks), are cases where unescaping
path separators or funky characters (like %01) is a problem. In those
places, we should generally fail instead of partially unescaping a URL.

Bug: 849998
Change-Id: I009667ec4ccf92e49810ece56d5e2f8e9604a096
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1598321
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657699}