Remove Google new_tab_url in prepopulated_engines.json
When Google is the DSP we now use the local NTP built into Chrome
instead of the remote NTP specified in prepopulated_engines.json.
Remove the entry, as the loading of the local NTP is handled by
logic in chrome/browser/search/search.cc.

Bug: 583289
Change-Id: Id28db805bc53eebab6d492be666f909e0ff0b800
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1591796
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Kyle Milka <kmilka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657763}