PerformanceManager: Close previous Document binding on navigation commit.
With this CL, when a navigation is committed, FrameNodeImpl closes the
binding from the previous document. This ensures that properties in the
FrameNodeImpl can't be updated by the previous document.

Additionally, ProcessNodeImpl and FrameNodeImpl use a Binding instead
of a BindingSet to convey that a single entity should send them messages
at a given time.

Change-Id: Ieceba785b691eadca6af7d6fccc6cc7aac105981
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1572459
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Cr-Commit-Position: refs/heads/master@{#657880}