[Autofill Assistant] ScrollIntoViewIfNeeded before click.
Before this change, WebController would attempt to scroll an element to
be clicked using the same logic as used to focus on an element. This
produced strange behavior on some websites, when zoomed, with the
element to click appearing and disappearing from the viewport.

This is a likely cause of reported issues with clicking when zoomed in.

With this change, WebController calls ScrollIntoViewIfNeeded instead of
trying to scroll nicely. It results in stabler behavior when zoomed: It
removes the flickering of element to click that was observed.

This is the cherry pick merge for the M-75 branch which was approved here: http://crbug.com/960391

(cherry picked from commit 85abeb6dff8839c728d5e007fc932a5a46667832)

Bug: 960391
Bug: b/123958732
Change-Id: Id890737ffe48ee6ed21061222a65c2488041b73c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1587027
Reviewed-by: Ganggui Tang <gogerald@chromium.org>
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#654847}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600853
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Cr-Commit-Position: refs/branch-heads/3770@{#497}
Cr-Branched-From: a9eee1c7c727ef42a15d86e9fa7b77ff0e63840a-refs/heads/master@{#652427}