Start CupsPrintersManager with BrowserContext
This change makes CupsPrintersManager start with the BrowserContext
rather than lazily. Printer discovery runs on construction of CPM, so in
order to support the autoconfigure USB case, we want printer discovery to
begin running as soon as a user logs in. Previously, CPM was started
lazily the first time either Print Preview or CrOS Settings was opened.

Bug: 951135
Change-Id: I343a8973ae2f4f07283878533fa8d2c0985c20ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1584996
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Commit-Queue: Bailey Berro <baileyberro@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658194}