Don't create ink drop container layer in LabelButton constructor
This layer can be created and destroyed as needed when an ink drop is
installed. Creating it on construction adds a cost to every
LabelButton, when there is usually only one ink drop active at a time.

InkDropContainerView already manages this upon adding or removing
ink drop layers, so no additional changes are needed to LabelButton.

Bug: 946738, 960442
Change-Id: If67b443296f85fc296b815d5c61b28683e8d233f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1603182
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658285}