Add SiteSettingsHelper
This adds SiteSettingsHelper which contains helper methods for
showing site settings and determining whether site settings is
available. Since this logic was previously part of PageInfoController,
it was difficult for other classes to use it. A downstream CL that
uses this class is included in the bug.

Bug: 952947
Change-Id: Idcf10232631605d2e2cf7a108c9b26a18b580965
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600436
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Mehran Mahmoudi <mahmoudi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658436}