Runs the "on selection removed" notification test after layout and paint
The affected test would count the number of selection changed notifications and succeed if 4 notifications had been received.
This is brittle because the test was counting both the notifications that resulted from the setup method in addition to the test itself.
To exclude the selection changed notificationsfired from the setup method, the test is mmodified so that it runs after layout and paint.
R=aboxhall@chromium.org

Bug: 639340
Change-Id: I766138415f6d36042214eeba3ada02645f54dc44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1601994
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658495}