Allow insecure updates for Kiosk apps
Insecure in that sence that they are useing CRX2 archive format. This
format is deprecating now (in favor of CRX3), and will be fully removed
(according to the plan) in M77. Until then it should be possible to use
old archive format for policy-based extensions (it makes sence only for
third-party-hosted ones, as Chrome Web Store already serves only CRX3).

For user sessions there is a policy “ExtensionAllowInsecureUpdates”
controlling if CRX2 is allowed (default to false from M75), but there is
no such one for Kiosk. This commit allows CRX2 for Kiosk apps.

Bug: 960428
Change-Id: I3f0599e209fee234cb216c11eeca79b170149e9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1602727
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Oleg Davydov <burunduk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658615}