Fix calculations for frame and tile threads with dav1d.
21699e36971304727ce89601ff9b669e57e63ecb incorrectly stated that when
configuring tilethreads=t and framethreads=f that the total number of
threads would be (t - 1)  + (f - 1), the actual total is t * f.

To avoid overpopulating thread counts for lower end systems, correct
this mistake when asking for the system limits to be applied to the
total.

If a system has the cores for it, we'll now use the following:
<300p: 2 tile threads, 2 frame threads = 4 total threads.
<700p: 3 tile threads, 2 frame threads = 6 total threads.

For higher resolutions we hit limit::kMaxVideoThreads (16):
<1000p: 5 tile threads, 3 frame thread = 15 total threads.
>1000p: 8 tile threads, 2 frame threads = 16 total threads.

Due to the (surprising) performance issues which occurred when setting
framethreads=1 (https://crbug.com/957511) the minimum number of threads
is 4 (two tile and two frame) regardless of core count. The maximum is
2 * core_count (up to limits::kMaxVideoThreads).

BUG=954659,957511
TEST=manual inspection of pthread_create() calls.
R=chcunningham

Change-Id: Ic261b3279b60959ab9fc3eb51edbc2767c5b9f9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1605160
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658643}