[Autofill] Updates unused bool parameter in PrepareSuggestions.
The contains_address bool previously determined which type of icon to
add to a suggestion in the autofill-use-improved-label-disambiguation
experiment. Ultimately, one icon was chosen, so contains_address is
unused.

This CL changes the boolean to store whether a LabelFormatter in the
experiment was successfully created. If so, an icon is added.

This change is necessary because unexpected forms, such as those
without name fields, add an icon to suggestions that are not made with
a LabelFormatter. However, in these cases it is preferred to not show
the icons.

Screenshot:
https://screenshot.googleplex.com/7gsK2RE8Wu9

Bug: 951174
Change-Id: I8aa1434506d2639cf59677f53cf1540eeffd37b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1602733
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658665}