[ MimeHandlerView ] Fix a loading race
When a resource is intercepted at PluginResponseInterceptorURLLoader, it lets
MimeHandlerViewAttachHelper override the response (which injects a static HTML
page that embeds a GuestView). At this stage, the attach helper post the
creation of MimeHandlerViewEmbedder class to UI (the MHVE will monitor the set
up of MimeHandlerView).

This opens up room for race: if the resource loading finishes sooner and
ReadyToCommitNavigation is called before MimeHandlerViewEmbedder is created then
the MimeHandlerView won't load.

This CL, fixes this problem by deferring load until MimeHandlerViewEmbedder is
created on UI thread.

Bug: 659750, 961731
Change-Id: I8b6e5cc0fb6035464384b50fd182cb00a73ec4f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1605162
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Ehsan Karamad <ekaramad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658743}