(fake) session_manager: StorePolicy callback waits for owner key write
Makes sure that in-memory FakeSessionManagerClient does not run
StoreDevicePolicy callback until the owner key file was written (if new
owner key has to be written).
SessionManagerOperation tries to load the owner key in response to
StoreDevivePolicy, and the task to load the key might be run before the
task to write it (as they are not using the same sequence task runner).

BUG=952855

Change-Id: I41c584869741ba3d06245c378f2ef0bccd06b0a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1604216
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#658756}