[css-grid] Update grid when changing auto repeat type
This patch makes two 'repeat()' values for 'grid-template' be considered
to be different if one uses 'auto-fill' and the other 'auto-fit'.

Previously, they were considered to be equal if the repeated values
were the same, without comparing the repeat type. Therefore, the grid
was not updated when setting both values one after the other.

BUG=961407
TEST=fast/css-grid-layout/grid-change-auto-repeat-tracks.html

Change-Id: I4e0202097097337aef600e5e8cb732b076bc8ade
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1606840
Reviewed-by: Manuel Rego <rego@igalia.com>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#658852}