Add a real geometry class for rounded corners instead of a vector.
Currently, we pass values into Layer::SetRoundedCornerRadius() as a
vector, which means that the values are just an ordered list with
implicit correspondence between order and what value they represent.
Consider the following code:
  layer->SetRoundedCornerRadius({3, 4, 5, 6});

Which corner is getting a radius of 3? We have to look at the
documentation for the function to find out that the corners are in
clockwise order starting from the upper left - but it's not obvious.
Passing values with specific semantic meaning in lists and vectors is
generally considered an anti-pattern as it is very easy to mix up the
order, and even sophisticated editors can't prompt you for the meaning
of each value as you enter them.

This CL introduces a new class: gfx::RoundedCornersF, which emulates the
rounded corners portion of a gfx::RRectF (and indeed, we've added a
version of the RRectF constructor that takes a RoundedCornersF).
RoundedCornersF provides an explicit way to pass corner parameters
(using named fields), and convenient constructors (with named
parameters) to create sets of rounded corners with:
 - all the same radius
 - four distinct corner radii with equal x and y radii

(See issue 961756 for discussion of extending this to distinct x and y
radii once certain technical limitations disappear.)

One consideration is that the new object is, internally, a list of four
float values, while the old API took a vector of four ints. However,
internally it was already being converted to an RRectF, which has eight
floats. And even better, since there is a four-float constructor for
RoundedCornersF, existing code that used a list initializer in calls to
Layer::SetRoundedCornerRadius() still compiles (with the values
representing the same corners as before, since we've preserved the
order).

Change-Id: I36d0037f8986468c146adf6fd5eea5c4d5dce9cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600965
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Reviewed-by: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659097}