chrome/browser/ui/webauthn: double-UI resident creds on Windows API v1.
This change adds a confirmation step when creating resident credentials
with Windows Webauthn DLL version one (currently in 19H1). This version
requires the use of native APIs for interaction with security keys, but
doesn't show a notice informing the user that creating a resident
credential will store information on their security key.

However, such a notice is a privacy requirement and therefore a
confirmation step is added prior to dispatching the API call.

Bug: 941120
Change-Id: Ia39aefe7e2b294a604346a66851dbda67372bd63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1601299
Reviewed-by: Martin Kreichgauer <martinkr@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659267}