Fix proxy_connection tests for M75
The tests use some neterrors that are internally falledback and handled
by the NetworkService. So the DRP loader throttle does not receive them
to record DRP InvalidResponseHeadersReceived.NetError histograms.

This test instead relies on warmup fetch failures being recorded.

Theere is some difference with M74 and M75, that the
WarmupURLFetcherCallback.SuccessfulFetch.InsecureProxy.Core is
recorded in M75 and InsecureProxy.NonCore. The test is fixed to run only
for M75+.

Bug: 934875
Change-Id: I2a2147b2d0d94343334494dd7d3558b1e6ec4201
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610508
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Commit-Queue: rajendrant <rajendrant@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659345}