[Autofill Assistant] Allow overriding initial terms&conditions state.
This is the M-75 equivalent to http://crrev/c/1605202. Due to a recent refactoring of the payment request which is not part of M-75 (http://crrev/c/1587023), this CL had to be developed on the branch directly.

Bug: 961778
Bug: b/131588758
Change-Id: I0a338d78da1d61bcba301aea5f457c71b43828c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1604116
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Cr-Commit-Position: refs/branch-heads/3770@{#586}
Cr-Branched-From: a9eee1c7c727ef42a15d86e9fa7b77ff0e63840a-refs/heads/master@{#652427}