[Merge to M75] Links containing positioned images should be annotated for accessibility.
The Get Image Descriptions from Google feature provides a
computer-generated description for unlabeled images on the web.
Importantly, when there's an unlabeled link that contains a single
image, it's supposed to apply the same automatic description to
the link, too - because some screen readers only pay attention to
the accessible name of the link.

There was a bug when the image was absolute-positioned - that caused
the link to have zero width and height, which made it ineligible for
annotation.

The fix is to only consider the size of the image when deciding
whether a node is eligible for annotation, rather than the size
of the link.

TBR=nektar@chromium.org
(cherry picked from commit c715358ade83d020a935db0cd373b6e0f672f90f)

Bug: 962370
Change-Id: I1a8d845666d2a508e17466da75bce0c88cd1d3fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1608922
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#659070}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1612037
Cr-Commit-Position: refs/branch-heads/3770@{#603}
Cr-Branched-From: a9eee1c7c727ef42a15d86e9fa7b77ff0e63840a-refs/heads/master@{#652427}