Use settings_test_util.h only in chrome_earl_grey.
Use [ChromeEarlGrey setContentSettingsBlockPopups]
instead of chrome_test_util::SetContentSettingsBlockPopups(..).

This is part of EG1 to EG2 migration, which involves moving EarlGrey code from
app-side helpers into test code. For GREYAsserts, the policy is to take assert-
containing-helpers, modify them to return NSError* with localizedDescription
containing the assert failure (in the event of a failure), and then assert
on the result of the helper on the test side.

Bug: 922813
Change-Id: Iba1f291af45c2676e0d04158542c4752716cf4bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600635
Commit-Queue: Maksym Onufriienko <monufriienko@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659586}