wpt: Support http://host:80/ at EventListener / incumbent test
MessageEvent's origin attribute will be
    http://hostname
instead of
    http://hostname:80
in case that the port number is HTTP standard port number, i.e. 80.

Usually this is not a problem because wptserve runs with port=8001
by default, but w3c-test.org runs with port=80.
http://w3c-test.org/dom/events/EventListener-incumbent-global-2.sub.html

This patch supports such a case.

Bug: 622974
Change-Id: I4567de3f5f12ae42c222229eafafecbbe72129e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610097
Commit-Queue: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Hitoshi Yoshida <peria@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659871}