Fix memleak in ProfileAttributesStorage's bitmap loading code
Before this CL, ProfileAttributesStorage used fancy pointer-to-pointer
acrobatics to load an image in a posted task and pass it back to the
calling thread. This was a) hard to follow (not obvious who owns the
pointer, and whether it gets properly deleted), and b) actually leaky,
if the whole thing happens during shutdown and the reply never runs.

Since gfx::Image is cheap to copy, this CL uses
PostTaskAndReplyWithResult to just pass it around by value. This fixes
the memleak and as a bonus, makes the whole thing much easier to follow.

Bug: 961069, 953845
Change-Id: I59a86cc96fb2312e9dab91a6b490a53cf3124b9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1611614
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659887}