[Touchless] Update style of touchless cards.
Refactored SuggestionsBinder to have overridable methods that the
touchless variant can use to control the behavior. It is a bit messy
this way because the contextual logic decision is made with a boolean
and if statements inside SuggestionsBinder and we are not consistent
with that approach. However, we did not have that option, because the
touchless code is not to be referenced from the normal Chrome, and I
didn't think it was safe to significantly refactor SuggestionsBinder
into multiple sub-classes, one for Zine and one for contextual.

(cherry picked from commit 8d5ea0efebd05f50bb0c088e0a847a75f219beee)

Bug: 937446
Change-Id: Ic638ddd5161413c2014a973ec40261420e4a7947
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1575434
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Theresa <twellington@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#654281}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1589684
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/branch-heads/3770@{#625}
Cr-Branched-From: a9eee1c7c727ef42a15d86e9fa7b77ff0e63840a-refs/heads/master@{#652427}