[ MimeHandlerView ] Reset PostMessageSupport at Commit
When the RenderFrame is ready to commit a new navigation, the
MimeHandlerViewContainerManager should reset its PostMessageSupport.
This is crucial when navigations commit in the same RenderFrame,
, i.e., setting a frame's URL from a.com/pdf1.pdf to a.com/pdf2.pdf
which would end up committing in the same RenderFrame.
With this fix, all PDFExtensinoLoad.* tests should now pass under the
MHVICPF flag.

Bug: 961786
Change-Id: Iaa9e897bb17d12e039fcc346623c0bec5938a8c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1611876
Commit-Queue: Ehsan Karamad <ekaramad@chromium.org>
Reviewed-by: Ehsan Karamad <ekaramad@chromium.org>
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659928}