Force ImmutableCSSPropertyValueSet to be properly aligned.
For the pointer arithmetic to produce aligned results, the struct itself
must have a size which has alignment at least as strong as the types that
are to be stapled to the end.

This is presently not guaranteed, since ImmutableCSSPropertyValueSet
actually happens to have no pointer members (or anything else that would
force 8-byte alignment on 64-bit).

Change-Id: Ie7536949abe5919e1ccc374a206321c26da6474c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1613231
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659933}