[Speed Waterfall] Schedule display_locking on bot platforms.
Since the benchmark is fairly short, I just manually added it
onto a shard that was supposed to be shorter. We can reshard
later to rebalance after this benchmark has run enough to
generate official timing data.

This is the first benchmark to explicitly only schedule on
a couple platforms, instead of defaulting to run on everything.
The plan is to move more benchmarks to this pattern in the
future in order to reduce load on platforms that we don't
have many devices for. (See full list here:
https://docs.google.com/spreadsheets/d/1wm6DNM9Zyr8Q_Z0ffjrycRTD5_bRP0vGob6p6kbLVSA/edit )


TBR: jbudorick@chromium.org
Bug: 961830, 961322
Change-Id: I9b80da77e1d945c82809487a02039c3370d43693
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1611313
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#659983}