Use TestGpuServiceHolder instead of TestGpuThreadHolder
This CL supports using TestGpuServiceHolder as a singleton.
One advantage of the singleton is that the SharedImageManager used
by different clients (e.g. InProcCmdBuffer and
SkiaRenderer) is common. This is a prerequisite for the clients
to share SharedImages between each other.

I will try to remove TestGpuThreadHolder in a separate CL.

TBR=rockot@google.com

Bug: 962481
Change-Id: I7ce4b802d58fde3080472bed324e301a06efe578
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610848
Commit-Queue: Jonathan Backer <backer@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660106}