[AF] More flexible autocomplete attributes
If the autocomplete attribute != off/no/false, and it's not equal to
any value predefined by Chrome, the field would not get autofilled.
This change tries to cover cases where the developer is trying to
give a hint to the af, but may make a mistake.
As in using "phone" instead of "tel".

Fixes: www.bluehost.com (phone and phone extension)
Bug: 962289
Change-Id: I395c4eab60801c2bcd67733ed0009f0a6a89df7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610716
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Parastoo Geranmayeh <parastoog@google.com>
Cr-Commit-Position: refs/heads/master@{#660150}