Add device parameter to scroll injection API
The newly introduced mechanism for injecting scroll gestures on the
main thread can also be extended to be used for keyboard scrolls.
However, we don't want these to show up as scrollbar scrolls (which the
API currently is hardcoded to assume).

Instead pass the WebGestureDevice along with the injection request.

Bug: 954007

Change-Id: If550d39158bc6a161d17ca18733b0ac94b850ab1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600483
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660193}