[Ephemeral Tab] Make ContextMenuTest more robust.
Makes the tests in ContextMenuTest handle the Ephemeral Tab being
either enabled or disabled.  This allows the tests to pass regardless
of whether the feature is enabled or not.  The Official build does
not have the feature enabled, so these tests are failing there.

BUG=938308

Change-Id: I66df69d78fc3c7bb24160fb4b77981e39a3e6230
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1613958
Auto-Submit: Donn Denman <donnd@chromium.org>
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Commit-Queue: Donn Denman <donnd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660198}