Move SuspendMode handling down to the socket level.
Entering suspend mode causes live sockets to hang on OSX and Windows
(Unclear what happens on other platforms). For brief suspends, it may
also be possible to get an ERR_CONNECTION_CLOSE, which is also a
problem, as it has semantic meaning for HTTP connections.

Previously, URLRequestJob watched for entering suspend mode, and
canceled the request when it happened. There are three problems with
this:
* Not all requests use the network.
* Other types of TCP sockets didn't get to take advantage of this logic.
* It cancels requests no matter where they are, and informs the caller of
  this. This has caused at least 4 or 5 crashers in the past, since it
  even informed callers when they weren't currently waiting on a
  notification from the URLRequest.

This CL moves the logic down to TCPClientSocket, and only calls back
into the caller with an error when the caller is expecting it.
One potential downside of this approach is that this doesn't catch DNS
lookups (At least when using UDP sockets, or GetAddrInfo), which may hang
when changing networks, too, though there are already timeouts and
retries at that layer, so there shouldn't be too big a regression
there (hopefully...).

Another change is that will mean the suspend logic no longer affects
live QUIC requests. Since HttpNetworkSession::CloseIdleConnections()
already doesn't close idle QUIC sessions on suspend, I assume we're
fine there, though it's possible live QUIC sessions have issues
related to suspend mode that idle ones did not have.

Bug: 651120
Change-Id: I7ce56ea520b7f3672c4a9593ded3cf1e44d09abd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610496
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660223}