fido: disable resident keys for Touch ID
This changes MakeCredentialRequestHandler to reject requests with
rk=true to the Touch ID authenticator, without dispatching them. The
authenticator itself is also changed to reject such requests should it
receive them, and to not return an arbitrary credential when receiving a
GetAssertion request with an empty allow list.

Bug: 941120
Change-Id: I7d70b2baaca66611ee8957445e54b17b91b6cc8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1600735
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660238}