Report tab switch failures to UMA.
The duration between when a tab switch is requested and a frame is
presented on the screen is recorded to the
Browser.Tabs.TotalSwitchDuration.* histogram.

We looked at traces collected via the "slow report" pipeline and
found failure cases that were not properly reported to UMA:

  Case 1:
    1. User switches to tab A.
    2. After a long time (e.g. a few seconds), a frame has still not
       been presented for tab A, so the user switches to tab B.
    3. User switches back to tab A.  This time a frame is represented.

    The first time the user switched to tab A, they had a bad
    experience, but this isn't reported to UMA.

  Case 2:
    1. User switches to tab A.
       A frame is not presented (kFailure flag is set in
       gfx::PresentationFeedback), but a duration is still reported to
       the Browser.Tabs.TotalSwitchDuration.* histogram.

This CL introduces Browser.Tabs.TabSwitchResult.* to report the result
of a tab switch. It ensures that only the duration of successful tab
switches are reported to Browser.Tabs.TotalSwitchDuration.*. It reports
the duration of incomplete tab switches to
Browser.Tabs.TotalIncompleteTabSwitchDuration.*. It also changes trace
events so that tab switches that did not succeed are properly
represented.

The goal of this CL is to make it easier to understand the impact of
tab freezing on tab switch time (it currently has an impact on
Browser.Tabs.TotalSwitchDuration.*, but it's hard to make conclusion
based on that observation given that
Browser.Tabs.TotalSwitchDuration.* includes failure cases).

Change-Id: I0b0c66f711a0389e6d8519b9a176ade043334f08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1592052
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660264}