Add recommendable / searchable to the App Service
In particular, the default values for these bool-typed fields is true /
true for almost all apps, but some built-in Chrome OS apps are
recommendable but not searchable, or vice versa.

Also add an AppUpdate unit test for the show_in_launcher field. That
field already existed, prior to this commit, but for some reason, it was
overlooked in the tests.

This fixes `unit_tests --enable-features=AppServiceAsh
--gtest_filter="AppSearchProviderTest.FetchRecommendations"`. It fails
before and passes after this commit.

Note that `unit_tests
--gtest_filter="AppSearchProviderTest.FetchRecommendations"`, without
AppServiceAsh enabled, passes both before and after this commit.

BUG=826982

Change-Id: Ifa748f98baca513e23e3c3df92d9f5379c1c5e70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1611433
Commit-Queue: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660267}