[wpt-sync] Fetch the most recent GitHub PRs
Without the `sort=updated` API param, we relied on the magic "best
match" algorithm of GitHub
(https://developer.github.com/v3/search/#ranking-search-results).

This was fine when we had <1000 (the maximum number of search results)
export PRs, as all of them would be included in the response. Now that
we have more than 1000 export PRs, we need to explicitly get the most
recently updated ones to make sure export PRs for recent commits are
included in the response.

Change-Id: Idff29b23c46d72a3df927a53d96dad4340c42d84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1613701
Commit-Queue: Luke Bj <lukebjerring@chromium.org>
Reviewed-by: Luke Bj <lukebjerring@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660367}