Phase out CrossThreadBind (3/n)
This CL is the third out of N CLs that phase out the use
of CrossThreadBind in favor of CrossThreadBindOnce.

Similarly to the first one [1] and [2], in this CL the uses of
CrossThreadBind that are being phased out are the ones used
combined with PostCrossThreadTask.

Reasoning: when a task is posted, it only executes once. Hence it is
logical to cross-thread the callback once.

[1] https://crrev.com/c/1615444
[2] https://crrev.com/c/1615446

R=haraken@chromium.org, hiroshige@chromium.org

BUG=963574

Change-Id: Ibf11b6ee52950c08f31ba77431c67b8ab630afaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1615447
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Antonio Gomes <tonikitoo@igalia.com>
Cr-Commit-Position: refs/heads/master@{#660531}