[Fuchsia] Enable OOP-R when Vulkan is supported
Support for OOP-R for GL is determined by SupportsOOPRaster in
gpu/config/gpu_info_collector.cc. This checks that GL has the extensions
the Skia needs to create a GrContext.

This code is not applicable to Vulkan. We will only return a valid
VulkanInProcessContextProvider if Skia is able to create a GrContext
with Vulkan.

Bug: 962617
Change-Id: I01783c8dbb88538bc6e3c2f7c348c84c7d7ff195
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1615369
Auto-Submit: Jonathan Backer <backer@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Commit-Queue: Jonathan Backer <backer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660560}