[omnibox] Create OmniboxFieldTrial::GetValueForRuleInContextByFeature
This CL makes a base::Feature based equivalent to
GetValueForRuleInContext.

We do this because in the near future, we want to be able to use our
page-context based rules system without necessarily being tied to the
bundled field trial.

A followup CL will migrate the ZeroSuggestProvider to use
GetValueForRuleInContextByFeature for the ZeroSuggestVariant parameter.

Bug: 963173
Change-Id: I00ae9cb8507b34d386fffd67dc447d5a3f0c61d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1613791
Commit-Queue: Tommy Li <tommycli@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660587}