[Autofill] Logs how often a formatter fails to make the suggestions.
When the use improved label disambiguation desktop experiment is
enabled and profile suggestions are available, the assumption is that
this will not--or will very rarely--happen.

However, bizarrely-structured forms exist, so it is not impossible that
the experiment is enabled, suggestions are available, and a formatter
failed to create them.

This logging is a sanity check to verify that this rarely happens.

Screenshot of chrome://histograms
https://screenshot.googleplex.com/OTSdHFUKmmM

Bug: 961334
Change-Id: I471e9c6078f11d46a972eaa2486d43b14830124f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1603242
Reviewed-by: Nik Bhagat <nikunjb@chromium.org>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Cr-Commit-Position: refs/heads/master@{#660600}