Create quarantine service
Creates a quarantine service to be used to perform file scanning
/ invoking attachment services.

This change creates the service but does not implement the quarantine
logic.

A follow up change will move quarantine implementation from
components/services/download to the service,
and add the logic to call the service.

Bug: 870998
Change-Id: I21a06a98bed59ce01774d4da5b8319c1dfe7b55d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1595156
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Alexander Woolf <alewoolf@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#660689}