OOR-CORS: OriginAccessList cleanups
This patch changes the data structure to have the allow/block lists.
Also, modifies the interfaces not to have list specific clear methods,
but just have them to clear both lists together.

Still |Patterns| isn't the best structure to have multiple patterns
for the same origin with priorities, but since matching rule is not
simple, and will be changed to count port number, it is not updated
in this patch.

Bug: 908756
Change-Id: Ie148b5f35c66968fb2a1c8a9332a1442b6091586
TBR: tsepez@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1614380
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660779}