Add AdFrameLoad Ukm event to AdsPLMO
Add a UKM event with metadata for each ad frame observed on a page load.
Records overall performance statistics (bytes/cpu), as well as metadata
such as whether the ad is cross origin, it's size, etc.

This will greatly simplify the number of metrics we are recording as
we will not need to split out all UMA by visible/non-visible/etc. These
metrics will prevent the need for new UMA to address the impact of ad
interventions.

As a lot of this UKM duplicates existing UMA, so we will reuse the existing
tests. Once these metrics are landed, duplicate UMA metrics can be
removed on a case by case basis depending on usefulness/simplicity.

Bug: 951480
Change-Id: I19c2a63f496fd84250e0318491da1753ca9d2771
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1562619
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#660988}