vk: Start the VK in INITIAL rather than UNKNOWN state.
We currently start the VK in an UNKNOWN state. When we enable the VK,
we first transition to INITIAL and then immediately transition to
LOADING_EXTENSION. When we disable the keyboard, we reset to the INITIAL
state.

This is kind of weird. We might as well start at INITIAL state and
transition to LOADING_EXTENSION when we enable the keyboard.

The only impact would be on the UNKNOWN -> INITIAL transition for
metrics, but the number of UNKNOWN -> INITIAL transitions should be
the same as INITIAL -> LOADING_EXTENSION anyway.

Bug: 845780
Change-Id: I39c6190a73daf9d9b4349916bfc0d1c825812ab1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1616824
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661033}