[ozone/wayland] Shm: Proxy: make mojo calls on the gpu thread
This is a prerequisite to unite the buffer managers and
share most of the buffer management APIs between software and
hardware paths.

When software compositing is used, make sure that the calls to mojo
happen on the gpu thread. Otherwise, OnSubmission and OnPresentation
may happen on a completely different thread and hit the DCHECK.

Bug: 963853
Change-Id: I537b7734f88dab23a9f97897ffd48e07bc36a537
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1617502
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#661180}