Percentage height becomes auto if containing block height is auto.
It's wrong to walk all the way to the root and look for something with a
fixed block size just to determine whether we can collapse through the
block or not. It only matters what the actual block size of the
containing block is.

This bug triggered DCHECK failures, because the cached state of
self-collapsing got out of sync with reality, because some ancestor went
from auto-height to fixed-height, a change that shouldn't affect the
descendant at all.

Bug: 962175
Change-Id: Ic24d9c46dfa515306563b195487310eedfae9ac3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1617833
Commit-Queue: Emil A Eklund <eae@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661380}