Initial (Prototype) Implementation of HTTPS image redirects.
Uses the ResourceRequest pointer in URLLoaderThrottle to redirect image
subresources on a page to a compressed version of the same resource.
This initial implementation will return a 200 when it successfully swaps
compressed version and a 307 when it needs to fall back onto the
original uncompressed resource.  It currently only handles those 2
scenarios.

This is a prototype implementation and is behind a flag.

Bug: 960513
Change-Id: I1d47e9f958a92e1ffb970d6852cdcab2acd9e698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1606220
Commit-Queue: Sean Harrison <harrisonsean@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661421}