Add metric to track momentum scroll input jank
Currently, momentum scroll input is provided once per frame. In cases
where the input is processed early (twice in a frame), we will skip
input in the next frame, leading to jank.

To track this we look at the number of times we coalesce gestures
and compare this to the expected number of gestures per frame.

To further isolate issue 947717 we also look at cases where we
coalesce a gesture near to its processing time.

Bug: 952930
Change-Id: Ic685109f518c2cd5ce6941558b606f5f04f2c192
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1580420
Reviewed-by: Nik Bhagat <nikunjb@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Eric Karl <ericrk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661474}