Delete group menu button from TabGroupHeader.
The current plan is to use a right-click context menu for the first
pass, which means the current group menu button is no longer necessary.
This also fixes an issue where browser tests could not tear down if a
group was active, because the group menu button didn't have the correct
accessibility settings.

Bug: 905491, 946263
Change-Id: I3f60f3d23686021bd777c45913930158578f2d25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1621245
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661510}