GWP-ASan: Refactor GuardedPageAllocator free lists
Refactor how we manage free lists in GuardedPageAllocator to use a
shared SimpleFreeList data structure. This is in anticipation of adding
a special PartitionAllocFreeList that ensures that once a page has been
allocated for one partition it is never allocated in a different
partition.

Bug: 956824

Change-Id: Iea75a625c341339a974f319e098c1529cbb9370c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1609603
Reviewed-by: Vitaly Buka <vitalybuka@chromium.org>
Commit-Queue: Vitaly Buka <vitalybuka@chromium.org>
Auto-Submit: Vlad Tsyrklevich <vtsyrklevich@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661563}