Move most WebGL 2.0 Compute stuff to its own class
When something in WebGLRenderingContextBase or
WebGL2RenderingContextBase needs to be extended for WebGL 2.0 Compute,
the CL favors overriding it in WebGL2ComputeRenderingContextBase rather
than doing that in the base class with a check on the context type. Not
sure about the performance impact though.

Bug: 859249
Change-Id: I17dbdf1bc3c02ecb4d5994555fe9787a47aa750e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1612156
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Jiajie Hu <jiajie.hu@intel.com>
Cr-Commit-Position: refs/heads/master@{#661574}