[Autofill Assistant] Disable action carousels change animations.
This CL disables change animation in the action carousel, to avoid chips
blinking when setting the same set of actions (e.g. when changing a form
or payment request value).

In practice, we don't lose any animations as we notify the adapter of a
change only when the text and type of chip is unchanged. Addition and
removal of chips are still animated.

 * Before: http://go/aa-chip-blink-before
 * After: http://go/aa-chip-blink-after

Videos: 
Change-Id: Ic7e02b138a8da685fce86982a09f402941e3f5f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1620631
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Jordan Demeulenaere <jdemeulenaere@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661677}