Fix (hopefully) and reenable BookmarksTest.testDisabled*
Before, ProfileSyncServiceAndroid::TriggerRefresh requested a refresh
only for the *active* data types. If the SyncService happened to be
reconfiguring at that point in time, the set of active data types was
empty, so the refresh request was a no-op. This caused the Java code
that triggered the whole thing to wait forever and eventually fail.
Since there is no harm is requesting a refresh for types that aren't
enabled, now we just ask for a refresh for all types.

Bug: 964342
Change-Id: I23c1f7fedce7ed90d56e7feb32aa62d1072ed0a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1621175
Auto-Submit: Marc Treib <treib@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661744}