Increase the Chrome OS cookie merge session timeout to 20 sec.
This is a speculative fix for an issue where a user logs
into a Chrome OS device (which authenticates) and then navigates
to a Google site only to be asked to login again. The theory is
that slower devices may actually need more than 10 seconds for
the merge session to finish.

Bug: 960889
Change-Id: I22fb835681abff1cd5a79338cd796d466483137e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1621066
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Chris Mumford <cmumford@google.com>
Cr-Commit-Position: refs/heads/master@{#661814}