Re-add the base::Bind/Callback/Closure presubmit on a list of paths
I have gone through the code with the following script, delving into
sub-directories where it seemed appropriate, to build a list of places
that still use base::Bind/Callback/Closure and need to be migrated. The
PRESUBMIT check will ignore those places to avoid warning fatigue and
people using BindRepeating randomly and incorrectly.

This allows the check to continue to be run against places that have
been converted and are now clean (such as cc, components/viz, ui/views)
to prevent the addition of new uses there.

It also gives a convenient burn-down list of places to fix, and they
can be removed as they are done so.

Uses .join() because just giving the whole list makes the presubmit
check 7% slower against one c++ file.

R=ajwong@chromium.org, dcheng@chromium.org

Bug: 947414
Change-Id: I7cda063a0e02c6b4a18e8229093af6d48dfe7fa7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1622869
Reviewed-by: Albert J. Wong <ajwong@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#661906}