Merge M75 [ios] Improve page placeholder placement
This merge CL combines two trunk CLs.

1. [ios] Improve page placeholder placement
Previously, the grey page placeholder image attempted to use the
webProxyView's contentInset property to properly position the
placeholder. This used frame setting.

The preferred way to position content in the BVC is to use the
ContentArea NamedGuide with autolayout contraints.

Using the preferred autolayout + NamedGuide approach also fixes a
placeholder malpositioning bug related to fullscreen experiments.

2. [ios] Fix page placeholder crash on restore
Adding the page placeholder to a view requires that view to be in a
view hierarchy with the Content Area named guide.
Prior to this CL, the placeholder was attempting to be added to the
web view before the web view is visible. This CL adds a check to see
if the web view is visible. If not, the placeholder will be added when
the web view becomes shown.

TBR=eugenebut@chromium.org
(cherry picked from commit 38703614b353273ac8f5629eeecc9f983013b40d)

Bug: 956077
Change-Id: Ia3d115ae203bacb59d41cdb3c377a4a75c00dc95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1593921
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#657797}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1623548
Cr-Commit-Position: refs/branch-heads/3770@{#760}
Cr-Branched-From: a9eee1c7c727ef42a15d86e9fa7b77ff0e63840a-refs/heads/master@{#652427}