[ios] update suggestions after password generation
Force a focus event (js) which will propagate back to the suggestions mediator who
will update the suggestion bar.  Applied to both Cancel and using suggested password.

Removed call to completion when suggesting a password, because form filling is
not complete, and we want the keyboard to stay (or come back) up.

Bug: 964941
Change-Id: I31900792b0c3452913cb71c31a6c89b502a0b27b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1620640
Commit-Queue: David Jean <djean@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662107}