service worker: Simplify the force delete operation.
DeleteVersion() was a function called when a service worker script
couldn't be read from disk. To avoid a sticky state where in-scope URLs
always fail to load, we remove the registration from storage and drop
the active version as a controller of any clients.

However it was trying to do a clever thing and allow any in-progress
register jobs for that scope to continue as normal, and also try to keep
the registration intact if there was a waiting version that wasn't
broken. This was causing complexity and wasn't reliable as per the TODO
in the code. It also wouldn't retain any user data attached to the
registration. It's simpler to ignore these edge cases and simply always
remove the registration.

Bug: 964201
Change-Id: I2e08bb2cd9faf7ec071bf7ae63e40540933aa462
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1621473
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662117}