Media Intent Handler: Disable audio handling on Android Go
This CL creates a new AudioLauncherActivity alias to the
MediaLauncherActivity that handles audio intents (while the main intent
handles video and images). We disable AudioLauncherActivity on Android
Go while leaving it enabled on enterprise-managed devices. This allows
us to continue to handle video and image intents on Android Go without
breaking the memory restrictions imposed on music players on Android
Go.

TESTED=Able to open video/audio/image files with Chrome on an
enterprise-managed device. Able to open video/image files with Chrome
on an Android Go device, but not given the option to open audio files
with Chrome.

Bug: 965166
Change-Id: Ia2b774a048ac430b658c29b8fb11bf8835709842
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1620610
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662195}