Reshow password notifications if needed
If a notification is dismissed but not dealt with,
then show it again after 1 day.

If a notification is not dismissed, and password is not changed,
then notification is updated after one day, since the text showing
the how urgent it is should change.

Deleted some code that special cased the "expired" case -
this case actually has the same logic as "about to expire" case,
so it doesn't need a special case.

Added more unit tests for SAML password expiry notification.

Bug: 930109
Change-Id: Ie035ec788ee083d07261bca0e19f216e6bc7d6cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1620642
Commit-Queue: A Olsen <olsen@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662204}