[Chromoting] Update mouse clamp offset fix
Move the fix after the show_display_id_ has been set so that the correct
display origin is used for the mouse clamp filter.

This updates the comment to include more information about why we need
to update the mouse filter separately when switching between displays
that are the same size.

Change-Id: Idcd7618a2f8d52d090c60ad92abefbd9069cc2be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626391
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Commit-Queue: Gary Kacmarcik <garykac@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662351}