Make XR test input runtime-agnostic
Changes all the button and axis IDs used for sending controller input
during the XR browser tests to be runtime-agnostic instead of using the
OpenVR constants.

This fixes the issue of WMR having to duplicate the constants since it
can't rely directly on OpenVR, and also simplifies things by using the
notion of primary/secondary axes instead of having to conditionally
change which axis type we use based on runtime.

Bug: 965306
Change-Id: Ide659c0c25958c1a732a5f893ce01a9cd88499e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1625933
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662458}