Remove old Windows *.TMP files.
When renaming a file, Windows will use an intermediate name and leave
that file around if the final destination name already exists.  These
files are unknown to Chrome and so tend to accumulate which is a
problem when the files are 1MB "metrics" files.

Bug: 934164
Change-Id: I30fa0d81db81af411d7c8775f389206904062412
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1611104
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662473}