[Autofill Assistant] Don't clip sheet content if it is not scrollable.
This CL disables the clipping of the bottom sheet content if it is not
scrollable. This allows to correctly display the details shadow if it is
the only visible element in the scrollable area.

We still need to enable the clipping of the content if it is scrollable,
otherwise scrolled content will overlap the carousel and header (given
how the ScrollView works internally). This is fine as the details are
usually shown at the top of the scrollable area, so clipping it when it
happens shouldn't have an impact on the details shadow.

Given that the actions carousel was relying on the clipChildren flag to
be disabled, the spacing before and above the carousel has been moved
into the carousel item decorations.

 * Before: https://screenshot.googleplex.com/p6DUiOBnyDp.png
 * After: https://screenshot.googleplex.com/AAKU42uvhgv.png

Screenshots: 
Change-Id: I900f47ab41a36cb3f2140d1387671e392d694d42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1624194
Commit-Queue: Jordan Demeulenaere <jdemeulenaere@chromium.org>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662543}