Fix and re-enable null VR pose tests
Fixes and re-enables the three tests in WebXrVrDeviceTest that were
disabled due to failing 100% of the time. The cause as a string in the
HTML file that they all load that was missing its closing quote, which
was causing issues when trying to load and run JavaScript on the page.

TBR=billorr@chromium.org

Bug: 966254
Change-Id: I89bb21bea7f539c976f9c90f63635951418d64a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626743
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662723}