[LayoutNG] Ensure that new formatting-context margins separate correctly.
When positioning a new formatting-context in the presence of adjoining
floats sometimes a "forced" BFC block-offset will be set twice.

When this new formatting-context was wrapped in a <div> (a parent which
hadn't had its BFC block-offset resolved yet) we ran into a bug with our
state machine.

Upon the second time resolving the BFC block-offset the parent would
resolve *its* BFC block-offset to the previous "forced" BFC
block-offset.

Upon the final relayout the parent would set the first "forced" BFC
block-offset to its children causing an "infinite" loop (if we had a
loop driving this machinery.

This ensures we use the child's new BFC block-offset in this case.

Bug: 966278, 966214
Change-Id: I5da2f3e90b9d2ff78f124247f6f68cd3a289ae42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1627692
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662763}