Persist histogram data when Chrome is running in the reduced mode.
Recently we noticed that the histogram data is lost when Chrome is
running in the reduced mode and get killed before the full browser
starts. This is because there isn't a SystemProfile set in the
reduced mode. Thus, the previous data in the BrowserMetrics-active.pma
file isn't picked up by Chrome the next time it is launched in the
full browser mode.

In this CL, we write the initial system profile to the mapped memory
file, so Chrome will pick up the data when it restarts in the full
browser mode.

Note: the histogram data is only persistent on the memory mapped file,
but not being uploaded when Chrome is running in the reduced mode.

Bug: 959854
Change-Id: Idc1cea6edfa854edf94f791f2df8a9ebb6a0af8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1601307
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Commit-Queue: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662815}