Remove Synchronous Child LocalSurfaceIdAllocation in CC
In order to support Mus, we added a series of methods to allow Aura to request
the allocation of a new LocalSurfaceId. This allowed both aura/mus and cc/ to
generate these Ids. As Aura was not responsible for ParentLocalSurfaceId allocation,
the Window Service was.

With the removal of Mus, this code path is no longer needed. This change reverts
that path. Leaving cc/ solely in charge of child allocation.

This change is a merged revert of four changes:
Revert "chromeos: gets aura as mus client library working with client ids"
This reverts commit 2b5a978b9351a8108cade7a003653d74b83f2ab8.

Revert "chromeos: gets aura as mus client library working with client ids"
This reverts commit 2b5a978b9351a8108cade7a003653d74b83f2ab8.

Revert "cc: provides function to synchronously allocate LocalSurfaceIds"
This reverts commit 39db70bbbaf815031d88305000ae8cee1c7264ac.

Revert "Call function when LayerTreeHostImpl generates a LocalSurfaceId"
This reverts commit 7a340f4091c6d1ec256938d3ca9d2d9f6395aacf

TEST=cc_unittests

Bug: 958239
Change-Id: Ie81d68f9dc6a6ff0bc57acccc5e9182bf819af95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1625394
Reviewed-by: Saman Sami <samans@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662833}