Expose inlineVerticalFieldOfView on XRRenderState
The code has supported updating inlineVerticalFieldOfView for a while,
but has always used a hard-coded default.  This change exposes the
ability for a webpage to update or read this value, and changes the
default value to match the specced 90 degrees (pi/2) requirement.

As part of this change, magic-window is also updated to adjust this
value.

Bug: 953412
Change-Id: I583935c487ee58a0a588e5161e66976a3c0e9675
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1625924
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Bill Orr <billorr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#662851}