Don't modify the response when creating a ConvolverNode
When creating a ConvolverNode with normalization (the default), the
ConvolverNode was modifying the supplied AudioBuffer containing the
desired impulse response.  This happens because the normalization
factor was applied to the buffer, convolver stages were computed, and
then the normalization was undone.  Due to floating-point roundoff,
this can (and does) modify the buffer.

Instead, compute the scale factor and pass it into the various
routines so that the scaling can be done there.  For stages using an
FFT, the resulting FFT is scaled by the factor.  For the direct
convolver, the response is modified there.  (The direct convolver
makes a copy of the reponse, so this is safe.)

This requires adding a new method, FFTFrame::ScaleFFT, to do the
scaling because only the FFTFrame knows the size of the underlying
arrays.

Manually verified that the test fails without this fix and passes with
this fix.

Bug: 961911
Test: unmmodified-buffer.html
Change-Id: I36a8ef9ac7751874bd82551f27ded707aca46ef1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1623457
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663200}