Enables new automatic clicks UI.
Removes need for automatic clicks feature flag to see new UI in
preparation for M76 launch. Does not delete the automatic clicks
feature flag because it will be used to hide a new 'scroll'
feature until launch in an upcoming change.

The bulk of this change is in removing old code to draw the old
automatic click ring widget, which will no longer be used in M76+.

Bug: 952476
Change-Id: I614d699527a46e83039802f3e72349a9ad39a6c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626399
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663216}