overview: Show drop target even if split view is disabled
The purpose of the present CL is to add a drop target in a certain
workflow. As of [1], the "drag to snap" mode has become a "normal drag"
mode which is available even when split view mode is not available. For
testing, you can note that in clamshell mode, with the functionality of
[2] disabled (the current default), split view mode is not available,
but you can still do a "normal drag" from overview. There is presently
no drop target in that workflow. There is no drop target depicted in [3]
either, but I am pretty sure a drop target is actually appropriate,
because of [4] (which is assigned to me) if for no other reason.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1612121
[2] https://crbug.com/890029
[3] https://bugs.chromium.org/p/chromium/issues/detail?id=866622#c33
[4] https://crbug.com/961170

Test: manual
Bug: 866622, 961170
Change-Id: I079d01c3e449b10f23a19a9993fbdd44d0c1482e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1627562
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663305}