Add prototype panel entry for feedback panel.
Implementation of a custom element representing a feedback panel
entry that can be used stand-alone or stacked to display the
status of file operations in progress in the File Manager.

This CL implements a prototype of the panel that integrates
a progress indicator, and has placeholders for text content
and action buttons.

Bug: 947388
Change-Id: I1f910260ea91d5366cc3c31f19dc234ff723f3d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1614826
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663463}