Support null timeline and start time.
Adds support for null timelines. Having a null timeline also requires
support for null start times, as only one of the current time and
start time can be resolved if the timeline is null.  This patch is a
precursor to adding support for mutable timelines.

Bug: 771751, 771722
Change-Id: Ibadf40ed146117d88cf5f55085116e8664715191
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1615314
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663551}