Allow path removal in favicon server request and check feature eagerly
We add a parameter to kEnableHistoryFaviconsGoogleServerQuery that
allows FaviconRequestHandler to query the favicon server by an url
with its path stripped. We also check the feature earlier in the
public methods and plumb a boolean, for experiment reasons. A typo in
the feature name is also fixed.

Bug: 955475
Change-Id: Ia6ca88d671567b82ea980c8ad4028766aed913ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1624814
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Jan Krcal <jkrcal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663561}