Generator for creating fuzzable protobuf files.
As part of the wayland/exo fuzzer, this generator will be used to
create a protobuf spec for the given wayland protocols. The protobuf
defines a message for every request, with the intention that libfuzzer
will be making message sequences that the wayland server should be
able to deal with.

See crrev.com/c/1627839 for a mockup of what the protobuf file looks
like.

Bug: 961564
Change-Id: Ib6ebc303d51b97ac3ca6f9f417c7b1125920f7c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1631033
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Cr-Commit-Position: refs/heads/master@{#663634}