Don't use |fetch_request_windows_id| if a ServiceWorker controller is unavailable
The CL https://crrev.com/c/1290435 introduced the |fetch_request_windows_id|
for ServiceWorkerProviderHost, however, it looks this token is being used in a
request where a ServiceWorker controller is not being involved.

Setting |fetch_request_window_id| should not affect correctness, however, we have the
extensions bug, https://crbug.com/963748, which we don't yet understand, where a
ServiceWorker is not being used in a test.

That is why we are making this change, at least, until we can fix the extension bug.

Bug: 623464,963748
Change-Id: I35a6ef80fbab42d4a00b6b881026518e06b4eb90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1616712
Commit-Queue: Hayato Ito <hayato@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663643}