[Autofill Assistant] Fix peek height when peekMode shows carousels.
This CL fixes the height of the BottomSheet when minimized and
peekMode=HANDLE_HEADER_CAROUSELS. Before http://crrev/c/1624194, the
vertical spacing before and after actions were set as layout margins
(i.e. it was not considered when computing the actions height), so we
had to add this spacing when computing the peek height. After that CL,
the spacing is contained into the actions height so there is no need to
add it when computing the peek height.

Change-Id: Ic83a983c085d2dc7e2d91515c1d7389eb004c955
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1631373
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Jordan Demeulenaere <jdemeulenaere@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663680}