base/android: Enable ordered prefetch on low-end Android devices only.
From the field trial (see
https://bugs.chromium.org/p/chromium/issues/detail?id=758566#c71 for details),
ordered prefetch improves startup performance significantly on low-end devices,
and regresses it on high-end ones.
As a consequence, enable only on low-memory.

From the field trial, this is time to commit for cold starts:
- Devices with >=2GB of RAM: Regression of 110ms (5.44%) at the median, 175ms
  (2.96%) @95th percentile
- Devices with <=1.2GB of RAM: Improvement of 135ms (2.71%) at the median, 650ms
  (4.34%) at the 95th percentile

Also removes the command-line switch and the synthetic field trial setting. As
such, this is partially a revert of
https://chromium-review.googlesource.com/c/chromium/src/+/1270936.

Bug: 758566
Change-Id: I3a95f97bd5c9eda7840a107fa22b7d016a4a08c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1631380
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Robert Kaplow (slow) <rkaplow@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663738}