Quota: Remove available disk space from host quota calculation
This change alters how per host quota is calculated. Specifically, there is a case where users have low disk space availability where the quota is based off available disk space. This change ignores that branch of logic for users that have the StaticHostQuota Finch study enabled. The study is disabled by default, but will be enabled for 100% of users, starting with Dev/Canary and moving up the channels to Stable. Keeping this behind a Finch study allows us to disable this change with a single line change to Finch.

Bug: 965710
Change-Id: Icd5ed59c1c592ccd9554395427f6707210376271
Signed-off-by: Jarryd Goodman <jarrydg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1625744
Reviewed-by: Nik Bhagat <nikunjb@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663795}