Fix page scale limits override
In https://crrev.com/53f9afcdecf35a8, we changed the Force Enable Zoom
setting to also force reset the initial scale factor so that pages that
specify an initial-scale in the viewport meta tag load fully zoomed out.

Unfortunately, this caused the bug in 960518 since we always reset the
initial scale factor to -1 when calling SetIgnoreViewportTagLimits. On
AndroidTV devices the initial scale is overridden using
SetInitialPageScaleOverride. This override would be clobbered when
SetIgnoreViewportTagLimits was called (with false, to ensure the
override is off).

This CL fixes the issue by storing the user agent constraints when we
override them, and resetting them to the original values when turning
off the override.

Bug: 960518,929709
Change-Id: I9d7236ec0991025067504c867fe0d06c9d12c455
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1631710
Commit-Queue: David Bokan <bokan@chromium.org>
Auto-Submit: David Bokan <bokan@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#663939}