Update AXPlatformNodeTextProviderWinBrowserTest.GetVisibleBounds
I wasn't able to reproduce the failure, but according to the error
logs they suggested that for some reason, word-wrapping was likely
different from public Chromium and closed Chrome.

This could be something subtle like differences in fonts,
scroll bars, or potentially DPI differences in the test environment.

Updating the test to avoid these types of differences.
Adjusted the text so that line breaks will be more predictable,
leaving ample room for and scrollbar differences.

The error was :
---
error: Expected equality of these values:
  L"G H I J K L "
  provider_content
    Which is: L"G H I J K L M "
---

Bug: 966344
Change-Id: If978ce1fe03984c64151241e6241cdfc29d900c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1627927
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Adam Ettenberger <adettenb@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#663961}