Fix bug in RRectF constructor and mojo unpacking
gfx::RRectF is a wrapper for SkRRect. SkRRect normalizes the corner
radii values if the sum of 2 corner radii is larger than the length of
the side between these two corners. This means that the constructor
cannot initialize a rounded rect by sequentially setting the corner
radius for each corner. All the corners have to be set atomically using
the API provided by SkRRect so that the normalization happens correctly.

Bug: 966488
Change-Id: I0e3f373e4425af905f1be1fc1bc9fda983bdf600
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1629010
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Malay Keshav <malaykeshav@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664000}