Print Preview: Reset custom page range if set automatically
It is possible that by changing various settings the length of the
previewed document can change, which may invalidate the custom page
range. Reset the custom page range to the full set of pages in the
following 2 cases:
(1) The custom page range was set to the full page range in JS, due to
the user blurring an empty custom input. Note: reset even if the
length change would not invalidate the value, so that leaving a
custom input blank matches the behavior of selecting "All" in the
dropdown.
(2) The custom page range is set to the full page range, so it is
clear that the user intends to print all the content, and the length
of the previewed document decreases.

Bug: 938214, 940386
Change-Id: Icdee3970c041f2d269c16967cc07a58514bdf811
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1629045
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664039}