[Mfill Android] Add native bridge code for addresses
This CL wires the address controller up with the manual filling
controller. To do that, it splits the bridge function
"OnItemsAvailable" which is responsible to propagate the items to
the address sheet.
The usage of that split is demonstrated  in https://crrev.com/c/1614202.
The details might change again once the visibility are more fletched
out (see https://crbug.com/965478).

This CL introduces proper metrics for fields that are contained in
sheets.
Minor clean-up: the CREDIT_CARDS suggestion type had no representation
in enums.xml, so I renamed it and put it in there - along with the new
ADDRESS_INFO type.

Bug: 926372, 962548
Change-Id: I93fde325d7a462f74a8c86023d607033ba9a6a51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1631651
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664175}