Do not block Mojo call on waiting for font table, store callback instead
After clarifying with the Mojo team, the initial approach of waiting for
a WaitableEvent before running the asynchronous Mojo callback, this was
identified as the incorrect approach. Instead, store the callbacks until
the font table builder is ready, then run them on the same sequence on
which the Mojo call came in.

Bug: 960263, 889864
Tbr: wfh, rsesek
Change-Id: Ief3fe6a21df28f5aa27a4daa6d65540bddf2b43e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1605985
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664223}