[ MimeHandlerView ] Fix Recording 'PDFLoadStatus'
Before a MimeHandlerViewContainer is created, the chrome layer records
the PDFLoadStatus (main frame vs embedded) in chrome/ layer. This CL
adds the count for frame-based MimeHandlerView. Essentially,

  * The metric is counted in the browser after MimeHandlerView is and
has an embedder frame.

  * Adds tests to verify the metric is collected for both frame-based
and BrowserPlugin-based MHV.

The reason for collecting the metrics in the browser process is that the
method ChromeContentRendererClient::MaybeCreateMimeHandlerView is
currently called twice per PDF load: once when an <embed> is appended to
DOM (with PDF type) (in which case it is navigated to the PDF resource)
and once when the injected <embed> from the browser is updated (this one
is used to attach GuestView).

Bug: 659750
Change-Id: Ide39660464c0e9ab4cc2b9357c4887841df700a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1632426
Commit-Queue: Ehsan Karamad <ekaramad@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664286}