Do not call StaticResource.getBitmap().recycle()
StaticResource.getBitmap() is and should only be called once in its
life cycle. This CL adds an assertion to guarantee this property.

Since StaticResource.getBitmap() would only be called once, it can
release the Bitmap after getBitmap() is called, and it's no longer
necessary to call recycle() on the Bitmap.

Bug: 965580
Change-Id: I070f9530aefea12a6dc72330a10e021103cfdd16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1625901
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664411}