Fix WebApkIntegrationTests with in-process network service
We set the cert verifier on network context unconditionally since the
FeatureList is not set up at the time the MockCertVerifier is set up in
the WebApkIntegrationTests.  This should be fine if the network service
is run OOP or if network service is disabled, since the cert verifier
set as a global in the browser process for NetworkContext will just be
unused in those cases.

Bug: 917467
Change-Id: Id97f877c955959b88f05167c779abfcd3a10f831
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1635951
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Auto-Submit: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664492}