Adding files from a post-branch patch to grd to stop build failures
Context:

In an earlier post-branch patch [1] I added two new files, to the
folder [2], namely [3] and [4]. In my original change on the master
branch [5] I added the new files to the necessary .grd file [6].

However, at the time of the M-75 branch, the "settings split" migration
[7] was in an early stage and the new os-settings page's .grd file [8]
was copied almost verbatim from the existing .grd file [6].

Between that time and my change that added the new files [3] and [4],
the migration removed several files from the new .grd file [8] because
the os-settings page has no depedendence on them. In particular, when I
added the new files [3] and [4] to the folder [2], there was no need to
add them to the new .grd file [8] because all files that depended on
[3] and [4] had been removed from [8] already.

Therefore, when I merged the new change into M-75 in [1], the M-75
version of [8] still contained files that would depend on [3] and [4]
once [1] was merged even though [3] and [4] had never been added to
[8], causing a build failure.

Fix:

This CL adds the files [3] and [4] to the new .grd file [8] and adds a
"preprocess=true" argument to the files that have a dependence on [3]
and [4] to process .html conditional tags because the dependences are
all wrapped in a conditionals checking that the OS is CrOS.

Testing:

I used a set of gn args that caused a build failure on a current
M-75 build (bd0529b42db02c71d24f4f77b637aba46792d5aa) but built and
deployed successfully with this fix, working as intended on a dev
Pixelbook.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1632794
[2] chrome/browser/resources/settings/autofill_page/
[3] blocking_request_manager.html
[4] blocking_request_manager.js
[5] https://chromium-review.googlesource.com/c/chromium/src/+/1574684
[6] chrome/browser/resources/settings/settings_resources.grd
[7] http://crbug.com/950007
[8] chrome/browser/resources/settings/os_settings_resources.grd


Bug: 968328
Change-Id: I2910c7ebd72a09a0ad3c73e2f49d1c451c5b651e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1637604
Reviewed-by: May Lippert <maybelle@chromium.org>
Reviewed-by: Geo Hsu <geohsu@chromium.org>
Cr-Commit-Position: refs/branch-heads/3770@{#916}
Cr-Branched-From: a9eee1c7c727ef42a15d86e9fa7b77ff0e63840a-refs/heads/master@{#652427}