Update accept-language doc
No change to logic, this only updates a method doc to reflect the
method's current behavior.

GenerateAcceptLanguageHeader() decrements q-values by 0.1, not 0.2. I'm
not sure when the behavior changed. I observed this when using the
method (http://crrev.com/c/1637451), but this does not depend on nor
block that change.

Test: N/A
Change-Id: Ib45bd6f6da66ad80a6c675898000f71e165dbe13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1637600
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664921}