[Autofill] Fixes a screen reader problem.
Screen readers repeat suggestions' disambiguation information, and this
is not a good user experience.

Example (Googlers only):
https://docs.google.com/presentation/d/1mjd7DOCV3cjms5A95fnBKJ4MYYPEWY0XkXu1C0DXcE0/edit#slide=id.g5a92c07359_0_345

Screen readers are instructed to read suggestions' label and
additional_label fields; however, in the credit cards and profiles
contexts, label stores a string that is used on mobile platforms and
additional_label stores a string that is used on desktop.

This CL ensures that, in the credit cards and profiles contexts, label
stores the correct disambiguation text for the platform and
additional_label is unused.

Bug: 968100, 967241
Change-Id: I49ffb781614488867615d5a3797b1342c49b3bc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1635413
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#664925}