[AF][Status Chip] Fix PageActionIconView::UpdateBorder()
Currently it overrides the function in IconLabelBubbleViews(), but the
base class method is called in constructor, so the derived class method
is never called in this case. Calling virtual method in constructor
should be avoided.

Therefore change the function simply to a private function.

Bug: 932818
Change-Id: I790ade09db63bb4c90843722cc3a23ee5525c434
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1636288
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Siyu An <siyua@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665263}