Disable OSSettingsSmbTest.All for debug builds
This test is flaky on debug builds; mostly fails with timeouts.
For now, it's better to disable this test case on debug builds.
This is somewhat a common pattern among webui tests -- and I think
it's fine as long as it's tested on the release builds.

Bug: 968608
Test: none
Change-Id: I09b19f1e1854f9e5943b55b75e61abef777c493e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1639143
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Jordy Greenblatt <jordynass@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665309}