[remoting][FTL] HeartbeatSender report UNAUTHENTICATED
If for some reason the host gets removed from the directory but it fails
to receive notification from the server, it will get into a zombie state
where it keeps heartbeating and fails until FtlSignalStrategy decides
to restart the stream and fails because of unauthenticated request,
which terminates the host. This means the host will be in zombie state
for up to 13 minutes.

This CL makes HeartbeatSender report the UNAUTHENTICATED state to ME2ME
host so that it can terminate the host earlier. This reduces the zombie
time to <= ~6 minutes.

Bug: 969176
Change-Id: Ie1f6af700234f8716197e9a841e17d2407f9b998
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1639582
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665386}