Captured mouse event local_point should use transform
When mouse events are captured, we use the capturing element instead of
doing hit test. The local_point is computed by AbsoluteToLocalPoint, but
we didn't apply transfroms. This is wrong and causes selection incorrect
when there is transform.
This CL removes the incorrect kIgnoreTransforms.

Bug: 969695
Change-Id: Icfcaf627f58999f037b793feeb3fe712a827ca8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1634719
Commit-Queue: Ella Ge <eirage@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665554}