Use testdriver Action API in WPT pointerevent tests - Part 9
In order to run the web-platform-tests automatically, we will use
testdriver Action API in all the wpt to simulate inputs. Here we are
changing one pointerevent input test which has "touch-action: none" on
the table's row and cell, which tests that touch-action does not apply
to table rows, row groups, table columns, and column groups.

This is the ninth CL that changes part of the pointerevent input
tests.

Bug: 606367
Change-Id: Ie3b1d4a192c5c538f0de4c60ab5c9c492185d561
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1639080
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665568}