Fix in-GPU-process media service in WebView
WebView environments are configured to run the media service in the GPU
process even though GPU runs in browser process. This means we have a
valid GpuProcessHost with null process_id(). Recent refactoring of
Service Manager code in content misinterpreted this to mean the
GpuProcessHost was invalid and that services couldn't be launched by
it.

This fixes that assumption and always attempts to launch an
in-GPU-process service if the GpuProcessHost is non-null. Also in such
cases where the reported PID is null, this correctly uses the browser's
own PID as the PID for the PID of the launched service instance.

This mirrors the behavior of in-GPU-process services prior to the
refactoring in r663141.

TBR=rockot@google.com

(cherry picked from commit 90d07f040396eeacf4fb942d0aeed421651c0ef0)

Bug: 967288
Change-Id: Idc9223815f45a2e65766077fa4e695af3e6ceff5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1637980
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#665057}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1640786
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/branch-heads/3809@{#20}
Cr-Branched-From: d82dec1a818f378c464ba307ddd9c92133eac355-refs/heads/master@{#665002}