Use async targeting to find drag and drop targets
Add callback supports to the RenderWidgetTargeter
and use the async targetting in drag and drop code
and register a callback to send drag events using
the callback. Note that since drag and drop events
aren't WebInputEvent it is not possible to use
FindTargetAndDispatch function of RenderWidgetTargeter
at this time.

The original CL
https://chromium-review.googlesource.com/c/chromium/src/+/1596840
was reverted due the added test being flaky. This CL is the same
as the original CL expect it has the same test with another
way of checking the logic.

TBR=sadrul@chromium.org

Bug: 804633
Change-Id: I46f15a2b694c3ad9d3574473c2a46ba13dcbfb79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1641383
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Commit-Queue: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665601}