Optimize thumbnail capturing for Tab-to-Grid transition animation
Split the forced live thumbnail capturing out of the normal caching
path. The forced live thumbnail is captured as low-res, and won't be
saved to the cache. This makes it faster so that using it during
the Tab-to-Grid transition animation won't affect the frame rate too
much.

Bug: 968829
Change-Id: I32ffa8553b5aefb29944781404292a8b891d8d80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1637979
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665691}