Wire up signal for plugin vm state changes
This CL wires up the signal for plugin vm state changes and uses it to
improve the launching experience. If we are in a suspending, stopping,
resetting, or pausing state, we now wait for the operation to complete
before starting the vm rather than simply failing to launch the vm. In
these case we also show a spinner in the shelf.

As we can now keep track of the state of the vm, we also update the
shelf and launcher context menu to not show the Shut Down option when
the VM is not actually running.

Bug: 940319
Change-Id: Ic0141275fce0ec792d6aa809ebb50d2f944be6c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1640777
Commit-Queue: Timothy Loh <timloh@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665785}