Implement DNS padding for DoH
All DoH queries will be padded to the next multiple of 128 bytes.
Implemented as a feature of DnsQuery to add the padding to the OPT
record, rather than just having it specified in the |opt_rdata| passed
to the DnsQuery constructor because adding the correct amount of padding
requires knowledge of the DnsQuery being constructed.

Bug: 967848
Change-Id: I6be13c24c770e80d06ac3f7e988cdf814ce5c00e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1614454
Commit-Queue: Eric Orth <ericorth@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#665935}