[DevTools] Serialize to JSON via CBOR.
I'm planning to remove protocol::Serializable::serialize(bool),
removing the ability there to make a ProtocolMessage
that carries JSON. This is the last usage of such messages
in both Chromium and V8; so I've put a bit of conversion
boilerplate code which is borrowed from
content/browser/devtools/devtools_protocol_encoding.cc.
Sadly it's longer, but it is a price to pay for simplifying
the API. Thanks much for considering.

Change-Id: I24a01e1ed6134f58fa65f03d29506afa33a7ff01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1641551
Commit-Queue: Johannes Henkel <johannes@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666095}