Added OnOpeningUrl to AssistantControllerObserver.
Previously, AssistantControllerObserver::OnUrlOpened would fire after
we requested that a url be opened in a new browser tab but before our
widget would lose focus. Now, due to the de-mojoization of the
relationship between ash/browser, this is no longer the case.

This causes our widget in standalone Assistant UI to lose widget
activation state and become hidden before we receive the OnUrlOpened
event. As a result, we don't know to close Assistant UI in OnUrlOpened
because we are already in hidden state.

To address this, I've added a new event that fires before the new
browser tab is opened so that observers requiring early notification
can continue to do so.

Bug: b:134409807
Change-Id: I1c827102813fcdadddfe76a777880e4616bd38af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1643537
Commit-Queue: David Black <dmblack@google.com>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666167}