ime: Revert MODECHANGE key behaviour back to original.
In a previous patch, we changed the MODECHANGE key behaviour so that
pressing it once will only show the current input method. This was to
follow the UX specs.

However, upon consideration, we're going to change it back to the
original behaviour of switching to next input method on first key press.

The original code used a event rewriter to convert the MODECHANGE key
into Ctrl-Shift-Space (which switches to the next key press). We then
changed the code to be in ash accelerators because of the extra logic
involved. Now that we don't need this extra logic, we could revert back
to event rewriter, but we think it's better to keep it in accelerators
because we still have some extra logic around histogram logging.

TBR=holte@chromium.org

Bug: 953901
Change-Id: I2cf4dc41300b62ee6f57085408e199a2a55062b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642487
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666233}