Disable leaking testcases in unit_tests (the test binary) on LSAN.
See crrev.com/c/1590404 and the linked bug for context -- the goal is to
fix the broken memory-leak checking in Chromium's ASAN
(AddressSanitizer) trybots, which are supposed to also run LSAN
(LeakSanitizer).

Currently-leaking tests must be fixed or disabled before leak checking
can be enabled in the ASAN trybot.

WARNING: This will result in a loss of ASAN coverage for the affected
tests given that LSAN runs as part of the ASAN trybot.

If that's not acceptable, we can wait until the leaks are fixed before
fixing leak detection in trybots, but keep in mind that newly-introduced
leaks Chromium-wide won't be caught until then.

Bug: 961073,969812
Change-Id: Iff5ca80465118a03cb6d5b544ad67acc7b22b9d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1640784
Commit-Queue: Caleb Raitto <caraitto@chromium.org>
Reviewed-by: Alexei Filippov <alph@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666329}