[CrOS Network] Fix erroneous warning text from configure WiFi dialog.
The "configure network" dialog displays a configuration error in some
cases when it should not. This was caused by the fact that we were
inspecting the last_error attribute of NetworkState, even if there is no
current error.

To fix this issue, I removed the NetworkState::GetErrorState() function
and updated clients to simply call the error() function instead.

Bug: 968474
Change-Id: I2d351057ec8b0d2d35a0ea8a46a4515413bec9f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642692
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666332}