[IndexedDB] Disable missing blob database corruption on Docs
This change removes the kWipeCorruptV2IDBDatabases feature to now mark
databases that have missing blobs (so permantently corrupt databases)
as corrupt, so they are cleared appropriately. Since Docs has error
handling here, and we are in active communications with them, disable
this corruption handling only for Docs. The reporting will stay in
place to monitor how small of a problem this is.

The numbers on this are pretty small, there are less than 0.005% of
users that have these non-docs corrupt databases.

R=pwnall@chromium.org

Bug: 857159
Change-Id: Icf8feac9ad6b626108e00213447a706bbe264e5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1633334
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666468}