Implement the full Supervision Onboarding flow.
This change adds the "Details" and "All Set" pages to the flow, as well
as a class to manage the (now multiple) onboarding states.

Another change is that we stop removing some webview listeners when we
stop loading the page. This was causing the webview to never finish
loading subsequent pages, getting stuck in the first blocking listener.

Bug: 958995
Change-Id: Ib28f8d271263ffa832e05a5c6cefb7f7675b02a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1630223
Commit-Queue: Lucas Tenório <ltenorio@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Reviewed-by: Oliver Chang <ochang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666535}