[Editing] Allow NextWordPosition to move across a block's trailing BR
When a block ends with <br>, |NextWordPositionInternal| used to return
BR@AfterNode as the next word position of BR@BeforeNode. This is
incorrect as there's no visible position after a block's trailing BR,
and therefore the caret ends up not moved.

This patch detects this case and moves to the next block if we have
reached the end of a block that ends with a BR.

Bug: 968334
Change-Id: I277821ef5e85f54e080047750d1c54f68c4f4120
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642035
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Auto-Submit: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666577}