[ozone/wayland] Fix broken software rendering path.
The "[Ozone/Wayland] Manager: make mojo calls on IO thread." CL:
https://crrev.com/c/1640398 broke the software rendering path,
which results in a DCHECK now.

It turned out that when software rendering is used, the buffers
are committed on the VizCompositorThread, whereas hw accelerated
rendering uses GpuMainThread instead.

To resolve the conflict, rename the |gpu_thread_runner_| to
the |commit_thread_runner_|, and use it when OnSubmission
and OnPresentation calls come.

Bug: 969603
Change-Id: I3600e35fdc9d4fd0817ce9948316a2af86108bdb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642558
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#666629}