Media Controls: Show pause button when trying to play while loading
This CL modifies the media controls CSS to show the pause icon whenever
media is trying to play (even when it is stopped due to metadata load
or buffering). This fixes an issue where if users tried to pause media
while metadata was loading, there was no indication whether the video
was going to play after load or remain paused.

Bug: 772766
Change-Id: If8dd943f62e4f0fd9ed9625f289cc6467ca1e764
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1645640
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666741}