Add gtk_disable_setlocale() to fix single-process DCHECK
gtk_init() calls setlocale(LC_ALL, ""), which sometimes sets the
LC_NUMERIC locale to something other than "C" (e.g. if the LANG
environment variable is set to a locale). Blink needs it to be
set to "C", so in single-process mode, Blink DCHECKs when it finds
that the browser process has set it to something else (via gtk_init).

So tell gtk to not call setlocale(LC_ALL, "") as we do that elsewhere
already, and add a DCHECK that the  LC_NUMERIC locale is set to "C".

Bug: 948917
Change-Id: I5f9f66f06e3e58b27372cdd3b10962df0e352284
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642975
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666817}