[CORS] Partially rely on factory bound initiator origin
Chrome content scripts (currently) sets the script origin to the
request initiator for requests made by XHR and fetch. Usually such a
request is treated as a same-origin request when the request URL lives
in the origin of the page with relying on the "allowlist" feature.

But, the doesn't work for some kind of content scripts when OOR-CORS is
enabled. Hence we relax the same-originness check by partially relying
on the origin bound to the associated URLLoaderFactory. It *usually* is
set to the frame's origin.

There is one exception, HTML imports. HTML imports are not recognized
by the browser process, so relying on the factory bound origin would
lead to a security issue, hence we don't use the factory bound origin
in that case. This means the original problem may not be solved when
HTML imports are involved (I don't know if there is such a
circumstance), but given HTML imports are being deprecated I would
like to accept such a small breakage.

Bug: 967626
Change-Id: Ic10ee51f4967be8abea7ed290e350dc4907690ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642752
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666946}