Shrink scroll_begin WebGestureEvent data
crrev.com/c/1574950 added a field to the scroll_begin struct on
WebGestureEvent. When doing so, I noticed a few opportunities to shrink
the size of the class - pack the bools together and have the enums
contained be 8-bits in size.

This shrinks the size of WebGestureEvent to one int smaller than prior
to that change.

Change-Id: I61ff66b1e79b534de1fed2fecf8b856ef27bb320
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1646558
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#666972}