Distinguish ServiceWorkerDatabase::ResourceRecord's error state
The context is http://crbug.com/946719. We want to fix a crash, however,
it would be difficult to reproduce it locally. To get an insight on a crash,
it would be better to distinguish an error state of ResourceRecord,
so that we can know who sets "-1" here.

Now an error state has two kinds of values, instead of just -1, and we can
know who sets an error state clearly when WriteResourceRecoredInBatch is
called and DCHECK is hit.

Note that |ResourceRecord::size_bytes| is a public field, so this CL is not
perfect, however, that could be okay for the purpose of debugging.

Bug: 946719

Change-Id: I592e1a4622220f5c7ff4eb5112947100b5562e24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1645113
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667053}