[Passwords] Touch To Fill Fixes
This change introduces a couple of small fixes to the Touch To Fill
feature. In particular, it breaks a circular dependency between the
constructors of the TouchToFillController and the
ManualFillingController and exposes the Touch to Fill feature flag to
Android.

Bug: 957532
Change-Id: Id1c9f227139d7b5273d1db6b0ca13a5004addf52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1649513
Commit-Queue: Vasilii Sukhanov <vasilii@chromium.org>
Auto-Submit: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667173}