[DeviceSync v2] Check if CryptAuthDeviceRegistry changes
The previously void methods AddDevice(), DeleteDevice(), and
SetRegistry() now return a boolean, indicating whether or not the
registry changed as a result of the call. Also, the pref will now only
be updated when the registry changes. And, DeleteDevice() no longer
DCHECKs that a device with the input |instance_id| is in the registry,
instead returning false if such a device doesn't exist.

The RemoteDeviceProvider needs to know when a DeviceSync updates the
local device cache. The aforementioned changes to
CryptAuthDeviceRegistry will help make that determination for v2
DeviceSyncs.

Bug: 951969
Change-Id: I6ad98cadc659d02895da3b16eed90b0ba50773bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1638386
Commit-Queue: Josh Nohle <nohle@chromium.org>
Auto-Submit: Josh Nohle <nohle@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667227}