[remoting host] Use offline-mode for token-exchange.
This changes OAuthTokenExchanger's default mode to offline. This ensures
that a non-empty refresh token will actually be delivered to the caller
on Windows, instead of an empty string.

This CL also updates OAuthTokenExchanger so that it no longer triggers
every hour if offline-mode is used to provide a refresh token.

The effect of this CL on non-Windows platforms is that OAuthTokenGetter
will get an exchanged refresh token and will cache it in memory, using
it to get new access tokens every hour. This will avoid hitting the
TokenExchange server every hour.

Bug: 971920
Change-Id: Ic0d3e3a9ca5a5e627ae135bb005dc186ebfb4e0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1648730
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667267}