M76 merge: [ios] Fix crash in page placeholder
WebState::WasShown() is called even when the webState's view is
not being presented. For example, this can happen during clear
browsing data when the BVC is deactivated and activated. That
is described in crbug.com/971364.

Due to the complexity and ambiguity of BVC's |-setActive:| method,
it is not modified in this CL. Instead, a lower risk approach is
taken by not attempting to display the page placeholder if it is
impossible to do so. The worst that can happen is that the page
placeholder is not shown, which would not be that terrible to the
user. This will prevent crashes seen in the real world.

TBR=eugenebut@chromium.org
(cherry picked from commit a4c70dd48a6eaaaecfb5d9175b372c948f33fe7d)

Bug: 962843, 968096, 970876
Change-Id: I00f68b78aff89aa154d0f98841310ab80723fe79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1645479
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#666761}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1650601
Reviewed-by: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/branch-heads/3809@{#145}
Cr-Branched-From: d82dec1a818f378c464ba307ddd9c92133eac355-refs/heads/master@{#665002}