[threadpool] Do not run BEST_EFFORT tasks during startup (reland).
This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/1631668.
The CL was reverted because of tests failures on chromeos-kevin-rel.
These failures was fixed by changing the priority of a task
in locale_util::SwitchLanguage. See
https://chromium-review.googlesource.com/c/chromium/src/+/1642563

With this CL, a task posted with:

base::PostTaskWithTraits(
    FROM_HERE,
    {base::TaskPriority::BEST_EFFORT, ...},
    base::BindOnce(...));

Will not be allowed to run before startup is complete.

The signal used to know that "startup is complete" is the same as
AfterStartupTaskUtils. It is not perfect. It will be improved in a
separate CL.

This CL aims to be the smallest change possible to allow seamless
migration of BrowserThread::PostAfterStartupTask() to
base::PostTaskWithTraits().

Bug: 887407, 968554
Change-Id: I2c17ccedae9bd9909a988311a295d896dfda695b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642732
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667470}