[net/http] Add smoke tests for external GSSAPI library loading.
There were no tests for ensuring that the GSSAPI library loading logic
works. This CL adds a couple of smoke tests that exercises the loading
logic.

It is not the intention of this CL to ensure that the method invocations
are correct since that depends on there being a real GSSAPI library.
Such tests are being considered for adding as end-to-end tests, but are
beyond the scope of this CL.

Bug: 807741
Change-Id: I9fa6d9427cee16902a766657bcba4d9ec6273960
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1643909
Commit-Queue: Asanka Herath <asanka@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667490}