Use DragEventGenerator for SplitView Resize test
This is reland of: https://chromium-review.googlesource.com/c/chromium/src/+/1637244
Modified test to explicity exit the tablet mode.
There seems to be an issue in shutdown path and I'll file a separate bug for it.

This also adds extra scenarios.
* Resizing with overview on the other side.
* touch event

I also found the scenario that splitview controller can be
called recursively during shutdown, which is fixed by the change
in SplitViewController

TBR=mukai,xdai,sky

Bug: None
Change-Id: I5ebd70a2f24e2544261374f6ac4794bdb152338e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1637244
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#666091}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1650368
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667575}