Add histogram and enums for tracking Cr50 flash log
The Cr50 flash log event types are one byte values, but in reality
very few event values are defined.

NVMEM flash log events are all logged as subevents of a single flash
log event. For the purposes of representation in the metrics space, we
offset NVMEM subevent values by 200 and log them as separate events.

The patch introducing the metrics generating script is
https://crrev.com/c/1649196

BUG=b:132287488
TEST=none

Change-Id: I7a44b69452f816a34012a88e7d12ad7eeaf9aa69
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1601932
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667690}