Fix gestures not being handled
It seems that when the compositor touch action is enabled, the way we
stop timeout monitor is incorrect. This results in two things:
1. Some of the pointer event are not getting acked and hence gesture
events are not handled correctly.
2. There are still quite a lot of gesture events getting timeout out.

In this CL, we have a temporary fix for problem 1 so that some of the
website can be fixed. Then we will have another follow up to solve
problem 2.

Bug: 971937, 971186
Change-Id: Id3ef77f0772c75895deb291323cd9cd45553204a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1650369
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667725}