Have AccountReconcilor check if tokens are loaded on ChromeOS
AccountReconcilor currently defines a check for IdentityManager being
ready as refresh tokens being loaded on all platforms except ChromeOS,
where it has a special-case kludge to account for the fact that refresh
tokens were not loaded in all circumstances on ChromeOS
(crbug.com/749535). However, we have since fixed crbug.com/749535, so
the logic on ChromeOS can now be the same as on other platforms.

Making this change now is necessary to unblock
https://chromium-review.googlesource.com/c/chromium/src/+/1642549, which
changes the semantics of IdentityManager::GetAccountsWithRefreshTokens()
to return an empty set until the tokens are loaded.

TEST=./out/chromeos_debug/components_unittests --gtest_filter="*AccountReconcilor*"

Bug: 919793
Change-Id: I3d48388a1efc165ae8d73eac9f180f12a44df112
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1645309
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Kush Sinha <sinhak@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667904}