OAuth2TokenService: Enforce GetAccounts() is empty until tokens loaded
OAuth2TokenService has a documented invariant that its GetAccounts()
method returns an empty list until refresh tokens are loaded from
disk (as denoted by the firing of the OnRefreshTokensLoaded() observer
callback). However, nothing enforces this invariant, as GetAccounts()
is implemented in platform-specific delegates. The lack of enforcement
of this invariant is blocking our simplification of O2TS observer APIs,
as that simplification requires that clients be able to reliably query
the state of the refresh tokens and make inferences based on that state
and the documented invariants of the various methods.

This CL enforces the invariant via the following:
- Moving tracking of whether tokens are loaded from
  ProfileOAuth2TokenService down to OAuth2TokenService
- Having OAuth2TokenService return an empty vector from GetAccounts() if
  the tokens are not yet loaded

This CL also adds a unittest of this behavior.

Note that this CL is making a behavioral change, and thus it might have
behavioral impact. Any client that was relying on the ability to see
accounts from OAuth2TokenService::GetAccounts() before tokens were
loaded from disk was relying on undocumented and buggy behavior; hence,
it will be good to ferret out and fix such instances if they exist.

Bug: 919793
Change-Id: I002348a28db68c8a7245ef50f0384992dc046f96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642549
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Kush Sinha <sinhak@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667940}