Make TextAutosizer work for OOPIFs.
This CL modifies TextAutosizer to work in OOPIFs by (i) revising the
UpdatePageInfoInAllFrames() logic to be able to handle a frame tree
with a remote main frame, and (ii) separating the logic in
UpdatePageInfo() into separate cases for remote and local main frames.
Local main frames are handled as before; in the remote main frame case
the information that used to be calculated on the fly is pulled from
the WebTextAutosizerPageInfo data stored on the Page object.

This CL is the completion of the work started in
https://chromium-review.googlesource.com/c/chromium/src/+/1627930.

Bug: 393285
Change-Id: I706e64d243ec3bbb889b9898edae6e7e37fe7274
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1643104
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#667945}