Allowing datalist option values that don't match pattern
The html list spec [1] leaves the behavior of the suggestions list
to the UA. It does not reference the pattern [2] spec at all. And
the pattern spec only says that a pattern mismatch should just set
patternMismatch flag to false, it does not mention validating
datalist options. It would seem that it would be beneficial to
show options values regardless of their pattern match status.

With this change, fields all still validate and enforce limits
correctly, they just show the user all options now. There are cases
to be made on both sides of this behavior, of course, but this
feels more permissive. Said another way, with this change, you
can achieve most desired behaviors. Before this change, there was
no way, for example, to show a list of options that don't
validate, but still apply validation to the end result. The bug
that prompted this CL (crbug.com/607097) is just such an example:
it provides several "starting points" for the text field, which
still require the user to tack a number on to the end to
validate. Helpful. Also closer to Firefox behavior.

[1] https://html.spec.whatwg.org/multipage/input.html#the-list-attribute
[2] https://html.spec.whatwg.org/multipage/input.html#the-pattern-attribute
[3] https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#suffering-from-a-pattern-mismatch

Bug: 607097
Change-Id: I27fec4239044365fa4800ce1064f5ab56f2781bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1639278
Commit-Queue: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668001}