[OT-PW] Add a custom_property_name in KeyframeModel
In order for CC animations to work for off-thread paint worklet, we
must know the name of an animated custom property. This CL adds the
custom property name to the KeyframeModel class, with the default value
of empty meaning that it is an animated native property.

This is part of the design here:
https://docs.google.com/document/d/1a7gO6cBxsJhn53akuJuieUiXvB74vvEmFDyKww8NKdw/edit

Even though the design is not finalized yet, adding the custom
property name is a must step in the doc.

Bug: 956650
Change-Id: I6d680dc1638a4d95db366f87c3fdc08fe949ff40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1648732
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668013}