Synchronize NavigableContentsView's LocalHostView focus w/ window focus.
Previously, if the user clicked into the WebContents of an
AssistantCardElementView, the corresponding RenderWidgetHostViewAura
would become focused. However, the focused view in the Assistant view
hierarchy (usually the Textfield in the Assistant Dialog Plate) wouldn't
receive a blur event. As such, the user would think they could keep
typing because of the blinking caret in the Textfield. When typing,
they would find nothing entered and think keyboard was broken.

Now, we synchronize the focus state of NavigableContentsView's
LocalHostView with the associated native window so that as the native
window subtree takes focus, so does the LocalHostView.

This fixes AssistantCardElementView further upstream.

Bug: b:134170714
Change-Id: I03d67a3d9fbc779a87449065c97d7e4a877fe10a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1649084
Commit-Queue: David Black <dmblack@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668077}