Fix spectator-mode.html
The base layer passed to updateRenderState needs compositionDisabled =
true when session mode = inline. Otherwise, there is a JS exception.

In order to see the canvas rendered on the page, I removed a line that
is not actually necessary. Even without the line to hide the canvas, the
button to enter spectator mode is still visible.

Change-Id: I645917f0a517ec4daa3d2ed4c0510e1e64093b3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1653632
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Jacob DeWitt <jacde@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668125}