MacPWAs Save Panels 3/3: Create select file dialog in app shim process
Add a helper function remote_cocoa::GetWindowMojoInterface, which will
get the mojo interface for a gfx::NativeWindow in the browser process.

Add a mojo method to create a SelectFileDialog to NativeWidgetNSWindow.

Use remote_cocoa::GetWindowMojoInterface and the new method on that
interface to create app shim dialogs in the app shim process.

Bug: 913303
Change-Id: I0f608a15c67c49f916b92e37ecf7fd1d966145f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1650801
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668226}