Fix settings pref storage for incognito profiles / Chrome OS guest mode
This fixes a longstanding issue where guest mode users cannot change
the browser's preferred language list.

The main profile in guest mode is an off-the-record/incognito profile.
This CL from 2.5 years ago https://codereview.chromium.org/2499093002
changed the settingsPrivate API in guest mode to use the backing
("recording") profile instead of the main incognito profile for pref
writes. The languages UI continued to function because the incognito
profile's preferences were "overlaid" onto the recording profile prefs.

By default, all prefs were overlaid, except for a block-list. That
block-list was changed to a whitelist, and the whitelisted prefs
were gradually removed. This CL removed the language prefs from the
list:
https://chromium-review.googlesource.com/c/chromium/src/+/1174118/
which made the guest mode language section stop working.

For guest mode we want to use the incognito pref store where possible.
The incognito pref store is in-memory. The backing user profile is
deleted after logout (it is on a tmpfs) but it's better to just avoid
writing the prefs.

Change the BCKSFactories to prefer the incognito profile. Return null
services in tests because unit tests with TestingProfile generally
don't need these services.

Bug: 887383
Test: added to browser_tests, manual test in guest mode on eve

Change-Id: Ib4bbe45662b13aa3b733eaddc4342de5b68086b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1653748
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668434}