Attempting to deflake tab_drag_controller_interactive_uitest.cc tests.
When QuitDraggingObserver is constructed, it registers a callback for
when the current drag finishes. If the drag has already finished, it
will be too late to receive that signal. Thus it is racy to construct
QuitDraggingObserver after starting the drag. This patch marks Wait as
l-value only to prompt clients to construct QuitDraggingObserver in the
correct manner, and fixes the existing uses.

Bug: 973604
Change-Id: I29d700870d710610359fc069671a867c61186d89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1649313
Reviewed-by: Charlene Yan <cyan@chromium.org>
Commit-Queue: Bret Sepulveda <bsep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668619}