PaymentRequest: Error message should span the width of the dialog
Before this patch, the width of the error message label is computed
depending on the width of the input field. If the input field is a
short field, the error message label is also short like attached
screenshot[1]. So, this patch makes the error message label flow all
the way to the right edge in the payment sheet dialog.

[1] https://bugs.chromium.org/p/chromium/issues/attachment?aid=397016&signed_aid=GrGW6Pbwx7-Pso6ha35ENQ==&inline=1

Bug: 972072
Change-Id: I92cc3d8b775a6ebe90be154501ce37620784fb4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1654730
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Jinho Bang <jinho.bang@samsung.com>
Cr-Commit-Position: refs/heads/master@{#668652}