Reland: Disable install event.prompt() after user has already installed the site
Before this CL we would only disable prompt() if the user installs the site
via the menu or the web app banner. This updates the web app omnibox install
button to also disable prompt() from being called.
Without disabling prompt() the site can re-prompt the user to install
which results in the browser crashing.

Note: This CL is a "quick fix" to resolve the browser crash in a mergeable
CL. A set of follow up CLs will change AppBannerManager to listen to web app
install events and clear the prompt() handle that way instead.

This is a reland of:
https://chromium-review.googlesource.com/c/chromium/src/+/1640742
The previous CL's test hit an MSAN error where the test class was being cast
as a sibling class erroneously. This CL moves the tests to
AppBannerManagerDesktopBrowserTest so the test class can subclass the existing
AppBannerManagerDesktop and avoid a sibling cast.
See reland diff:
https://chromium-review.googlesource.com/c/chromium/src/+/1653789/1..8

Bug: 956810
Change-Id: I31a3c4a6d6ae78a6a629d47119a89a4118efb702
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1653789
Auto-Submit: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668655}