tablet: Cleanup some code in TabletModeController.
Now that we start observing before arranging the windows instead of
after starting the animation. There is no need for the workaround to
emplace nullptr window states in the tablet mode window manager.

Removes this and updates some of the code to use constexpr and make
unique.

Test: none
Bug: 948706
Change-Id: Icbbeb67c0419ca996e0824ead1c6374ab0cb63dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1657548
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668703}