[Autofill Assistant] Make positioning of highlighted area independent of viewport height.
This change fixes issues with the event filter of the overlay not
adapting to viewport height changes.

With this change, the position of highlighted elements are reported as
CSS pixels, relative to the page, instead of coordinates relative to the
viewport size. The viewport size and position within the page is tracked
and reported separately.

When displaying or comparing against filter coordinates, CSS pixels are
converted into physical pixels using the viewport width, which allows
adapting to zoom.

This change also simplifies the display of the highlighted area by
ignoring the offset reported to the onScroll methods as it was causing
issues. This allows the model to cache the element position and the
visual viewport position independently; when scrolling, unless the
element position is absolute, element position normally stay the same
and only the viewport changes.

This is the M-76 merge for http://crrev/c/1634737.

(cherry picked from commit 553002120f1f5381f2905f26c17e298b75a452b2)

Bug: 973034
Bug: b/133669408
Bug: b/129050125
Change-Id: I745584b496d960a2d73231238dd2006afb3f7eaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1634737
Reviewed-by: Jordan Demeulenaere <jdemeulenaere@chromium.org>
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#665536}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1657911
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Cr-Commit-Position: refs/branch-heads/3809@{#280}
Cr-Branched-From: d82dec1a818f378c464ba307ddd9c92133eac355-refs/heads/master@{#665002}