[Autofill Assistant] Deferred module install.
Before this patch the autofill_assistant dynamic feature module was
installed via on-demand. This introduces a 10+ second delay before AA
is available and leads to a bad user experience. Moving the loading
and installing step out of the critical path and into the app's
upgrade step mitigates that.

This patch adds a deferred install trigger in Chrome's package replace
hook and the CCT startup. The condition for the deferred trigger is
high end phones with Android L+. The fallback to on-demand is still kept
as a safety net.

Bug: b/132424979
Change-Id: I8327b4acd60cef28e22502ae9cfd29e8e352310a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1619686
Commit-Queue: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#668798}